all_vars {dplyr}	R Documentation

Apply predicate to all variables

Description

These quoting functions signal to scoped filtering verbs (e.g. filter_if() or filter_all()) that a predicate expression should be applied to all relevant variables. The all_vars() variant takes the intersection of the predicate expressions with & while the any_vars() variant takes the union with |.

Usage

all_vars(expr)

any_vars(expr)

Arguments

expr
A predicate expression. This variable supports unquoting and will be evaluated in the context of the data frame. It should return a logical vector.

This argument is automatically quoted and later evaluated in the context of the data frame. It supports unquoting. See vignette("programming") for an introduction to these concepts.

See Also

funs() and vars() for other quoting functions that you can use with scoped verbs.

[Package dplyr version 0.7.8 Index]