as.table.tbl_cube {dplyr}	R Documentation

Coerce a tbl_cube to other data structures

Description

Supports conversion to tables, data frames, tibbles.

For a cube, the data frame returned by tibble::as_data_frame() resulting data frame contains the dimensions as character values (and not as factors).

Usage

## S3 method for class 'tbl_cube'
as.table(x, ..., measure = 1L)

## S3 method for class 'tbl_cube'
as.data.frame(x, ...)

## S3 method for class 'tbl_cube'
as_tibble(x, ...)

Arguments

x
a tbl_cube

...
Passed on to individual methods; otherwise ignored.

measure
A measure name or index, default: the first measure

[Package dplyr version 0.7.8 Index]