as.tbl_cube {dplyr}	R Documentation

Coerce an existing data structure into a tbl_cube

Description

Coerce an existing data structure into a tbl_cube

Usage

as.tbl_cube(x, ...)

## S3 method for class 'array'
as.tbl_cube(x, dim_names = names(dimnames(x)),
  met_name = deparse(substitute(x)), ...)

## S3 method for class 'table'
as.tbl_cube(x, dim_names = names(dimnames(x)),
  met_name = "Freq", ...)

## S3 method for class 'matrix'
as.tbl_cube(x, dim_names = names(dimnames(x)),
  met_name = deparse(substitute(x)), ...)

## S3 method for class 'data.frame'
as.tbl_cube(x, dim_names = NULL,
  met_name = guess_met(x), ...)

Arguments

x
an object to convert. Built in methods will convert arrays, tables and data frames.

...
Passed on to individual methods; otherwise ignored.

dim_names
names of the dimensions. Defaults to the names of

met_name
a string to use as the name for the measure the dimnames().

[Package dplyr version 0.7.8 Index]