Do values in a numeric vector fall in specified range?

This is a shortcut for x >= left & x <= right, implemented efficiently in C++ for local values, and translated to the appropriate SQL for remote tables.

Usage

between(x, left, right)

Arguments

x

A numeric vector of values

left, right

Boundary values (must be scalars).

Examples

between(1:12, 7, 9)
#>  [1] FALSE FALSE FALSE FALSE FALSE FALSE  TRUE  TRUE  TRUE FALSE FALSE
#> [12] FALSE

x <- rnorm(1e2)
x[between(x, -1, 1)]
#>  [1]  0.460640070 -0.553553325 -0.968124516 -0.189945642  0.796875216
#>  [6]  0.031533712 -0.748305188  0.039424412  0.905661594  0.833231638
#> [11]  0.203089785 -0.944601862  0.434532914  0.701573421  0.904793147
#> [16]  0.853780057 -0.060064147  0.488737145 -0.955695310 -0.365085695
#> [21]  0.362594892 -0.196970098  0.067021236  0.993174172 -0.007072562
#> [26] -0.869634573 -0.803073076  0.203672136  0.236531661 -0.667446407
#> [31]  0.566548194 -0.435262589 -0.742869114 -0.294231531  0.210147873
#> [36] -0.807662063  0.904527142 -0.133120762 -0.257811031 -0.800931113
#> [41] -0.684899821  0.830634532 -0.116895313  0.198119341 -0.156174177
#> [46]  0.703246061 -0.924180017  0.716021174 -0.514185795 -0.513469216
#> [51]  0.194167284  0.470861257  0.431940279  0.826985384 -0.370538963
#> [56]  0.457114695  0.744017457 -0.482764254 -0.842839162  0.171335307
#> [61]  0.507521742  0.075137921 -0.143547774  0.587593426  0.284745166
#> [66]  0.061675227  0.096750704 -0.409620611

## Or on a tibble using filter
filter(starwars, between(height, 100, 150))
#> # A tibble: 5 × 14
#>   name       height  mass hair_color skin_color eye_color birth_year sex  
#>   <chr>       <int> <dbl> <chr>      <chr>      <chr>          <dbl> <chr>
#> 1 Leia Orga…    150    49 brown      light      brown             19 fema…
#> 2 Mon Mothma    150    NA auburn     fair       blue              48 fema…
#> 3 Watto         137    NA black      blue, grey yellow            NA male 
#> 4 Sebulba       112    40 none       grey, red  orange            NA male 
#> 5 Gasgano       122    NA none       white, bl… black             NA male 
#> # … with 6 more variables: gender <chr>, homeworld <chr>, species <chr>,
#> #   films <list>, vehicles <list>, starships <list>