Condense values into list-columns

condense() is an experimental variant of summarise() designed to make it as easy as possible to work with list-columns. It is particularly powerful when working with models: you can fit models per group with condense() and then flexibly extract components with summarise().

It has three special properties:

It only ever returns a single row per group.

Each new column is always wrapped in a list.

It returns a rowwise() output.

It's hard to explain condense() in isolation, so if you find these ideas intriguing we recommend starting with vignette("rowwise").

condense(.data, ...)

Arguments

.data
A data frame, data frame extension (e.g. a tibble), or a lazy data frame (e.g. from dbplyr or dtplyr). See Methods, below, for more details.

...
<data-masking> Name-value pairs of functions. All outputs will be automatically wrapped in lists, making it most suitable for functions that that return non-vectors (e.g. linear models) or vectors of length greater than one.

Value

A rowwise object.

The rows and initial columns come from group_data(.data).

Each additional column that you provide will be wrapped in a list column.

Output is grouped by row.

Data frame attributes are not preserved, because condense() fundamentally creates a new data frame.

Methods

This function is a generic, which means that packages can provide implementations (methods) for other classes. See the documentation of individual methods for extra arguments and differences in behaviour.

The following methods are currently available in loaded packages: dplyr (data.frame) .

Examples

# Modelling ------------------------------------------------
mods <- mtcars %>%
  group_by(cyl) %>%
  condense(mod = lm(mpg ~ disp, data = across()))
mods

#> # A tibble: 3 x 2
#> # Rowwise:  cyl
#>     cyl mod   
#>   <dbl> <list>
#> 1     4 <lm>  
#> 2     6 <lm>  
#> 3     8 <lm>  

mods %>%
  mutate(rsq = summary(mod)$r.squared)

#> # A tibble: 3 x 3
#> # Rowwise:  cyl
#>     cyl mod       rsq
#>   <dbl> <list>  <dbl>
#> 1     4 <lm>   0.648 
#> 2     6 <lm>   0.0106
#> 3     8 <lm>   0.270 

if (requireNamespace("broom")) {
  mods %>% summarise(broom::glance(mod))
  mods %>% summarise(broom::tidy(mod))
  mods %>% summarise(broom::augment(mod))
}

#> Loading required namespace: broom

#> # A tibble: 32 x 10
#> # Rowwise:  cyl
#>      cyl   mpg  disp .fitted .se.fit .resid   .hat .sigma .cooksd .std.resid
#>    <dbl> <dbl> <dbl>   <dbl>   <dbl>  <dbl>  <dbl>  <dbl>   <dbl>      <dbl>
#>  1     4  22.8 108      26.3   0.855 -3.48  0.0920   2.70 0.0849     -1.29  
#>  2     4  24.4 147.     21.0   1.62   3.35  0.330    2.62 0.521       1.45  
#>  3     4  22.8 141.     21.8   1.46   0.956 0.267    2.96 0.0286      0.396 
#>  4     4  32.4  78.7    30.2   1.22   2.16  0.188    2.87 0.0838      0.852 
#>  5     4  30.4  75.7    30.6   1.29  -0.242 0.211    2.99 0.00125    -0.0965
#>  6     4  33.9  71.1    31.3   1.41   2.64  0.251    2.79 0.196       1.08  
#>  7     4  21.5 120.     24.6   0.984 -3.14  0.122    2.74 0.0982     -1.19  
#>  8     4  27.3  79      30.2   1.21  -2.90  0.186    2.77 0.148      -1.14  
#>  9     4  26   120.     24.6   0.988  1.39  0.123    2.94 0.0193      0.525 
#> 10     4  30.4  95.1    28.0   0.913  2.38  0.105    2.85 0.0466      0.892 
#> # … with 22 more rows