Locale used by dplyr

dplyr_locale() returns a single string representing the default locale used by dplyr when ordering character vectors. It is used as the default value of .locale in arrange().

Default locale

The default locale returned by dplyr_locale() is the C locale, identical to explicitly supplying .locale = "C".

The C locale is not exactly the same as English locales, such as "en". The main difference is that the C locale groups the English alphabet by case, while most English locales group the alphabet by letter. For example, c("a", "b", "C", "B", "c") will sort as c("B", "C", "a", "b", "c") in the C locale, with all uppercase letters coming before lowercase letters, but will sort as c("a", "b", "B", "c", "C") in an English locale. This often makes little practical difference during data analysis, because both return identical results when case is consistent between observations.

Global override

To override the above default behavior, you can set the global option, dplyr.locale, to a stringi locale identifier from stringi::stri_locale_list() to globally alter the default locale. This requires stringi >=1.5.3.

We generally recommend that you set the .locale argument of arrange() explicitly rather than overriding the global locale, if possible.

Another alternative is to only change the global locale within a limited scope through the use of rlang::local_options() or rlang::with_options(). This can be useful when a package that you don't control calls arrange() internally.

Reproducibility

The C locale has the benefit of being completely reproducible across all supported R versions and operating systems with no extra effort.

If you set .locale to an option from stringi::stri_locale_list(), then stringi must be installed by anyone who wants to run your code. If you utilize this in a package, then stringi should be placed in Imports.

Usage

dplyr_locale()

Examples

# Default locale is C
dplyr_locale()
#> [1] "C"

df <- tibble(x = c("a", "b", "C", "B", "c"))
df
#> # A tibble: 5 × 1
#>   x    
#>   <chr>
#> 1 a    
#> 2 b    
#> 3 C    
#> 4 B    
#> 5 c    

# The C locale groups the English alphabet by case, placing uppercase letters
# before lowercase letters. This is the default.
arrange(df, x)
#> # A tibble: 5 × 1
#>   x    
#>   <chr>
#> 1 B    
#> 2 C    
#> 3 a    
#> 4 b    
#> 5 c    

# The American English locale groups the alphabet by letter.
# Explicitly override `.locale` with `"en"` for this ordering.
arrange(df, x, .locale = "en")
#> # A tibble: 5 × 1
#>   x    
#>   <chr>
#> 1 a    
#> 2 b    
#> 3 B    
#> 4 c    
#> 5 C    

# Or temporarily override the `dplyr.locale` global option, which is useful
# if `arrange()` is called from a function you don't control
col_sorter <- function(df) {
  arrange(df, x)
}

rlang::with_options(dplyr.locale = "en", {
  col_sorter(df)
})
#> # A tibble: 5 × 1
#>   x    
#>   <chr>
#> 1 a    
#> 2 b    
#> 3 B    
#> 4 c    
#> 5 C    

# This Danish letter is expected to sort after `z`
df <- tibble(x = c("o", "p", "\u00F8", "z"))
df
#> # A tibble: 4 × 1
#>   x    
#>   <chr>
#> 1 o    
#> 2 p    
#> 3 ø    
#> 4 z    

# The American English locale sorts it right after `o`
arrange(df, x, .locale = "en")
#> # A tibble: 4 × 1
#>   x    
#>   <chr>
#> 1 o    
#> 2 ø    
#> 3 p    
#> 4 z    

# Using `"da"` for Danish ordering gives the expected result
arrange(df, x, .locale = "da")
#> # A tibble: 4 × 1
#>   x    
#>   <chr>
#> 1 o    
#> 2 p    
#> 3 z    
#> 4 ø