dr_dplyr {dplyr}	R Documentation

Dr Dplyr checks your installation for common problems.

Description

Only run this if you are seeing problems, like random crashes. It's possible for dr_dplyr to return false positives, so there's no need to run if all is ok.

Usage

dr_dplyr()

Examples

## Not run: 
dr_dplyr()

## End(Not run)

[Package dplyr version 0.7.8 Index]