funs {dplyr}	R Documentation

Create a list of functions calls.

Description

funs() provides a flexible way to generate a named list of functions for input to other functions like summarise_at().

Usage

funs(..., .args = list())

Arguments

...
A list of functions specified by:

Their name, "mean"

The function itself, mean

A call to the function with . as a dummy argument, mean(., na.rm = TRUE)

These arguments are automatically quoted. They support unquoting and splicing. See vignette("programming") for an introduction to these concepts.

The following notations are not supported, see examples:

An anonymous function, function(x) mean(x, na.rm = TRUE)

An anonymous function in purrr notation, ~mean(., na.rm = TRUE)

.args, args
A named list of additional arguments to be added to all function calls.

Examples

funs(mean, "mean", mean(., na.rm = TRUE))

# Override default names
funs(m1 = mean, m2 = "mean", m3 = mean(., na.rm = TRUE))

# If you have function names in a vector, use funs_
fs <- c("min", "max")
funs_(fs)

# Not supported
## Not run: 
funs(function(x) mean(x, na.rm = TRUE))
funs(~mean(x, na.rm = TRUE))
## End(Not run)

[Package dplyr version 0.7.8 Index]