group_by_prepare {dplyr}	R Documentation

Prepare for grouping.

Description

Performs standard operations that should happen before individual methods process the data. This includes mutating the tbl to add new grouping columns and updating the groups (based on add)

Usage

group_by_prepare(.data, ..., .dots = list(), add = FALSE)

Value

A list

data
Modified tbl

groups
Modified groups

[Package dplyr version 0.7.8 Index]