grouped_df {dplyr}	R Documentation

A grouped data frame.

Description

The easiest way to create a grouped data frame is to call the group_by() method on a data frame or tbl: this will take care of capturing the unevaluated expressions for you.

Usage

grouped_df(data, vars, drop = TRUE)

is.grouped_df(x)

is_grouped_df(x)

Arguments

data
a tbl or data frame.

vars
a character vector or a list of name()

drop
if TRUE preserve all factor levels, even those without data.

[Package dplyr version 0.7.8 Index]