Join methods for data frames

This page describes the details of the join generics when applied to data frames and tibbles.

All methods treat the x input as primary: the return value will be the same type as x and the rows will in be the same order (duplicated where necessary, and where needed missing rows from y will be added to the end.)

# S3 method for data.frame
inner_join(
  x,
  y,
  by = NULL,
  copy = FALSE,
  suffix = c(".x", ".y"),
  ...,
  na_matches = pkgconfig::get_config("dplyr::na_matches")
)

# S3 method for data.frame
left_join(
  x,
  y,
  by = NULL,
  copy = FALSE,
  suffix = c(".x", ".y"),
  ...,
  na_matches = pkgconfig::get_config("dplyr::na_matches")
)

# S3 method for data.frame
right_join(
  x,
  y,
  by = NULL,
  copy = FALSE,
  suffix = c(".x", ".y"),
  ...,
  na_matches = pkgconfig::get_config("dplyr::na_matches")
)

# S3 method for data.frame
full_join(
  x,
  y,
  by = NULL,
  copy = FALSE,
  suffix = c(".x", ".y"),
  ...,
  keep = FALSE,
  na_matches = pkgconfig::get_config("dplyr::na_matches")
)

# S3 method for data.frame
semi_join(
  x,
  y,
  by = NULL,
  copy = FALSE,
  ...,
  na_matches = pkgconfig::get_config("dplyr::na_matches")
)

# S3 method for data.frame
anti_join(
  x,
  y,
  by = NULL,
  copy = FALSE,
  ...,
  na_matches = pkgconfig::get_config("dplyr::na_matches")
)

# S3 method for data.frame
nest_join(x, y, by = NULL, copy = FALSE, keep = FALSE, name = NULL, ...)

Arguments

x, y
Data frames

by
A character vector of variables to join by.

If NULL, the default, *_join() will perofrm a natural join, using all variables in across x and y. A message lists the variables so that you can check they're correct; suppress the message by supply by explicitly.

To join by different variables on x and y use a named vector. For example, by = c("a" = "b") will match x$a to y$b.

To perform a cross-join, generating all combinations of x and y, use by = character().

copy
If x and y are not from the same data source, and copy is TRUE, then y will be copied into the same src as x. This allows you to join tables across srcs, but it is a potentially expensive operation so you must opt into it.

suffix
If there are non-joined duplicate variables in x and y, these suffixes will be added to the output to disambiguate them. Should be a character vector of length 2.

...
Included for compatibility with the generic; otherwise ignored.

na_matches
Should NA and NaN values match one another?

Use "never" to always treat two NA or NaN values as different, like joins for database sources, similarly to merge(incomparables = FALSE). The default, "na", always treats two NA or NaN values as equal, like %in%, match(), merge().

Users and package authors can change the default behavior by calling pkgconfig::set_config("dplyr::na_matches" = "never").

keep
Should the join keys from y be preserved in the output? Only applies to nest_join() and full_join().

name
The name of the list column nesting joins create. If NULL the name of y is used.

Examples

df1 <- data.frame(x = c(1, NA), y = 2)
df2 <- data.frame(x = c(1, NA), z = 3)

# By default, NAs match other NAs so that there are two
# rows in the output:
left_join(df1, df2)

#> Joining, by = "x"

#>    x y z
#> 1  1 2 3
#> 2 NA 2 3

# You can optionally request that NAs don't match, giving a
# a result that more closely resembles SQL joins
left_join(df1, df2, na_matches = "never")

#> Joining, by = "x"

#> Warning: `na_matches = 'never' currently unsupported

#>    x y z
#> 1  1 2 3
#> 2 NA 2 3