last_dplyr_warnings {dplyr}	R Documentation

Show warnings from the last command

Description

Warnings that occur inside a dplyr verb like mutate() are caught and stashed away instead of being emitted to the console. This prevents rowwise and grouped data frames from flooding the console with warnings. To see the original warnings, use last_dplyr_warnings().

Usage

last_dplyr_warnings(n = 5)

Arguments

n
Passed to head() so that only the first n warnings are displayed.

[Package dplyr version 1.1.3 Index]