n {dplyr}	R Documentation

The number of observations in the current group.

Description

This function is implemented specifically for each data source and can only be used from within summarise(), mutate() and filter().

Usage

n()

Examples

if (require("nycflights13")) {
carriers <- group_by(flights, carrier)
summarise(carriers, n())
mutate(carriers, n = n())
filter(carriers, n() < 100)
}

[Package dplyr version 0.7.8 Index]