same_src {dplyr}	R Documentation

Figure out if two sources are the same (or two tbl have the same source)

Description

Figure out if two sources are the same (or two tbl have the same source)

Usage

same_src(x, y)

Arguments

x, y
src or tbls to test

Value

a logical flag

[Package dplyr version 1.1.3 Index]