Select helpers

These functions allow you to select variables based on their names.

starts_with(): starts with a prefix

ends_with(): ends with a prefix

contains(): contains a literal string

matches(): matches a regular expression

num_range(): a numerical range like x01, x02, x03.

one_of(): variables in character vector.

everything(): all variables.

current_vars()

starts_with(match, ignore.case = TRUE, vars = current_vars())

ends_with(match, ignore.case = TRUE, vars = current_vars())

contains(match, ignore.case = TRUE, vars = current_vars())

matches(match, ignore.case = TRUE, vars = current_vars())

num_range(prefix, range, width = NULL, vars = current_vars())

one_of(..., vars = current_vars())

everything(vars = current_vars())

Arguments

match
A string.

ignore.case
If TRUE, the default, ignores case when matching names.

vars
A character vector of variable names. When called from inside select() these are automatically set to the names of the table.

prefix
A prefix that starts the numeric range.

range
A sequence of integers, like 1:5

width
Optionally, the "width" of the numeric range. For example, a range of 2 gives "01", a range of three "001", etc.

...
One or more character vectors.

Value

An integer vector giving the position of the matched variables.

Examples

iris <- tbl_df(iris) # so it prints a little nicer
select(iris, starts_with("Petal"))

#> # A tibble: 150 × 2
#>    Petal.Length Petal.Width
#>           <dbl>       <dbl>
#> 1           1.4         0.2
#> 2           1.4         0.2
#> 3           1.3         0.2
#> 4           1.5         0.2
#> 5           1.4         0.2
#> 6           1.7         0.4
#> 7           1.4         0.3
#> 8           1.5         0.2
#> 9           1.4         0.2
#> 10          1.5         0.1
#> # ... with 140 more rows

select(iris, ends_with("Width"))

#> # A tibble: 150 × 2
#>    Sepal.Width Petal.Width
#>          <dbl>       <dbl>
#> 1          3.5         0.2
#> 2          3.0         0.2
#> 3          3.2         0.2
#> 4          3.1         0.2
#> 5          3.6         0.2
#> 6          3.9         0.4
#> 7          3.4         0.3
#> 8          3.4         0.2
#> 9          2.9         0.2
#> 10         3.1         0.1
#> # ... with 140 more rows

select(iris, contains("etal"))

#> # A tibble: 150 × 2
#>    Petal.Length Petal.Width
#>           <dbl>       <dbl>
#> 1           1.4         0.2
#> 2           1.4         0.2
#> 3           1.3         0.2
#> 4           1.5         0.2
#> 5           1.4         0.2
#> 6           1.7         0.4
#> 7           1.4         0.3
#> 8           1.5         0.2
#> 9           1.4         0.2
#> 10          1.5         0.1
#> # ... with 140 more rows

select(iris, matches(".t."))

#> # A tibble: 150 × 4
#>    Sepal.Length Sepal.Width Petal.Length Petal.Width
#>           <dbl>       <dbl>        <dbl>       <dbl>
#> 1           5.1         3.5          1.4         0.2
#> 2           4.9         3.0          1.4         0.2
#> 3           4.7         3.2          1.3         0.2
#> 4           4.6         3.1          1.5         0.2
#> 5           5.0         3.6          1.4         0.2
#> 6           5.4         3.9          1.7         0.4
#> 7           4.6         3.4          1.4         0.3
#> 8           5.0         3.4          1.5         0.2
#> 9           4.4         2.9          1.4         0.2
#> 10          4.9         3.1          1.5         0.1
#> # ... with 140 more rows

select(iris, Petal.Length, Petal.Width)

#> # A tibble: 150 × 2
#>    Petal.Length Petal.Width
#>           <dbl>       <dbl>
#> 1           1.4         0.2
#> 2           1.4         0.2
#> 3           1.3         0.2
#> 4           1.5         0.2
#> 5           1.4         0.2
#> 6           1.7         0.4
#> 7           1.4         0.3
#> 8           1.5         0.2
#> 9           1.4         0.2
#> 10          1.5         0.1
#> # ... with 140 more rows

select(iris, everything())

#> # A tibble: 150 × 5
#>    Sepal.Length Sepal.Width Petal.Length Petal.Width Species
#>           <dbl>       <dbl>        <dbl>       <dbl>  <fctr>
#> 1           5.1         3.5          1.4         0.2  setosa
#> 2           4.9         3.0          1.4         0.2  setosa
#> 3           4.7         3.2          1.3         0.2  setosa
#> 4           4.6         3.1          1.5         0.2  setosa
#> 5           5.0         3.6          1.4         0.2  setosa
#> 6           5.4         3.9          1.7         0.4  setosa
#> 7           4.6         3.4          1.4         0.3  setosa
#> 8           5.0         3.4          1.5         0.2  setosa
#> 9           4.4         2.9          1.4         0.2  setosa
#> 10          4.9         3.1          1.5         0.1  setosa
#> # ... with 140 more rows

vars <- c("Petal.Length", "Petal.Width")
select(iris, one_of(vars))

#> # A tibble: 150 × 2
#>    Petal.Length Petal.Width
#>           <dbl>       <dbl>
#> 1           1.4         0.2
#> 2           1.4         0.2
#> 3           1.3         0.2
#> 4           1.5         0.2
#> 5           1.4         0.2
#> 6           1.7         0.4
#> 7           1.4         0.3
#> 8           1.5         0.2
#> 9           1.4         0.2
#> 10          1.5         0.1
#> # ... with 140 more rows