summarise_each {dplyr}	R Documentation

Summarise and mutate multiple columns.

Description

mutate_each() and summarise_each() are deprecated in favour of a more featureful family of functions: mutate_all(), mutate_at(), mutate_if(), summarise_all(), summarise_at() and summarise_if().

The _each() functions have two replacements depending on what variables you want to apply funs to. To apply a function to all variables, use mutate_all() or summarise_all(). To apply a function to a selection of variables, use mutate_at() or summarise_at().

See the relevant section of vignette("compatibility") for more information.

Usage

summarise_each(tbl, funs, ...)

summarise_each_(tbl, funs, vars)

mutate_each(tbl, funs, ...)

mutate_each_(tbl, funs, vars)

summarize_each(tbl, funs, ...)

summarize_each_(tbl, funs, vars)

[Package dplyr version 0.7.8 Index]