tidyeval-compat {dplyr}	R Documentation

Other tidy eval tools

Description

These tidy eval tools are no longer recommended for normal usage, but are still exported for compatibility. See ?tidyeval for the recommended tools.

[Package dplyr version 0.8.3 Index]