with_groups {dplyr}	R Documentation

Perform an operation with temporary groups

Description

This was an experimental function that allows you to modify the grouping variables for a single operation; it is superseded in favour of using the .by argument to individual verbs.

Usage

with_groups(.data, .groups, .f, ...)

Arguments

.data
A data frame

.groups
<tidy-select> One or more variables to group by. Unlike group_by(), you can only group by existing variables, and you can use tidy-select syntax like c(x, y, z) to select multiple variables.

Use NULL to temporarily ungroup.

.f
Function to apply to regrouped data. Supports purrr-style ~ syntax

...
Additional arguments passed on to ....

Examples

df <- tibble(g = c(1, 1, 2, 2, 3), x = runif(5))

# Old
df %>%
  with_groups(g, mutate, x_mean = mean(x))
# New
df %>% mutate(x_mean = mean(x), .by = g)

[Package dplyr version 1.1.3 Index]