Apply a function or a set of functions to a set of columns

Creates a data frame by applying a set of functions to a tidy selection of columns in the current slice

current_key()

current_column()

across(select, fns = identity)

Arguments

select
tidy selection of columns

fns
Functions to apply to each of the selected columns. Possible values are:

A single function or a single quosure style lambda, e.g. ~ mean(.x, na.rm = TRUE)

A named list of functions and/or lambdas

Value

A tibble

Details

This allows you to use select() semantics inside in summarise() and mutate()

When a single function is given, it is applied to each of the selected columns and the output columns are named after the selected input columns.

When a named list of functions is given, the result is made of as many columns as the number of functions, each of these output columns is a pack, i.e. a data frame column made from the result of the function applied across the selected columns.

current_column() gives the name of the column currently being processed.

current_key() return metadata about the "current" element: a single row tibble, with one column per grouping variable

Examples

# A single function
iris %>%
  group_by(Species) %>%
  summarise(across(starts_with("Sepal"), mean))

#> # A tibble: 3 x 3
#>   Species    Sepal.Length Sepal.Width
#>   <fct>             <dbl>       <dbl>
#> 1 setosa             5.01        3.43
#> 2 versicolor         5.94        2.77
#> 3 virginica          6.59        2.97

iris %>%
  group_by(Species) %>%
  summarise(across(starts_with("Sepal"), ~mean(.x, na.rm = TRUE)))

#> # A tibble: 3 x 3
#>   Species    Sepal.Length Sepal.Width
#>   <fct>             <dbl>       <dbl>
#> 1 setosa             5.01        3.43
#> 2 versicolor         5.94        2.77
#> 3 virginica          6.59        2.97

# a named list of functions
iris %>%
  group_by(Species) %>%
  summarise(across(starts_with("Sepal"), list(mean = mean, sd = sd)))

#> # A tibble: 3 x 3
#>   Species    mean$Sepal.Length $Sepal.Width sd$Sepal.Length $Sepal.Width
#>   <fct>                  <dbl>        <dbl>           <dbl>        <dbl>
#> 1 setosa                  5.01         3.43           0.352        0.379
#> 2 versicolor              5.94         2.77           0.516        0.314
#> 3 virginica               6.59         2.97           0.636        0.322