Apply a function (or a set of functions) to a set of columns

across() makes it easy to apply the same transformation to multiple columns, allowing you to use select() semantics inside in summarise() and mutate(). across() supersedes the family of "scoped variants" like summarise_at(), summarise_if(), and summarise_all().

See vignette("colwise") for more details.

c_across() is designed to work with rowwise() to make it easy to perform rowwise aggregations; it works like c() but uses tidy select semantics so you can easily select multiple variables. See vignette("rowwise") for more details.

across(cols = everything(), fns = NULL, names = NULL, ...)

c_across(cols = everything())

Arguments

cols
<tidy-select> Columns to transform. Because across() is used within functions like summarise() and mutate(), you can't select or compute upon grouping variables.

fns
Functions to apply to each of the selected columns. Possible values are:

NULL, to returns the columns untransformed.

A function, e.g. mean.

A purrr-style lambda, e.g. ~ mean(.x, na.rm = TRUE)

A list of functions/lambdas, e.g. list(mean = mean, n_miss = ~ sum(is.na(.x))

Within these functions you can use cur_column() and cur_group() to access the current column and grouping keys respectively.

names
A glue specification that describes how to name the output columns. This can use {col} to stand for the selected column name, and {fn} to stand for the name of the function being applied. The default (NULL) is equivalent to "{col}" for the single function case and "{col}_{fn}" for the case where a list is used for fns.

...
Additional arguments for the function calls in fns.

Value

A tibble with one column for each column in cols and each function in fns.

Examples

# across() -----------------------------------------------------------------
iris %>%
  group_by(Species) %>%
  summarise(across(starts_with("Sepal"), mean))

#> # A tibble: 3 x 3
#>   Species    Sepal.Length Sepal.Width
#>   <fct>             <dbl>       <dbl>
#> 1 setosa             5.01        3.43
#> 2 versicolor         5.94        2.77
#> 3 virginica          6.59        2.97

iris %>%
  as_tibble() %>%
  mutate(across(is.factor, as.character))

#> # A tibble: 150 x 5
#>    Sepal.Length Sepal.Width Petal.Length Petal.Width Species
#>           <dbl>       <dbl>        <dbl>       <dbl> <chr>  
#>  1          5.1         3.5          1.4         0.2 setosa 
#>  2          4.9         3            1.4         0.2 setosa 
#>  3          4.7         3.2          1.3         0.2 setosa 
#>  4          4.6         3.1          1.5         0.2 setosa 
#>  5          5           3.6          1.4         0.2 setosa 
#>  6          5.4         3.9          1.7         0.4 setosa 
#>  7          4.6         3.4          1.4         0.3 setosa 
#>  8          5           3.4          1.5         0.2 setosa 
#>  9          4.4         2.9          1.4         0.2 setosa 
#> 10          4.9         3.1          1.5         0.1 setosa 
#> # … with 140 more rows

# A purrr-style formula
iris %>%
  group_by(Species) %>%
  summarise(across(starts_with("Sepal"), ~mean(.x, na.rm = TRUE)))

#> # A tibble: 3 x 3
#>   Species    Sepal.Length Sepal.Width
#>   <fct>             <dbl>       <dbl>
#> 1 setosa             5.01        3.43
#> 2 versicolor         5.94        2.77
#> 3 virginica          6.59        2.97

# A named list of functions
iris %>%
  group_by(Species) %>%
  summarise(across(starts_with("Sepal"), list(mean = mean, sd = sd)))

#> # A tibble: 3 x 5
#>   Species    Sepal.Length_mean Sepal.Length_sd Sepal.Width_mean Sepal.Width_sd
#>   <fct>                  <dbl>           <dbl>            <dbl>          <dbl>
#> 1 setosa                  5.01            3.43            0.352          0.379
#> 2 versicolor              5.94            2.77            0.516          0.314
#> 3 virginica               6.59            2.97            0.636          0.322

# Use the names argument to control the output names
iris %>%
  group_by(Species) %>%
  summarise(across(starts_with("Sepal"), mean, names = "mean_{col}"))

#> # A tibble: 3 x 3
#>   Species    mean_Sepal.Length mean_Sepal.Width
#>   <fct>                  <dbl>            <dbl>
#> 1 setosa                  5.01             3.43
#> 2 versicolor              5.94             2.77
#> 3 virginica               6.59             2.97

iris %>%
  group_by(Species) %>%
  summarise(across(starts_with("Sepal"), list(mean = mean, sd = sd), names = "{col}.{fn}"))

#> # A tibble: 3 x 5
#>   Species    Sepal.Length.mean Sepal.Length.sd Sepal.Width.mean Sepal.Width.sd
#>   <fct>                  <dbl>           <dbl>            <dbl>          <dbl>
#> 1 setosa                  5.01            3.43            0.352          0.379
#> 2 versicolor              5.94            2.77            0.516          0.314
#> 3 virginica               6.59            2.97            0.636          0.322

iris %>%
  group_by(Species) %>%
  summarise(across(starts_with("Sepal"), list(mean, sd), names = "{col}.fn{fn}"))

#> # A tibble: 3 x 5
#>   Species    Sepal.Length.fn1 Sepal.Length.fn2 Sepal.Width.fn1 Sepal.Width.fn2
#>   <fct>                 <dbl>            <dbl>           <dbl>           <dbl>
#> 1 setosa                 5.01             3.43           0.352           0.379
#> 2 versicolor             5.94             2.77           0.516           0.314
#> 3 virginica              6.59             2.97           0.636           0.322

# c_across() ---------------------------------------------------------------
df <- tibble(id = 1:4, w = runif(4), x = runif(4), y = runif(4), z = runif(4))
df %>%
  rowwise() %>%
  mutate(
    sum = sum(c_across(w:z)),
    sd = sd(c_across(w:z))
 )

#> # A tibble: 4 x 7
#> # Rowwise: 
#>      id      w       x     y      z   sum     sd
#>   <int>  <dbl>   <dbl> <dbl>  <dbl> <dbl>  <dbl>
#> 1     1 0.0808 0.00740 0.733 0.0342 0.855 0.347 
#> 2     2 0.834  0.466   0.773 0.320  2.39  0.245 
#> 3     3 0.601  0.498   0.875 0.402  2.38  0.204 
#> 4     4 0.157  0.290   0.175 0.196  0.818 0.0590