all_vars {dplyr}	R Documentation

Apply predicate to all variables

Description

all_vars() and any_vars() were only needed for the scoped verbs, which have been superseded by the use of across() in an existing verb. See vignette("colwise") for details.

These quoting functions signal to scoped filtering verbs (e.g. filter_if() or filter_all()) that a predicate expression should be applied to all relevant variables. The all_vars() variant takes the intersection of the predicate expressions with & while the any_vars() variant takes the union with |.

Usage

all_vars(expr)

any_vars(expr)

Arguments

expr
<data-masking> An expression that returns a logical vector, using . to refer to the "current" variable.

See Also

vars() for other quoting functions that you can use with scoped verbs.

[Package dplyr version 1.0.7 Index]