args_by {dplyr}	R Documentation

Helper for consistent documentation of .by

Description

Use @inheritParams args_by to consistently document .by.

Arguments

.by

<tidy-select> Optionally, a selection of columns to group by for just this operation, functioning as an alternative to group_by(). For details and examples, see ?dplyr_by.

[Package dplyr version 1.1.1 Index]