Arrange rows by a selection of variables

These scoped variants of arrange() sort a data frame by a selection of variables. Like arrange(), you can modify the variables before ordering with funs().

arrange_all(.tbl, .funs = list(), ...)

arrange_at(.tbl, .vars, .funs = list(), ...)

arrange_if(.tbl, .predicate, .funs = list(), ...)

Arguments

.tbl
A tbl object.

.funs
List of function calls generated by funs(), or a character vector of function names, or simply a function.

Bare formulas are passed to rlang::as_function() to create purrr-style lambda functions. Note that these lambda prevent hybrid evaluation from happening and it is thus more efficient to supply functions like mean() directly rather than in a lambda-formula.

...
Additional arguments for the function calls in .funs. These are evaluated only once, with explicit splicing.

.vars
A list of columns generated by vars(), or a character vector of column names, or a numeric vector of column positions.

.predicate
A predicate function to be applied to the columns or a logical vector. The variables for which .predicate is or returns TRUE are selected. This argument is passed to rlang::as_function() and thus supports quosure-style lambda functions and strings representing function names.

Examples

df <- as_tibble(mtcars)
df

#> # A tibble: 32 × 11
#>      mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#>  * <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1  21.0     6 160.0   110  3.90 2.620 16.46     0     1     4     4
#>  2  21.0     6 160.0   110  3.90 2.875 17.02     0     1     4     4
#>  3  22.8     4 108.0    93  3.85 2.320 18.61     1     1     4     1
#>  4  21.4     6 258.0   110  3.08 3.215 19.44     1     0     3     1
#>  5  18.7     8 360.0   175  3.15 3.440 17.02     0     0     3     2
#>  6  18.1     6 225.0   105  2.76 3.460 20.22     1     0     3     1
#>  7  14.3     8 360.0   245  3.21 3.570 15.84     0     0     3     4
#>  8  24.4     4 146.7    62  3.69 3.190 20.00     1     0     4     2
#>  9  22.8     4 140.8    95  3.92 3.150 22.90     1     0     4     2
#> 10  19.2     6 167.6   123  3.92 3.440 18.30     1     0     4     4
#> # ... with 22 more rows

arrange_all(df)

#> # A tibble: 32 × 11
#>      mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#>    <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1  10.4     8 460.0   215  3.00 5.424 17.82     0     0     3     4
#>  2  10.4     8 472.0   205  2.93 5.250 17.98     0     0     3     4
#>  3  13.3     8 350.0   245  3.73 3.840 15.41     0     0     3     4
#>  4  14.3     8 360.0   245  3.21 3.570 15.84     0     0     3     4
#>  5  14.7     8 440.0   230  3.23 5.345 17.42     0     0     3     4
#>  6  15.0     8 301.0   335  3.54 3.570 14.60     0     1     5     8
#>  7  15.2     8 275.8   180  3.07 3.780 18.00     0     0     3     3
#>  8  15.2     8 304.0   150  3.15 3.435 17.30     0     0     3     2
#>  9  15.5     8 318.0   150  2.76 3.520 16.87     0     0     3     2
#> 10  15.8     8 351.0   264  4.22 3.170 14.50     0     1     5     4
#> # ... with 22 more rows

# You can supply a function that will be applied before taking the
# ordering of the variables. The variables of the sorted tibble
# keep their original values.
arrange_all(df, desc)

#> # A tibble: 32 × 11
#>      mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#>    <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1  33.9     4  71.1    65  4.22 1.835 19.90     1     1     4     1
#>  2  32.4     4  78.7    66  4.08 2.200 19.47     1     1     4     1
#>  3  30.4     4  95.1   113  3.77 1.513 16.90     1     1     5     2
#>  4  30.4     4  75.7    52  4.93 1.615 18.52     1     1     4     2
#>  5  27.3     4  79.0    66  4.08 1.935 18.90     1     1     4     1
#>  6  26.0     4 120.3    91  4.43 2.140 16.70     0     1     5     2
#>  7  24.4     4 146.7    62  3.69 3.190 20.00     1     0     4     2
#>  8  22.8     4 140.8    95  3.92 3.150 22.90     1     0     4     2
#>  9  22.8     4 108.0    93  3.85 2.320 18.61     1     1     4     1
#> 10  21.5     4 120.1    97  3.70 2.465 20.01     1     0     3     1
#> # ... with 22 more rows

arrange_all(df, funs(desc(.)))

#> # A tibble: 32 × 11
#>      mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#>    <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1  33.9     4  71.1    65  4.22 1.835 19.90     1     1     4     1
#>  2  32.4     4  78.7    66  4.08 2.200 19.47     1     1     4     1
#>  3  30.4     4  95.1   113  3.77 1.513 16.90     1     1     5     2
#>  4  30.4     4  75.7    52  4.93 1.615 18.52     1     1     4     2
#>  5  27.3     4  79.0    66  4.08 1.935 18.90     1     1     4     1
#>  6  26.0     4 120.3    91  4.43 2.140 16.70     0     1     5     2
#>  7  24.4     4 146.7    62  3.69 3.190 20.00     1     0     4     2
#>  8  22.8     4 140.8    95  3.92 3.150 22.90     1     0     4     2
#>  9  22.8     4 108.0    93  3.85 2.320 18.61     1     1     4     1
#> 10  21.5     4 120.1    97  3.70 2.465 20.01     1     0     3     1
#> # ... with 22 more rows