Arrange rows by a selection of variables
Source: R/colwise-arrange.R

These scoped variants of arrange() sort a data frame by a selection of variables. Like arrange(), you can modify the variables before ordering with funs().

arrange_all(.tbl, .funs = list(), ...)

arrange_at(.tbl, .vars, .funs = list(), ...)

arrange_if(.tbl, .predicate, .funs = list(), ...)

Arguments

.tbl
A tbl object.

.funs
List of function calls generated by funs(), or a character vector of function names, or simply a function.

Bare formulas are passed to rlang::as_function() to create purrr-style lambda functions. Note that these lambda prevent hybrid evaluation from happening and it is thus more efficient to supply functions like mean() directly rather than in a lambda-formula.

...
Additional arguments for the function calls in .funs. These are evaluated only once, with tidy dots support.

.vars
A list of columns generated by vars(), or a character vector of column names, or a numeric vector of column positions.

.predicate
A predicate function to be applied to the columns or a logical vector. The variables for which .predicate is or returns TRUE are selected. This argument is passed to rlang::as_function() and thus supports quosure-style lambda functions and strings representing function names.

Examples

df <- as_tibble(mtcars)
df

#> # A tibble: 32 x 11
#>      mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#>  * <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1  21.0    6.  160.  110.  3.90  2.62  16.5    0.    1.    4.    4.
#>  2  21.0    6.  160.  110.  3.90  2.88  17.0    0.    1.    4.    4.
#>  3  22.8    4.  108.   93.  3.85  2.32  18.6    1.    1.    4.    1.
#>  4  21.4    6.  258.  110.  3.08  3.22  19.4    1.    0.    3.    1.
#>  5  18.7    8.  360.  175.  3.15  3.44  17.0    0.    0.    3.    2.
#>  6  18.1    6.  225.  105.  2.76  3.46  20.2    1.    0.    3.    1.
#>  7  14.3    8.  360.  245.  3.21  3.57  15.8    0.    0.    3.    4.
#>  8  24.4    4.  147.   62.  3.69  3.19  20.0    1.    0.    4.    2.
#>  9  22.8    4.  141.   95.  3.92  3.15  22.9    1.    0.    4.    2.
#> 10  19.2    6.  168.  123.  3.92  3.44  18.3    1.    0.    4.    4.
#> # ... with 22 more rows

arrange_all(df)

#> # A tibble: 32 x 11
#>      mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#>    <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1  10.4    8.  460.  215.  3.00  5.42  17.8    0.    0.    3.    4.
#>  2  10.4    8.  472.  205.  2.93  5.25  18.0    0.    0.    3.    4.
#>  3  13.3    8.  350.  245.  3.73  3.84  15.4    0.    0.    3.    4.
#>  4  14.3    8.  360.  245.  3.21  3.57  15.8    0.    0.    3.    4.
#>  5  14.7    8.  440.  230.  3.23  5.34  17.4    0.    0.    3.    4.
#>  6  15.0    8.  301.  335.  3.54  3.57  14.6    0.    1.    5.    8.
#>  7  15.2    8.  276.  180.  3.07  3.78  18.0    0.    0.    3.    3.
#>  8  15.2    8.  304.  150.  3.15  3.44  17.3    0.    0.    3.    2.
#>  9  15.5    8.  318.  150.  2.76  3.52  16.9    0.    0.    3.    2.
#> 10  15.8    8.  351.  264.  4.22  3.17  14.5    0.    1.    5.    4.
#> # ... with 22 more rows

# You can supply a function that will be applied before taking the
# ordering of the variables. The variables of the sorted tibble
# keep their original values.
arrange_all(df, desc)

#> # A tibble: 32 x 11
#>      mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#>    <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1  33.9    4.  71.1   65.  4.22  1.84  19.9    1.    1.    4.    1.
#>  2  32.4    4.  78.7   66.  4.08  2.20  19.5    1.    1.    4.    1.
#>  3  30.4    4.  95.1  113.  3.77  1.51  16.9    1.    1.    5.    2.
#>  4  30.4    4.  75.7   52.  4.93  1.62  18.5    1.    1.    4.    2.
#>  5  27.3    4.  79.0   66.  4.08  1.94  18.9    1.    1.    4.    1.
#>  6  26.0    4. 120.    91.  4.43  2.14  16.7    0.    1.    5.    2.
#>  7  24.4    4. 147.    62.  3.69  3.19  20.0    1.    0.    4.    2.
#>  8  22.8    4. 141.    95.  3.92  3.15  22.9    1.    0.    4.    2.
#>  9  22.8    4. 108.    93.  3.85  2.32  18.6    1.    1.    4.    1.
#> 10  21.5    4. 120.    97.  3.70  2.46  20.0    1.    0.    3.    1.
#> # ... with 22 more rows

arrange_all(df, funs(desc(.)))

#> # A tibble: 32 x 11
#>      mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#>    <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1  33.9    4.  71.1   65.  4.22  1.84  19.9    1.    1.    4.    1.
#>  2  32.4    4.  78.7   66.  4.08  2.20  19.5    1.    1.    4.    1.
#>  3  30.4    4.  95.1  113.  3.77  1.51  16.9    1.    1.    5.    2.
#>  4  30.4    4.  75.7   52.  4.93  1.62  18.5    1.    1.    4.    2.
#>  5  27.3    4.  79.0   66.  4.08  1.94  18.9    1.    1.    4.    1.
#>  6  26.0    4. 120.    91.  4.43  2.14  16.7    0.    1.    5.    2.
#>  7  24.4    4. 147.    62.  3.69  3.19  20.0    1.    0.    4.    2.
#>  8  22.8    4. 141.    95.  3.92  3.15  22.9    1.    0.    4.    2.
#>  9  22.8    4. 108.    93.  3.85  2.32  18.6    1.    1.    4.    1.
#> 10  21.5    4. 120.    97.  3.70  2.46  20.0    1.    0.    3.    1.
#> # ... with 22 more rows