Coerce a tbl_cube to other data structures

Supports conversion to tables, data frames, tibbles.

For a cube, the data frame returned by tibble::as_data_frame() resulting data frame contains the dimensions as character values (and not as factors).

# S3 method for tbl_cube
as.table(x, ..., measure = 1L)

# S3 method for tbl_cube
as.data.frame(x, ...)

# S3 method for tbl_cube
as_data_frame(x, ...)

Arguments

x
a tbl_cube

...
Passed on to individual methods; otherwise ignored.

measure
A measure name or index, default: the first measure