bench_compare {dplyr}	R Documentation

Evaluate, compare, benchmark operations of a set of srcs.

Description

These functions are deprecated because we now believe that you're better of performing the comparisons directly, yourself, in order to generate more informative test failures.

Usage

bench_tbls(tbls, op, ..., times = 10)

compare_tbls(tbls, op, ref = NULL, compare = equal_data_frame, ...)

compare_tbls2(tbls_x, tbls_y, op, ref = NULL, compare = equal_data_frame, ...)

eval_tbls(tbls, op)

eval_tbls2(tbls_x, tbls_y, op)

Arguments

tbls, tbls_x, tbls_y
A list of tbl()s.

op
A function with a single argument, called often with each element of tbls.

...
For compare_tbls(): additional parameters passed on the compare() function

For bench_tbls(): additional benchmarks to run.

times
For benchmarking, the number of times each operation is repeated.

ref
For checking, a data frame to test results against. If not supplied, defaults to the results from the first src.

compare
A function used to compare the results. Defaults to equal_data_frame which ignores the order of rows and columns.

Value

eval_tbls(): a list of data frames.

compare_tbls(): an invisible TRUE on success, otherwise an error is thrown.

bench_tbls(): an object of class microbenchmark::microbenchmark()

[Package dplyr version 1.0.7 Index]