between {dplyr}	R Documentation

Do values in a numeric vector fall in specified range?

Description

This is a shortcut for x >= left & x <= right, implemented efficiently in C++ for local values, and translated to the appropriate SQL for remote tables.

Usage

between(x, left, right)

Arguments

x
A numeric vector of values

left, right
Boundary values (must be scalars).

Examples

between(1:12, 7, 9)

x <- rnorm(1e2)
x[between(x, -1, 1)]

## Or on a tibble using filter
filter(starwars, between(height, 100, 150))

[Package dplyr version 1.0.7 Index]