Detect where values fall in a specified range

This is a shortcut for x >= left & x <= right, implemented for local vectors and translated to the appropriate SQL for remote tables.

Usage

between(x, left, right)

Arguments

x

A vector

left, right

Boundary values. Both left and right are recycled to the size of x and are cast to the type of x.

Examples

between(1:12, 7, 9)
#>  [1] FALSE FALSE FALSE FALSE FALSE FALSE  TRUE  TRUE  TRUE FALSE FALSE
#> [12] FALSE

x <- rnorm(1e2)
x[between(x, -1, 1)]
#>  [1]  0.218121853  0.948688047 -0.033507202  0.040491321 -0.553157845
#>  [6] -0.258530206 -0.564647026  0.589433362  0.448653374 -0.992302424
#> [11] -0.291560765 -0.970899921  0.661608895 -0.140688650 -0.397237832
#> [16]  0.271489987  0.256398227  0.641167824  0.708338408 -0.414173100
#> [21]  0.680947367 -0.215688643 -0.656953386  0.853470558 -0.312159436
#> [26]  0.850546104 -0.334839025  0.665827727  0.098209990 -0.372547655
#> [31]  0.489031175 -0.363287273  0.284326182  0.034756935  0.284074399
#> [36] -0.921484269  0.276753116 -0.054096566  0.756830060 -0.171048053
#> [41] -0.540911749 -0.146835242  0.696601461 -0.008679849 -0.321476668
#> [46] -0.217045157  0.192953243 -0.731245918  0.402445750  0.841073212
#> [51] -0.874512827  0.810890494  0.254959695  0.272416010  0.185323112
#> [56] -0.546222952 -0.054046103 -0.532531002  0.581856885 -0.639158004
#> [61]  0.102324525 -0.228856459  0.193485907  0.333661743 -0.514978576
#> [66] -0.298707580  0.154439532 -0.525721259  0.780438560  0.787403796
#> [71] -0.845631251

# On a tibble using `filter()`
filter(starwars, between(height, 100, 150))
#> # A tibble: 5 × 14
#>   name       height  mass hair_color skin_color eye_color birth_year sex  
#>   <chr>       <int> <dbl> <chr>      <chr>      <chr>          <dbl> <chr>
#> 1 Leia Orga…    150    49 brown      light      brown             19 fema…
#> 2 Mon Mothma    150    NA auburn     fair       blue              48 fema…
#> 3 Watto         137    NA black      blue, grey yellow            NA male 
#> 4 Sebulba       112    40 none       grey, red  orange            NA male 
#> 5 Gasgano       122    NA none       white, bl… black             NA male 
#> # … with 6 more variables: gender <chr>, homeworld <chr>, species <chr>,
#> #   films <list>, vehicles <list>, starships <list>