Do values in a numeric vector fall in specified range?

This is a shortcut for x >= left & x <= right, implemented efficiently in C++ for local values, and translated to the appropriate SQL for remote tables.

Usage

between(x, left, right)

Arguments

x

A numeric vector of values

left, right

Boundary values (must be scalars).

Examples

between(1:12, 7, 9)
#>  [1] FALSE FALSE FALSE FALSE FALSE FALSE  TRUE  TRUE  TRUE FALSE FALSE
#> [12] FALSE

x <- rnorm(1e2)
x[between(x, -1, 1)]
#>  [1] -0.546377651 -0.593116863  0.175985159 -0.490919608  0.768709084
#>  [6] -0.039843483  0.261635057  0.861066071 -0.963555013  0.351846356
#> [11]  0.727032797  0.609425003  0.062826614 -0.105775930  0.327955955
#> [16] -0.422241319  0.474824089  0.049649094  0.053576765  0.964733553
#> [21]  0.528026808 -0.191964017 -0.911797434  0.159322067 -0.005392053
#> [26] -0.233469681  0.655083274 -0.465679514 -0.835770938 -0.019655646
#> [31]  0.380793412 -0.480262267  0.083799850  0.727235761  0.711993534
#> [36]  0.954041590 -0.092506315  0.561640395 -0.073476365 -0.844959996
#> [41]  0.046428679 -0.077145172  0.296444936  0.599131473 -0.271199130
#> [46] -0.848919561 -0.280392734  0.720437853 -0.077399401 -0.477003534
#> [51]  0.011929876  0.653872048  0.115095859 -0.443200497  0.129473642
#> [56]  0.057872523 -0.362710812  0.077408424  0.205596664 -0.760109193
#> [61] -0.329489004 -0.002823810 -0.019200949 -0.700893629

## Or on a tibble using filter
filter(starwars, between(height, 100, 150))
#> # A tibble: 5 × 14
#>   name       height  mass hair_color skin_color eye_color birth_year sex  
#>   <chr>       <int> <dbl> <chr>      <chr>      <chr>          <dbl> <chr>
#> 1 Leia Orga…    150    49 brown      light      brown             19 fema…
#> 2 Mon Mothma    150    NA auburn     fair       blue              48 fema…
#> 3 Watto         137    NA black      blue, grey yellow            NA male 
#> 4 Sebulba       112    40 none       grey, red  orange            NA male 
#> 5 Gasgano       122    NA none       white, bl… black             NA male 
#> # … with 6 more variables: gender <chr>, homeworld <chr>, species <chr>,
#> #   films <list>, vehicles <list>, starships <list>