Detect where values fall in a specified range

This is a shortcut for x >= left & x <= right, implemented for local vectors and translated to the appropriate SQL for remote tables.

Usage

between(x, left, right)

Arguments

x

A vector

left, right

Boundary values. Both left and right are recycled to the size of x.

Value

A logical vector the same size as x.

Details

x, left, and right are all cast to their common type before the comparison is made.

See also

Examples

between(1:12, 7, 9)
#>  [1] FALSE FALSE FALSE FALSE FALSE FALSE  TRUE  TRUE  TRUE FALSE FALSE
#> [12] FALSE

x <- rnorm(1e2)
x[between(x, -1, 1)]
#>  [1] -0.91984399 -0.01974861  0.51874661 -0.05903348  0.94932770
#>  [6] -0.70808160  0.35971665  0.30995957  0.22169338 -0.43611339
#> [11] -0.30497382  0.94422885 -0.73905711  0.43587399  0.11200308
#> [16] -0.89861457 -0.37289221 -0.61142966 -0.67172479  0.64712159
#> [21]  0.41028292  0.10322622 -0.31679373  0.75047794 -0.63386864
#> [26] -0.50737084 -0.33650277 -0.06033292  0.59169436 -0.74707994
#> [31] -0.05813514  0.13291908 -0.60760225  0.84592525 -0.02277037
#> [36]  0.33095608  0.57622297  0.05411518 -0.29546890  0.55598115
#> [41] -0.09745428  0.70745847 -0.11350330 -0.92110219 -0.75431477
#> [46]  0.73964002  0.76397448 -0.45451259  0.19944116  0.64176142
#> [51] -0.35439493  0.85036860 -0.22118695 -0.79027520 -0.14546208
#> [56] -0.07823371  0.36078726 -0.06693282  0.55170545 -0.58957438
#> [61]  0.29750313

# On a tibble using `filter()`
filter(starwars, between(height, 100, 150))
#> # A tibble: 5 × 14
#>   name       height  mass hair_color skin_color eye_color birth_year sex  
#>   <chr>       <int> <dbl> <chr>      <chr>      <chr>          <dbl> <chr>
#> 1 Leia Orga…    150    49 brown      light      brown             19 fema…
#> 2 Mon Mothma    150    NA auburn     fair       blue              48 fema…
#> 3 Watto         137    NA black      blue, grey yellow            NA male 
#> 4 Sebulba       112    40 none       grey, red  orange            NA male 
#> 5 Gasgano       122    NA none       white, bl… black             NA male 
#> # ℹ 6 more variables: gender <chr>, homeworld <chr>, species <chr>,
#> #   films <list>, vehicles <list>, starships <list>