Detect where values fall in a specified range

This is a shortcut for x >= left & x <= right, implemented for local vectors and translated to the appropriate SQL for remote tables.

Usage

between(x, left, right)

Arguments

x

A vector

left, right

Boundary values. Both left and right are recycled to the size of x.

Value

A logical vector the same size as x.

Details

x, left, and right are all cast to their common type before the comparison is made.

Examples

between(1:12, 7, 9)
#>  [1] FALSE FALSE FALSE FALSE FALSE FALSE  TRUE  TRUE  TRUE FALSE FALSE
#> [12] FALSE

x <- rnorm(1e2)
x[between(x, -1, 1)]
#>  [1]  0.5564569786  0.0085242348  0.4964026113 -0.1579975436  0.6813509000
#>  [6] -0.2764720729 -0.7677638097 -0.1190151615 -0.8966560729 -0.4736920324
#> [11] -0.8563559971  0.4148179623 -0.3611946679  0.2937871743  0.7475680775
#> [16] -0.4277876744 -0.9307482411  0.3689745191 -0.5587475164 -0.3455983062
#> [21]  0.6716123597  0.6063745980  0.8975011503 -0.3652112864  0.6768628825
#> [26]  0.6482943084  0.4979409538 -0.6578393326 -0.1281811616  0.9942095717
#> [31]  0.3297754396 -0.0837185069  0.4764592568  0.4375930370 -0.3629954734
#> [36] -0.0007500322  0.2571660980 -0.7292868967 -0.0869977967 -0.4180108425
#> [41] -0.4998721505 -0.9834296145 -0.2958241664  0.0656772905  0.6243156483
#> [46] -0.4628019706 -0.8100099414 -0.0806611084 -0.0528625322  0.4606163136
#> [51]  0.6825709260 -0.6720348531 -0.5611448069 -0.4695095174 -0.8509898839
#> [56]  0.9589572324  0.6779528419  0.2489535529  0.4149817892  0.7498448966
#> [61]  0.1107293718  0.0784373621  0.8062485115 -0.6947697120  0.4323442992
#> [66] -0.9586510898 -0.7769671990  0.8365146447  0.5105256367 -0.6265067606
#> [71]  0.9954352488  0.6355729076

# On a tibble using `filter()`
filter(starwars, between(height, 100, 150))
#> # A tibble: 5 × 14
#>   name   height  mass hair_…¹ skin_…² eye_c…³ birth…⁴ sex   gender homew…⁵
#>   <chr>   <int> <dbl> <chr>   <chr>   <chr>     <dbl> <chr> <chr>  <chr>  
#> 1 Leia …    150    49 brown   light   brown        19 fema… femin… Aldera…
#> 2 Mon M…    150    NA auburn  fair    blue         48 fema… femin… Chandr…
#> 3 Watto     137    NA black   blue, … yellow       NA male  mascu… Toydar…
#> 4 Sebul…    112    40 none    grey, … orange       NA male  mascu… Malast…
#> 5 Gasga…    122    NA none    white,… black        NA male  mascu… Troiken
#> # … with 4 more variables: species <chr>, films <list>, vehicles <list>,
#> #   starships <list>, and abbreviated variable names ¹​hair_color,
#> #   ²​skin_color, ³​eye_color, ⁴​birth_year, ⁵​homeworld