Detect where values fall in a specified range

This is a shortcut for x >= left & x <= right, implemented for local vectors and translated to the appropriate SQL for remote tables.

Usage

between(x, left, right)

Arguments

x

A vector

left, right

Boundary values. Both left and right are recycled to the size of x and are cast to the type of x.

Examples

between(1:12, 7, 9)
#>  [1] FALSE FALSE FALSE FALSE FALSE FALSE  TRUE  TRUE  TRUE FALSE FALSE
#> [12] FALSE

x <- rnorm(1e2)
x[between(x, -1, 1)]
#>  [1]  0.7797697882  0.4293146607 -0.4189722461 -0.0685828984  0.1397743557
#>  [6]  0.1020386041 -0.5325389455 -0.4286074188 -0.3305014009  0.1815200911
#> [11]  0.2031836256  0.8790554542  0.5944811975  0.3472966829 -0.1671729572
#> [16] -0.0077028844 -0.8082397149  0.3989868639  0.2471428141 -0.8250980169
#> [21]  0.3687582814  0.9373782485  0.0751878523 -0.8901253161  0.6068612794
#> [26]  0.0690864151 -0.3914245983  0.8009260329  0.8550299563  0.4066275884
#> [31]  0.0006054723 -0.3015347677 -0.7322581817 -0.7366182197 -0.6194447310
#> [36]  0.2644487352  0.6409674454  0.9925141388  0.6180839746  0.0988662664
#> [41]  0.7913227161  0.5363511721  0.1799378990  0.4548822718  0.1512782556
#> [46]  0.8647758217  0.6787146952 -0.2336040720 -0.7945221738 -0.5899732513
#> [51] -0.5183309748  0.4763436212 -0.3519411154 -0.9148149980 -0.4479350876
#> [56]  0.9226015789  0.6282008777 -0.2064783602  0.4314286375 -0.5732127434
#> [61]  0.0375359540  0.1943952060  0.2821013743  0.1817440600  0.1566911377
#> [66] -0.0666580577 -0.4178228544  0.2940494677  0.9912444750 -0.5361355476
#> [71]  0.9295345050 -0.9565858365 -0.1696146797

# On a tibble using `filter()`
filter(starwars, between(height, 100, 150))
#> # A tibble: 5 × 14
#>   name       height  mass hair_color skin_color eye_color birth_year sex  
#>   <chr>       <int> <dbl> <chr>      <chr>      <chr>          <dbl> <chr>
#> 1 Leia Orga…    150    49 brown      light      brown             19 fema…
#> 2 Mon Mothma    150    NA auburn     fair       blue              48 fema…
#> 3 Watto         137    NA black      blue, grey yellow            NA male 
#> 4 Sebulba       112    40 none       grey, red  orange            NA male 
#> 5 Gasgano       122    NA none       white, bl… black             NA male 
#> # … with 6 more variables: gender <chr>, homeworld <chr>, species <chr>,
#> #   films <list>, vehicles <list>, starships <list>