Do values in a numeric vector fall in specified range?

This is a shortcut for x >= left & x <= right, implemented efficiently in C++ for local values, and translated to the appropriate SQL for remote tables.

Usage

between(x, left, right)

Arguments

x

A numeric vector of values

left, right

Boundary values (must be scalars).

Examples

between(1:12, 7, 9)
#>  [1] FALSE FALSE FALSE FALSE FALSE FALSE  TRUE  TRUE  TRUE FALSE FALSE
#> [12] FALSE

x <- rnorm(1e2)
x[between(x, -1, 1)]
#>  [1]  0.860818691 -0.621887654 -0.905065193  0.026276246  0.348002014
#>  [6]  0.376463091  0.007418543  0.455642493  0.758014153 -0.846844519
#> [11] -0.311864349 -0.894211411 -0.276770302 -0.668178035 -0.331613955
#> [16]  0.206876852  0.936766752 -0.593080594 -0.653186743 -0.543103907
#> [21] -0.148429789 -0.738514568 -0.626819961  0.794339571 -0.165032291
#> [26]  0.085722426  0.829160808 -0.848240991  0.940718023 -0.461979611
#> [31]  0.599207739  0.425332224  0.967116536  0.280240332 -0.463296150
#> [36] -0.826791315 -0.349311020  0.127035493 -0.254015561  0.911394632
#> [41]  0.873573857 -0.160622913  0.307199500 -0.646525683 -0.204619455
#> [46] -0.018739196 -0.420708538 -0.923615583  0.100933457 -0.189448020
#> [51]  0.620471588 -0.847072091  0.494328839  0.239024669 -0.403381550
#> [56] -0.933612226  0.776008801 -0.656015230 -0.802256463 -0.385051267

## Or on a tibble using filter
filter(starwars, between(height, 100, 150))
#> # A tibble: 5 × 14
#>   name       height  mass hair_color skin_color eye_color birth_year sex  
#>   <chr>       <int> <dbl> <chr>      <chr>      <chr>          <dbl> <chr>
#> 1 Leia Orga…    150    49 brown      light      brown             19 fema…
#> 2 Mon Mothma    150    NA auburn     fair       blue              48 fema…
#> 3 Watto         137    NA black      blue, grey yellow            NA male 
#> 4 Sebulba       112    40 none       grey, red  orange            NA male 
#> 5 Gasgano       122    NA none       white, bl… black             NA male 
#> # … with 6 more variables: gender <chr>, homeworld <chr>, species <chr>,
#> #   films <list>, vehicles <list>, starships <list>