Detect where values fall in a specified range

This is a shortcut for x >= left & x <= right, implemented for local vectors and translated to the appropriate SQL for remote tables.

Usage

between(x, left, right)

Arguments

x

A vector

left, right

Boundary values. Both left and right are recycled to the size of x and are cast to the type of x.

Examples

between(1:12, 7, 9)
#>  [1] FALSE FALSE FALSE FALSE FALSE FALSE  TRUE  TRUE  TRUE FALSE FALSE
#> [12] FALSE

x <- rnorm(1e2)
x[between(x, -1, 1)]
#>  [1] -0.11167789 -0.82966543 -0.21337340  0.63120389 -0.24830240
#>  [6] -0.36079721 -0.79251930  0.38474592 -0.65835274  0.54477256
#> [11]  0.91037579 -0.45563220  0.11153368  0.24191238  0.98591631
#> [16] -0.19238831 -0.61598002 -0.69266380  0.83182480  0.01652089
#> [21]  0.44643315  0.64329052 -0.51569678  0.14753276  0.90645918
#> [26]  0.69558399  0.67087742  0.53018143 -0.26003543 -0.43567031
#> [31]  0.42625768  0.68985084  0.78380166  0.90478648 -0.90002912
#> [36] -0.53905160 -0.14928555  0.07512613 -0.92093717 -0.30363898
#> [41] -0.22726179 -0.70401020  0.76952861  0.02528832  0.08889090
#> [46] -0.48761813 -0.21215013 -0.71891604 -0.25528080 -0.69750462
#> [51]  0.49232346  0.52245998 -0.63312867 -0.04506385  0.41588560
#> [56]  0.55898709 -0.02728837  0.19086657 -0.04613805  0.68891799
#> [61] -0.82065016  0.61355032  0.12418927  0.70289143 -0.58380083
#> [66] -0.77141015  0.96591993 -0.28505074  0.66187716  0.42199811
#> [71] -0.76409952

# On a tibble using `filter()`
filter(starwars, between(height, 100, 150))
#> # A tibble: 5 × 14
#>   name   height  mass hair_…¹ skin_…² eye_c…³ birth…⁴ sex   gender homew…⁵
#>   <chr>   <int> <dbl> <chr>   <chr>   <chr>     <dbl> <chr> <chr>  <chr>  
#> 1 Leia …    150    49 brown   light   brown        19 fema… femin… Aldera…
#> 2 Mon M…    150    NA auburn  fair    blue         48 fema… femin… Chandr…
#> 3 Watto     137    NA black   blue, … yellow       NA male  mascu… Toydar…
#> 4 Sebul…    112    40 none    grey, … orange       NA male  mascu… Malast…
#> 5 Gasga…    122    NA none    white,… black        NA male  mascu… Troiken
#> # … with 4 more variables: species <chr>, films <list>, vehicles <list>,
#> #   starships <list>, and abbreviated variable names ¹​hair_color,
#> #   ²​skin_color, ³​eye_color, ⁴​birth_year, ⁵​homeworld
#> # ℹ Use `colnames()` to see all variable names