Do values in a numeric vector fall in specified range?

This is a shortcut for x >= left & x <= right, implemented efficiently in C++ for local values, and translated to the appropriate SQL for remote tables.

Usage

between(x, left, right)

Arguments

x

A numeric vector of values

left, right

Boundary values (must be scalars).

Examples

between(1:12, 7, 9)
#>  [1] FALSE FALSE FALSE FALSE FALSE FALSE  TRUE  TRUE  TRUE FALSE FALSE
#> [12] FALSE

x <- rnorm(1e2)
x[between(x, -1, 1)]
#>  [1]  0.35838246  0.84514174 -0.19711524 -0.45505945 -0.04517161
#>  [6] -0.49550824 -0.14530117 -0.40661668 -0.61759603  0.11635680
#> [11] -0.25154959  0.73585961 -0.96874576 -0.81181956  0.58133312
#> [16] -0.73834773 -0.55212307 -0.93850913  0.87726626  0.71831447
#> [21]  0.26959690 -0.26159356  0.10064154  0.40112714  0.06000747
#> [26] -0.84960153 -0.06577643 -0.83368964 -0.91964683 -0.79998959
#> [31]  0.66523145 -0.08124753  0.24546857 -0.73101003 -0.13355699
#> [36] -0.59481967  0.55481458  0.37449605 -0.12138578 -0.46578358
#> [41] -0.96375990  0.35851775  0.27931179 -0.13394854 -0.77965630
#> [46] -0.62613090  0.30466107  0.02093655  0.39124550  0.15359712
#> [51]  0.22695697 -0.66931122 -0.22598752 -0.62552169  0.32762610
#> [56]  0.12782751 -0.82033011  0.57094314 -0.39700224

## Or on a tibble using filter
filter(starwars, between(height, 100, 150))
#> # A tibble: 5 × 14
#>   name       height  mass hair_color skin_color eye_color birth_year sex  
#>   <chr>       <int> <dbl> <chr>      <chr>      <chr>          <dbl> <chr>
#> 1 Leia Orga…    150    49 brown      light      brown             19 fema…
#> 2 Mon Mothma    150    NA auburn     fair       blue              48 fema…
#> 3 Watto         137    NA black      blue, grey yellow            NA male 
#> 4 Sebulba       112    40 none       grey, red  orange            NA male 
#> 5 Gasgano       122    NA none       white, bl… black             NA male 
#> # … with 6 more variables: gender <chr>, homeworld <chr>, species <chr>,
#> #   films <list>, vehicles <list>, starships <list>