Detect where values fall in a specified range

This is a shortcut for x >= left & x <= right, implemented for local vectors and translated to the appropriate SQL for remote tables.

Usage

between(x, left, right)

Arguments

x

A vector

left, right

Boundary values. Both left and right are recycled to the size of x.

Value

A logical vector the same size as x.

Details

x, left, and right are all cast to their common type before the comparison is made.

See also

Examples

between(1:12, 7, 9)
#>  [1] FALSE FALSE FALSE FALSE FALSE FALSE  TRUE  TRUE  TRUE FALSE FALSE
#> [12] FALSE

x <- rnorm(1e2)
x[between(x, -1, 1)]
#>  [1] -0.115683702 -0.528846563  0.545431103  0.897254311  0.399968043
#>  [6] -0.639974439  0.204475739 -0.792094306  0.896043409 -0.770671942
#> [11]  0.548574777  0.002335534 -0.704822718 -0.893263740 -0.861585343
#> [16]  0.945630792 -0.193562947 -0.336589315  0.901509899 -0.321576769
#> [21]  0.956016372 -0.205110580 -0.408584690 -0.084289370 -0.685101717
#> [26] -0.528846455  0.331740740  0.267051492  0.966024223  0.836655755
#> [31] -0.854641074 -0.092696938 -0.392833649 -0.178361346  0.180680831
#> [36] -0.954036532 -0.532159035 -0.801031056 -0.299631639  0.364692990
#> [41]  0.919116232 -0.403266859 -0.711258055  0.776471037 -0.997191837
#> [46] -0.479657022  0.474651156  0.252855888 -0.117494165 -0.030601619
#> [51] -0.302926533 -0.021962839 -0.029993566 -0.963922736  0.358005124
#> [56] -0.240833484  0.544729982 -0.889691607 -0.651900932  0.668622672
#> [61] -0.371676129  0.102626492 -0.752315127 -0.578149696  0.610011541
#> [66]  0.538634366  0.783168471 -0.491102397 -0.439912021  0.118641517
#> [71]  0.136506230  0.113658223

# On a tibble using `filter()`
filter(starwars, between(height, 100, 150))
#> # A tibble: 5 × 14
#>   name       height  mass hair_color skin_color eye_color birth_year sex  
#>   <chr>       <int> <dbl> <chr>      <chr>      <chr>          <dbl> <chr>
#> 1 Leia Orga…    150    49 brown      light      brown             19 fema…
#> 2 Mon Mothma    150    NA auburn     fair       blue              48 fema…
#> 3 Watto         137    NA black      blue, grey yellow            NA male 
#> 4 Sebulba       112    40 none       grey, red  orange            NA male 
#> 5 Gasgano       122    NA none       white, bl… black             NA male 
#> # ℹ 6 more variables: gender <chr>, homeworld <chr>, species <chr>,
#> #   films <list>, vehicles <list>, starships <list>