Detect where values fall in a specified range

This is a shortcut for x >= left & x <= right, implemented for local vectors and translated to the appropriate SQL for remote tables.

Usage

between(x, left, right)

Arguments

x

A vector

left, right

Boundary values. Both left and right are recycled to the size of x and are cast to the type of x.

Examples

between(1:12, 7, 9)
#>  [1] FALSE FALSE FALSE FALSE FALSE FALSE  TRUE  TRUE  TRUE FALSE FALSE
#> [12] FALSE

x <- rnorm(1e2)
x[between(x, -1, 1)]
#>  [1]  0.450052379 -0.352731128  0.110589005 -0.323436198  0.701830116
#>  [6] -0.535069679 -0.467093330  0.089805025  0.937990801  0.950678517
#> [11] -0.541989493  0.116391571 -0.292162198  0.765422553 -0.907625107
#> [16] -0.386579588  0.521158709 -0.181940889  0.952643539 -0.309918452
#> [21]  0.626869318 -0.117774378 -0.233756816  0.605825230  0.417248018
#> [26]  0.288163629 -0.134289013 -0.168540347  0.724202094  0.369370572
#> [31] -0.319967919  0.709575269  0.122457016  0.765536569 -0.201247103
#> [36] -0.330148559 -0.483773951  0.487504980 -0.862233105 -0.780826791
#> [41]  0.936359369  0.930133505  0.731570815  0.613106750  0.178326878
#> [46] -0.597755223  0.116027279  0.441582386 -0.308174582 -0.959081364
#> [51]  0.563328386  0.828879791  0.530936314 -0.369289919  0.374258211
#> [56] -0.727944695  0.121173050  0.075049271  0.368862772 -0.469428931
#> [61]  0.463782860 -0.541337524 -0.512984113 -0.522835350  0.996776858
#> [66] -0.001694233 -0.418304076 -0.276010330 -0.355976928

# On a tibble using `filter()`
filter(starwars, between(height, 100, 150))
#> # A tibble: 5 × 14
#>   name       height  mass hair_color skin_color eye_color birth_year sex  
#>   <chr>       <int> <dbl> <chr>      <chr>      <chr>          <dbl> <chr>
#> 1 Leia Orga…    150    49 brown      light      brown             19 fema…
#> 2 Mon Mothma    150    NA auburn     fair       blue              48 fema…
#> 3 Watto         137    NA black      blue, grey yellow            NA male 
#> 4 Sebulba       112    40 none       grey, red  orange            NA male 
#> 5 Gasgano       122    NA none       white, bl… black             NA male 
#> # … with 6 more variables: gender <chr>, homeworld <chr>, species <chr>,
#> #   films <list>, vehicles <list>, starships <list>