Detect where values fall in a specified range

This is a shortcut for x >= left & x <= right, implemented for local vectors and translated to the appropriate SQL for remote tables.

Usage

between(x, left, right)

Arguments

x

A vector

left, right

Boundary values. Both left and right are recycled to the size of x and are cast to the type of x.

Examples

between(1:12, 7, 9)
#>  [1] FALSE FALSE FALSE FALSE FALSE FALSE  TRUE  TRUE  TRUE FALSE FALSE
#> [12] FALSE

x <- rnorm(1e2)
x[between(x, -1, 1)]
#>  [1] -0.88901325 -0.68599741  0.03716127  0.08701639  0.12448210
#>  [6]  0.45973861  0.50056265  0.94645594 -0.17685277  0.11515926
#> [11] -0.23443983 -0.92229962  0.34275078 -0.96595797  0.27812373
#> [16] -0.49340198 -0.45939001 -0.54643386 -0.53480390  0.53352922
#> [21]  0.31844952 -0.27765663  0.08558934 -0.21374939 -0.21481154
#> [26]  0.52769256  0.49097406 -0.56330939 -0.73850536  0.26348345
#> [31] -0.92129883  0.80623764  0.91920348 -0.44788864 -0.20562398
#> [36] -0.64611959 -0.51544505 -0.22865008 -0.32234798  0.05310384
#> [41]  0.94117326  0.54316638  0.30408344 -0.69514194  0.23714152
#> [46] -0.43552484  0.27898086  0.28925659 -0.80672467 -0.02414667
#> [51]  0.02969439 -0.51744020 -0.71678519 -0.07655893 -0.82815519
#> [56]  0.21554574  0.00186936 -0.79158154 -0.04685671  0.54311065
#> [61]  0.26801735  0.45150529 -0.19538661 -0.20479160 -0.19572252
#> [66] -0.73865684

# On a tibble using `filter()`
filter(starwars, between(height, 100, 150))
#> # A tibble: 5 × 14
#>   name   height  mass hair_…¹ skin_…² eye_c…³ birth…⁴ sex   gender homew…⁵
#>   <chr>   <int> <dbl> <chr>   <chr>   <chr>     <dbl> <chr> <chr>  <chr>  
#> 1 Leia …    150    49 brown   light   brown        19 fema… femin… Aldera…
#> 2 Mon M…    150    NA auburn  fair    blue         48 fema… femin… Chandr…
#> 3 Watto     137    NA black   blue, … yellow       NA male  mascu… Toydar…
#> 4 Sebul…    112    40 none    grey, … orange       NA male  mascu… Malast…
#> 5 Gasga…    122    NA none    white,… black        NA male  mascu… Troiken
#> # … with 4 more variables: species <chr>, films <list>, vehicles <list>,
#> #   starships <list>, and abbreviated variable names ¹​hair_color,
#> #   ²​skin_color, ³​eye_color, ⁴​birth_year, ⁵​homeworld