Detect where values fall in a specified range

This is a shortcut for x >= left & x <= right, implemented for local vectors and translated to the appropriate SQL for remote tables.

Usage

between(x, left, right)

Arguments

x

A vector

left, right

Boundary values. Both left and right are recycled to the size of x and are cast to the type of x.

Examples

between(1:12, 7, 9)
#>  [1] FALSE FALSE FALSE FALSE FALSE FALSE  TRUE  TRUE  TRUE FALSE FALSE
#> [12] FALSE

x <- rnorm(1e2)
x[between(x, -1, 1)]
#>  [1] -0.1513338190  0.0287160428 -0.0315031597  0.6467014847 -0.9163867302
#>  [6]  0.9288712325 -0.8556001813 -0.5757412573  0.0146876304 -0.3459094009
#> [11] -0.1961224730  0.3555155577 -0.0621493327  0.3940803913 -0.5796302465
#> [16] -0.0443864240  0.6017924028  0.2146253184  0.7411796387  0.2079682748
#> [21]  0.4254329352 -0.4677791629  0.5068340782 -0.1752866923 -0.8335042244
#> [26]  0.7936630292  0.6844323636 -0.3756247643  0.6094696687  0.9268072405
#> [31] -0.8099153417  0.4222278591 -0.1173364307 -0.6525120636 -0.1532984232
#> [36]  0.2039925286 -0.1699447403 -0.5036189130  0.3672652656 -0.6018352117
#> [41]  0.9516207752 -0.0981093007 -0.3630067092  0.2509608319 -0.7519065924
#> [46] -0.4090697854  0.1316448084 -0.0007339527  0.7838693000  0.0430860627
#> [51] -0.3485587909 -0.5528356470  0.1711532070  0.9994257291  0.6638405271
#> [56]  0.6029757345 -0.4560832682 -0.7883431778  0.7128606689  0.0336162645
#> [61] -0.5168054411  0.2062152323  0.0547014886  0.7985854620 -0.4988340665
#> [66] -0.1541134673  0.3573718729 -0.5594035544  0.0331043671 -0.9339884868
#> [71] -0.5029439039

# On a tibble using `filter()`
filter(starwars, between(height, 100, 150))
#> # A tibble: 5 × 14
#>   name       height  mass hair_color skin_color eye_color birth_year sex  
#>   <chr>       <int> <dbl> <chr>      <chr>      <chr>          <dbl> <chr>
#> 1 Leia Orga…    150    49 brown      light      brown             19 fema…
#> 2 Mon Mothma    150    NA auburn     fair       blue              48 fema…
#> 3 Watto         137    NA black      blue, grey yellow            NA male 
#> 4 Sebulba       112    40 none       grey, red  orange            NA male 
#> 5 Gasgano       122    NA none       white, bl… black             NA male 
#> # … with 6 more variables: gender <chr>, homeworld <chr>, species <chr>,
#> #   films <list>, vehicles <list>, starships <list>