Detect where values fall in a specified range

This is a shortcut for x >= left & x <= right, implemented for local vectors and translated to the appropriate SQL for remote tables.

Usage

between(x, left, right)

Arguments

x

A vector

left, right

Boundary values. Both left and right are recycled to the size of x and are cast to the type of x.

Examples

between(1:12, 7, 9)
#>  [1] FALSE FALSE FALSE FALSE FALSE FALSE  TRUE  TRUE  TRUE FALSE FALSE
#> [12] FALSE

x <- rnorm(1e2)
x[between(x, -1, 1)]
#>  [1]  0.30636600 -0.34726867  0.26706829 -0.06567054 -0.22719082
#>  [6]  0.90483224 -0.30666988 -0.73507768  0.24039308 -0.01882698
#> [11] -0.68333335 -0.54448213 -0.75414127 -0.40214482  0.06967099
#> [16]  0.76307507 -0.98156755  0.09815358  0.85418822 -0.71984475
#> [21] -0.35371174 -0.31337150  0.85382733  0.28729577  0.81500954
#> [26] -0.33944496 -0.54675144 -0.10459325 -0.78749868 -0.79941767
#> [31]  0.86263651  0.91656964 -0.38215635 -0.55731977  0.33878393
#> [36] -0.27278930 -0.96917319  0.49312301 -0.26964590  0.87978668
#> [41] -0.14556653 -0.72187083  0.34877142 -0.58399820  0.52359983
#> [46]  0.57571652  0.83107669  0.09867946 -0.04730447  0.17430057
#> [51]  0.02600705 -0.99473160  0.23747621 -0.68541468 -0.05426174
#> [56]  0.67361189  0.77355288  0.53967441  0.05034704  0.42199702
#> [61] -0.27707036  0.84847291  0.98702663  0.28723906  0.04436389
#> [66]  0.43581935 -0.98177745 -0.59807256  0.10569225

# On a tibble using `filter()`
filter(starwars, between(height, 100, 150))
#> # A tibble: 5 × 14
#>   name       height  mass hair_color skin_color eye_color birth_year sex  
#>   <chr>       <int> <dbl> <chr>      <chr>      <chr>          <dbl> <chr>
#> 1 Leia Orga…    150    49 brown      light      brown             19 fema…
#> 2 Mon Mothma    150    NA auburn     fair       blue              48 fema…
#> 3 Watto         137    NA black      blue, grey yellow            NA male 
#> 4 Sebulba       112    40 none       grey, red  orange            NA male 
#> 5 Gasgano       122    NA none       white, bl… black             NA male 
#> # … with 6 more variables: gender <chr>, homeworld <chr>, species <chr>,
#> #   films <list>, vehicles <list>, starships <list>