Detect where values fall in a specified range

This is a shortcut for x >= left & x <= right, implemented for local vectors and translated to the appropriate SQL for remote tables.

Usage

between(x, left, right)

Arguments

x

A vector

left, right

Boundary values. Both left and right are recycled to the size of x and are cast to the type of x.

Examples

between(1:12, 7, 9)
#>  [1] FALSE FALSE FALSE FALSE FALSE FALSE  TRUE  TRUE  TRUE FALSE FALSE
#> [12] FALSE

x <- rnorm(1e2)
x[between(x, -1, 1)]
#>  [1]  0.405774135 -0.615447081  0.773391731 -0.445865209  0.183729795
#>  [6]  0.537107973  0.408348396 -0.695171309 -0.117930016  0.720843871
#> [11]  0.289457281 -0.559491834 -0.299778802  0.229686390  0.307390518
#> [16]  0.177441153 -0.243940886  0.482293229 -0.517105377  0.171952598
#> [21] -0.775020480 -0.671481014 -0.886449538 -0.096358301  0.086923767
#> [26]  0.371056896 -0.436959553 -0.478209684 -0.034590647 -0.840466552
#> [31] -0.222613203  0.151494953 -0.404299978  0.390712323  0.004090624
#> [36] -0.023766087 -0.284017731  0.230174414 -0.323955722 -0.458729055
#> [41]  0.362849788 -0.347284148  0.724757535  0.632890409  0.295914483
#> [46]  0.357609730 -0.853772307  0.553075409 -0.852832081  0.597690546
#> [51] -0.974736729  0.696859771 -0.846155581  0.043194358 -0.783143924
#> [56] -0.017037584  0.013282745 -0.555920752 -0.058255057 -0.250841505
#> [61]  0.741832525 -0.277428303  0.115338063 -0.528677728  0.622951953
#> [66] -0.277762063  0.064579302 -0.212421866  0.168330768  0.917101595
#> [71] -0.258831788  0.757084903 -0.605720395

# On a tibble using `filter()`
filter(starwars, between(height, 100, 150))
#> # A tibble: 5 × 14
#>   name   height  mass hair_…¹ skin_…² eye_c…³ birth…⁴ sex   gender homew…⁵
#>   <chr>   <int> <dbl> <chr>   <chr>   <chr>     <dbl> <chr> <chr>  <chr>  
#> 1 Leia …    150    49 brown   light   brown        19 fema… femin… Aldera…
#> 2 Mon M…    150    NA auburn  fair    blue         48 fema… femin… Chandr…
#> 3 Watto     137    NA black   blue, … yellow       NA male  mascu… Toydar…
#> 4 Sebul…    112    40 none    grey, … orange       NA male  mascu… Malast…
#> 5 Gasga…    122    NA none    white,… black        NA male  mascu… Troiken
#> # … with 4 more variables: species <chr>, films <list>, vehicles <list>,
#> #   starships <list>, and abbreviated variable names ¹​hair_color,
#> #   ²​skin_color, ³​eye_color, ⁴​birth_year, ⁵​homeworld