Do values in a numeric vector fall in specified range?

This is a shortcut for x >= left & x <= right, implemented efficiently in C++ for local values, and translated to the appropriate SQL for remote tables.

Usage

between(x, left, right)

Arguments

x

A numeric vector of values

left, right

Boundary values (must be scalars).

Examples

between(1:12, 7, 9)
#>  [1] FALSE FALSE FALSE FALSE FALSE FALSE  TRUE  TRUE  TRUE FALSE FALSE
#> [12] FALSE

x <- rnorm(1e2)
x[between(x, -1, 1)]
#>  [1] -0.78761486  0.78780294  0.78775055 -0.09153170 -0.54503256
#>  [6] -0.68277633  0.75164912 -0.40459968  0.41044633 -0.82641679
#> [11] -0.27660038 -0.18858428 -0.84613514 -0.52000057 -0.16651978
#> [16]  0.31554921  0.41310104 -0.80597005  0.95867263 -0.20756646
#> [21]  0.24007137  0.44495822 -0.28545593 -0.88409473 -0.23082020
#> [26]  0.97968318  0.07788861 -0.65780553  0.40286713 -0.78725426
#> [31] -0.06157306 -0.78863183 -0.14263737  0.14041015 -0.16661817
#> [36]  0.94174022  0.32795499 -0.41777054 -0.80983730 -0.25676083
#> [41]  0.81399193 -0.31021482  0.01583431 -0.06983452 -0.09934146
#> [46]  0.23745902 -0.16248330 -0.57426528 -0.78276305  0.81651250
#> [51] -0.23577305 -0.46726423 -0.64869233  0.52938798 -0.06449084
#> [56]  0.91317735  0.19360874 -0.09785793  0.08003160 -0.13086612
#> [61]  0.92778769 -0.56932974  0.55818127  0.17186555

## Or on a tibble using filter
filter(starwars, between(height, 100, 150))
#> # A tibble: 5 × 14
#>   name       height  mass hair_color skin_color eye_color birth_year sex  
#>   <chr>       <int> <dbl> <chr>      <chr>      <chr>          <dbl> <chr>
#> 1 Leia Orga…    150    49 brown      light      brown             19 fema…
#> 2 Mon Mothma    150    NA auburn     fair       blue              48 fema…
#> 3 Watto         137    NA black      blue, grey yellow            NA male 
#> 4 Sebulba       112    40 none       grey, red  orange            NA male 
#> 5 Gasgano       122    NA none       white, bl… black             NA male 
#> # … with 6 more variables: gender <chr>, homeworld <chr>, species <chr>,
#> #   films <list>, vehicles <list>, starships <list>