Do values in a numeric vector fall in specified range?

This is a shortcut for x >= left & x <= right, implemented efficiently in C++ for local values, and translated to the appropriate SQL for remote tables.

Usage

between(x, left, right)

Arguments

x

A numeric vector of values

left, right

Boundary values (must be scalars).

Examples

between(1:12, 7, 9)
#>  [1] FALSE FALSE FALSE FALSE FALSE FALSE  TRUE  TRUE  TRUE FALSE FALSE
#> [12] FALSE

x <- rnorm(1e2)
x[between(x, -1, 1)]
#>  [1] -0.778809329  0.097467595  0.317185512 -0.455037197  0.722127407
#>  [6]  0.251595735 -0.831697533  0.730532665  0.786318708 -0.529158020
#> [11] -0.835205434  0.101889577 -0.732198343 -0.565504124 -0.888034989
#> [16]  0.177022196 -0.624576137 -0.712126316 -0.787720856  0.761039937
#> [21]  0.186835110 -0.714356898 -0.337880209  0.098592950  0.729168084
#> [26] -0.577099086  0.160449902 -0.078488250 -0.424973380 -0.603550278
#> [31] -0.296779887  0.099883895 -0.241141847  0.332494350 -0.045449091
#> [36] -0.100020455 -0.705708161 -0.567581662 -0.203480877 -0.539230275
#> [41]  0.404846357  0.818339328 -0.686750060 -0.724011478  0.757661347
#> [46]  0.083637703 -0.278436797 -0.006921754 -0.487487444  0.530101462
#> [51]  0.235181120  0.211355888  0.985301211  0.447008096  0.635633731
#> [56] -0.189945850 -0.788549713  0.666442121 -0.641897164  0.599666520

## Or on a tibble using filter
filter(starwars, between(height, 100, 150))
#> # A tibble: 5 × 14
#>   name       height  mass hair_color skin_color eye_color birth_year sex  
#>   <chr>       <int> <dbl> <chr>      <chr>      <chr>          <dbl> <chr>
#> 1 Leia Orga…    150    49 brown      light      brown             19 fema…
#> 2 Mon Mothma    150    NA auburn     fair       blue              48 fema…
#> 3 Watto         137    NA black      blue, grey yellow            NA male 
#> 4 Sebulba       112    40 none       grey, red  orange            NA male 
#> 5 Gasgano       122    NA none       white, bl… black             NA male 
#> # … with 6 more variables: gender <chr>, homeworld <chr>, species <chr>,
#> #   films <list>, vehicles <list>, starships <list>