Detect where values fall in a specified range

This is a shortcut for x >= left & x <= right, implemented for local vectors and translated to the appropriate SQL for remote tables.

Usage

between(x, left, right)

Arguments

x

A vector

left, right

Boundary values. Both left and right are recycled to the size of x.

Value

A logical vector the same size as x.

Details

x, left, and right are all cast to their common type before the comparison is made.

See also

Examples

between(1:12, 7, 9)
#>  [1] FALSE FALSE FALSE FALSE FALSE FALSE  TRUE  TRUE  TRUE FALSE FALSE
#> [12] FALSE

x <- rnorm(1e2)
x[between(x, -1, 1)]
#>  [1] -0.85132459 -0.88208589 -0.73556547 -0.18127526 -0.32951065
#>  [6] -0.98631398  0.36652670  0.19627672 -0.53297505  0.58562815
#> [11] -0.81770823  0.29213703  0.37021036  0.57347933  0.75172965
#> [16] -0.24214717  0.92004516  0.95314700 -0.02861963 -0.13674918
#> [21] -0.29575685 -0.24859453 -0.23118020 -0.16794330 -0.73528188
#> [26] -0.31971717 -0.31212469  0.26001351  0.87960831  0.91254320
#> [31]  0.40037276  0.59768109 -0.03905560  0.19922626 -0.76789708
#> [36]  0.33497579  0.51358059  0.75268829 -0.91693135  0.09943548
#> [41]  0.07673404 -0.14616383 -0.49632932 -0.33812132  0.93296967
#> [46] -0.54502027 -0.61617765 -0.85992778 -0.12758225  0.62728321
#> [51]  0.55517801  0.37847893  0.25153878 -0.66005994 -0.45473127
#> [56]  0.62841423 -0.42927633 -0.32509914  0.53773339  0.04354723
#> [61] -0.43220089  0.82920387 -0.58016573  0.39533921 -0.36347645

# On a tibble using `filter()`
filter(starwars, between(height, 100, 150))
#> # A tibble: 5 × 14
#>   name   height  mass hair_…¹ skin_…² eye_c…³ birth…⁴ sex   gender homew…⁵
#>   <chr>   <int> <dbl> <chr>   <chr>   <chr>     <dbl> <chr> <chr>  <chr>  
#> 1 Leia …    150    49 brown   light   brown        19 fema… femin… Aldera…
#> 2 Mon M…    150    NA auburn  fair    blue         48 fema… femin… Chandr…
#> 3 Watto     137    NA black   blue, … yellow       NA male  mascu… Toydar…
#> 4 Sebul…    112    40 none    grey, … orange       NA male  mascu… Malast…
#> 5 Gasga…    122    NA none    white,… black        NA male  mascu… Troiken
#> # … with 4 more variables: species <chr>, films <list>, vehicles <list>,
#> #   starships <list>, and abbreviated variable names ¹​hair_color,
#> #   ²​skin_color, ³​eye_color, ⁴​birth_year, ⁵​homeworld