Do values in a numeric vector fall in specified range?

This is a shortcut for x >= left & x <= right, implemented efficiently in C++ for local values, and translated to the appropriate SQL for remote tables.

between(x, left, right)

Arguments

x
A numeric vector of values

left, right
Boundary values

Examples

between(1:12, 7, 9)

#>  [1] FALSE FALSE FALSE FALSE FALSE FALSE  TRUE  TRUE  TRUE FALSE FALSE FALSE

x <- rnorm(1e2)
x[between(x, -1, 1)]

#>  [1]  0.67638957 -0.23276176  0.63747287 -0.68326687 -0.97967535 -0.46251911
#>  [7]  0.74786883 -0.80534353  0.77599004  0.70359786  0.71540139  0.40152349
#> [13]  0.40427794 -0.77705556 -0.28045548  0.29738725 -0.23564615  0.04433469
#> [19] -0.32849189  0.46205231  0.69362435 -0.70859443 -0.20889688  0.35227917
#> [25] -0.77928540  0.70005531  0.11759261 -0.28362275  0.49790884  0.38624909
#> [31]  0.71752478  0.13191018 -0.92066340  0.26726762 -0.23854626 -0.78447669
#> [37] -0.82776894  0.94433756 -0.63969296 -0.21714460  0.57991478 -0.95260348
#> [43] -0.96767371  0.79245114 -0.93697260  0.86130054  0.62163991  0.48472168
#> [49] -0.44727427  0.54609708  0.24288873  0.16851210  0.23906467  0.23632142
#> [55] -0.25911917  0.64904570  0.84197042  0.26416593  0.09731070 -0.98868279
#> [61]  0.37043223