Copy a local data frame to a remote src
Source: R/copy-to.r

This function uploads a local data frame into a remote data source, creating the table definition as needed. Wherever possible, the new object will be temporary, limited to the current connection to the source.

copy_to(dest, df, name = deparse(substitute(df)), overwrite = FALSE, ...)

Arguments

dest
remote data source

df
local data frame

name
name for new remote table.

overwrite
If TRUE, will overwrite an existing table with name name. If FALSE, will throw an error if name already exists.

...
other parameters passed to methods.

Value

a tbl object in the remote source

See also

collect() for the opposite action; downloading remote data into a local dbl.

Examples

# NOT RUN {
iris2 <- dbplyr::src_memdb() %>% copy_to(iris, overwrite = TRUE)
iris2
# }