Keep rows that match a condition

The filter() function is used to subset a data frame, retaining all rows that satisfy your conditions. To be retained, the row must produce a value of TRUE for all conditions. Note that when a condition evaluates to NA the row will be dropped, unlike base subsetting with [.

Usage

filter(.data, ..., .preserve = FALSE)

Arguments

.data

A data frame, data frame extension (e.g. a tibble), or a lazy data frame (e.g. from dbplyr or dtplyr). See Methods, below, for more details.

...

<data-masking> Expressions that return a logical value, and are defined in terms of the variables in .data. If multiple expressions are included, they are combined with the & operator. Only rows for which all conditions evaluate to TRUE are kept.

.preserve

Relevant when the .data input is grouped. If .preserve = FALSE (the default), the grouping structure is recalculated based on the resulting data, otherwise the grouping is kept as is.

Value

An object of the same type as .data. The output has the following properties:

Rows are a subset of the input, but appear in the same order.

Columns are not modified.

The number of groups may be reduced (if .preserve is not TRUE).

Data frame attributes are preserved.

Details

The filter() function is used to subset the rows of .data, applying the expressions in ... to the column values to determine which rows should be retained. It can be applied to both grouped and ungrouped data (see group_by() and ungroup()). However, dplyr is not yet smart enough to optimise the filtering operation on grouped datasets that do not need grouped calculations. For this reason, filtering is often considerably faster on ungrouped data.

Useful filter functions

There are many functions and operators that are useful when constructing the expressions used to filter the data:

==, >, >= etc

&, |, !, xor()

is.na()

between(), near()

Grouped tibbles

Because filtering expressions are computed within groups, they may yield different results on grouped tibbles. This will be the case as soon as an aggregating, lagging, or ranking function is involved. Compare this ungrouped filtering:

starwars %>% filter(mass > mean(mass, na.rm = TRUE))

With the grouped equivalent:

starwars %>% group_by(gender) %>% filter(mass > mean(mass, na.rm = TRUE))

In the ungrouped version, filter() compares the value of mass in each row to the global average (taken over the whole data set), keeping only the rows with mass greater than this global average. In contrast, the grouped version calculates the average mass separately for each gender group, and keeps rows with mass greater than the relevant within-gender average.

Methods

This function is a generic, which means that packages can provide implementations (methods) for other classes. See the documentation of individual methods for extra arguments and differences in behaviour.

The following methods are currently available in loaded packages: dbplyr (tbl_lazy), dplyr (data.frame, ts) .

See also

Examples

# Filtering by one criterion
filter(starwars, species == "Human")
#> # A tibble: 35 × 14
#>    name  height  mass hair_…¹ skin_…² eye_c…³ birth…⁴ sex   gender homew…⁵
#>    <chr>  <int> <dbl> <chr>   <chr>   <chr>     <dbl> <chr> <chr>  <chr>  
#>  1 Luke…    172    77 blond   fair    blue       19   male  mascu… Tatooi…
#>  2 Dart…    202   136 none    white   yellow     41.9 male  mascu… Tatooi…
#>  3 Leia…    150    49 brown   light   brown      19   fema… femin… Aldera…
#>  4 Owen…    178   120 brown,… light   blue       52   male  mascu… Tatooi…
#>  5 Beru…    165    75 brown   light   blue       47   fema… femin… Tatooi…
#>  6 Bigg…    183    84 black   light   brown      24   male  mascu… Tatooi…
#>  7 Obi-…    182    77 auburn… fair    blue-g…    57   male  mascu… Stewjon
#>  8 Anak…    188    84 blond   fair    blue       41.9 male  mascu… Tatooi…
#>  9 Wilh…    180    NA auburn… fair    blue       64   male  mascu… Eriadu 
#> 10 Han …    180    80 brown   fair    brown      29   male  mascu… Corell…
#> # … with 25 more rows, 4 more variables: species <chr>, films <list>,
#> #   vehicles <list>, starships <list>, and abbreviated variable names
#> #   ¹​hair_color, ²​skin_color, ³​eye_color, ⁴​birth_year, ⁵​homeworld
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
filter(starwars, mass > 1000)
#> # A tibble: 1 × 14
#>   name   height  mass hair_…¹ skin_…² eye_c…³ birth…⁴ sex   gender homew…⁵
#>   <chr>   <int> <dbl> <chr>   <chr>   <chr>     <dbl> <chr> <chr>  <chr>  
#> 1 Jabba…    175  1358 NA      green-… orange      600 herm… mascu… Nal Hu…
#> # … with 4 more variables: species <chr>, films <list>, vehicles <list>,
#> #   starships <list>, and abbreviated variable names ¹​hair_color,
#> #   ²​skin_color, ³​eye_color, ⁴​birth_year, ⁵​homeworld
#> # ℹ Use `colnames()` to see all variable names

# Filtering by multiple criteria within a single logical expression
filter(starwars, hair_color == "none" & eye_color == "black")
#> # A tibble: 9 × 14
#>   name   height  mass hair_…¹ skin_…² eye_c…³ birth…⁴ sex   gender homew…⁵
#>   <chr>   <int> <dbl> <chr>   <chr>   <chr>     <dbl> <chr> <chr>  <chr>  
#> 1 Nien …    160    68 none    grey    black        NA male  mascu… Sullust
#> 2 Gasga…    122    NA none    white,… black        NA male  mascu… Troiken
#> 3 Kit F…    196    87 none    green   black        NA male  mascu… Glee A…
#> 4 Plo K…    188    80 none    orange  black        22 male  mascu… Dorin  
#> 5 Lama …    229    88 none    grey    black        NA male  mascu… Kamino 
#> 6 Taun …    213    NA none    grey    black        NA fema… femin… Kamino 
#> 7 Shaak…    178    57 none    red, b… black        NA fema… femin… Shili  
#> 8 Tion …    206    80 none    grey    black        NA male  mascu… Utapau 
#> 9 BB8        NA    NA none    none    black        NA none  mascu… NA     
#> # … with 4 more variables: species <chr>, films <list>, vehicles <list>,
#> #   starships <list>, and abbreviated variable names ¹​hair_color,
#> #   ²​skin_color, ³​eye_color, ⁴​birth_year, ⁵​homeworld
#> # ℹ Use `colnames()` to see all variable names
filter(starwars, hair_color == "none" | eye_color == "black")
#> # A tibble: 38 × 14
#>    name  height  mass hair_…¹ skin_…² eye_c…³ birth…⁴ sex   gender homew…⁵
#>    <chr>  <int> <dbl> <chr>   <chr>   <chr>     <dbl> <chr> <chr>  <chr>  
#>  1 Dart…    202   136 none    white   yellow     41.9 male  mascu… Tatooi…
#>  2 Gree…    173    74 NA      green   black      44   male  mascu… Rodia  
#>  3 IG-88    200   140 none    metal   red        15   none  mascu… NA     
#>  4 Bossk    190   113 none    green   red        53   male  mascu… Trando…
#>  5 Lobot    175    79 none    light   blue       37   male  mascu… Bespin 
#>  6 Ackb…    180    83 none    brown … orange     41   male  mascu… Mon Ca…
#>  7 Nien…    160    68 none    grey    black      NA   male  mascu… Sullust
#>  8 Nute…    191    90 none    mottle… red        NA   male  mascu… Cato N…
#>  9 Jar …    196    66 none    orange  orange     52   male  mascu… Naboo  
#> 10 Roos…    224    82 none    grey    orange     NA   male  mascu… Naboo  
#> # … with 28 more rows, 4 more variables: species <chr>, films <list>,
#> #   vehicles <list>, starships <list>, and abbreviated variable names
#> #   ¹​hair_color, ²​skin_color, ³​eye_color, ⁴​birth_year, ⁵​homeworld
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names

# When multiple expressions are used, they are combined using &
filter(starwars, hair_color == "none", eye_color == "black")
#> # A tibble: 9 × 14
#>   name   height  mass hair_…¹ skin_…² eye_c…³ birth…⁴ sex   gender homew…⁵
#>   <chr>   <int> <dbl> <chr>   <chr>   <chr>     <dbl> <chr> <chr>  <chr>  
#> 1 Nien …    160    68 none    grey    black        NA male  mascu… Sullust
#> 2 Gasga…    122    NA none    white,… black        NA male  mascu… Troiken
#> 3 Kit F…    196    87 none    green   black        NA male  mascu… Glee A…
#> 4 Plo K…    188    80 none    orange  black        22 male  mascu… Dorin  
#> 5 Lama …    229    88 none    grey    black        NA male  mascu… Kamino 
#> 6 Taun …    213    NA none    grey    black        NA fema… femin… Kamino 
#> 7 Shaak…    178    57 none    red, b… black        NA fema… femin… Shili  
#> 8 Tion …    206    80 none    grey    black        NA male  mascu… Utapau 
#> 9 BB8        NA    NA none    none    black        NA none  mascu… NA     
#> # … with 4 more variables: species <chr>, films <list>, vehicles <list>,
#> #   starships <list>, and abbreviated variable names ¹​hair_color,
#> #   ²​skin_color, ³​eye_color, ⁴​birth_year, ⁵​homeworld
#> # ℹ Use `colnames()` to see all variable names

# The filtering operation may yield different results on grouped
# tibbles because the expressions are computed within groups.
#
# The following filters rows where `mass` is greater than the
# global average:
starwars %>% filter(mass > mean(mass, na.rm = TRUE))
#> # A tibble: 10 × 14
#>    name  height  mass hair_…¹ skin_…² eye_c…³ birth…⁴ sex   gender homew…⁵
#>    <chr>  <int> <dbl> <chr>   <chr>   <chr>     <dbl> <chr> <chr>  <chr>  
#>  1 Dart…    202   136 none    white   yellow     41.9 male  mascu… Tatooi…
#>  2 Owen…    178   120 brown,… light   blue       52   male  mascu… Tatooi…
#>  3 Chew…    228   112 brown   unknown blue      200   male  mascu… Kashyy…
#>  4 Jabb…    175  1358 NA      green-… orange    600   herm… mascu… Nal Hu…
#>  5 Jek …    180   110 brown   fair    blue       NA   male  mascu… Bestin…
#>  6 IG-88    200   140 none    metal   red        15   none  mascu… NA     
#>  7 Bossk    190   113 none    green   red        53   male  mascu… Trando…
#>  8 Dext…    198   102 none    brown   yellow     NA   male  mascu… Ojom   
#>  9 Grie…    216   159 none    brown,… green,…    NA   male  mascu… Kalee  
#> 10 Tarf…    234   136 brown   brown   blue       NA   male  mascu… Kashyy…
#> # … with 4 more variables: species <chr>, films <list>, vehicles <list>,
#> #   starships <list>, and abbreviated variable names ¹​hair_color,
#> #   ²​skin_color, ³​eye_color, ⁴​birth_year, ⁵​homeworld
#> # ℹ Use `colnames()` to see all variable names

# Whereas this keeps rows with `mass` greater than the gender
# average:
starwars %>% group_by(gender) %>% filter(mass > mean(mass, na.rm = TRUE))
#> # A tibble: 14 × 14
#> # Groups:   gender [2]
#>    name         height   mass hair_…¹ skin_…² eye_c…³ birth…⁴ sex   gender
#>    <chr>         <int>  <dbl> <chr>   <chr>   <chr>     <dbl> <chr> <chr> 
#>  1 Darth Vader     202  136   none    white   yellow     41.9 male  mascu…
#>  2 Owen Lars       178  120   brown,… light   blue       52   male  mascu…
#>  3 Beru Whites…    165   75   brown   light   blue       47   fema… femin…
#>  4 Chewbacca       228  112   brown   unknown blue      200   male  mascu…
#>  5 Jabba Desil…    175 1358   NA      green-… orange    600   herm… mascu…
#>  6 Jek Tono Po…    180  110   brown   fair    blue       NA   male  mascu…
#>  7 IG-88           200  140   none    metal   red        15   none  mascu…
#>  8 Bossk           190  113   none    green   red        53   male  mascu…
#>  9 Ayla Secura     178   55   none    blue    hazel      48   fema… femin…
#> 10 Luminara Un…    170   56.2 black   yellow  blue       58   fema… femin…
#> 11 Zam Wesell      168   55   blonde  fair, … yellow     NA   fema… femin…
#> 12 Shaak Ti        178   57   none    red, b… black      NA   fema… femin…
#> 13 Grievous        216  159   none    brown,… green,…    NA   male  mascu…
#> 14 Tarfful         234  136   brown   brown   blue       NA   male  mascu…
#> # … with 5 more variables: homeworld <chr>, species <chr>, films <list>,
#> #   vehicles <list>, starships <list>, and abbreviated variable names
#> #   ¹​hair_color, ²​skin_color, ³​eye_color, ⁴​birth_year
#> # ℹ Use `colnames()` to see all variable names

# To refer to column names that are stored as strings, use the `.data` pronoun:
vars <- c("mass", "height")
cond <- c(80, 150)
starwars %>%
  filter(
    .data[[vars[[1]]]] > cond[[1]],
    .data[[vars[[2]]]] > cond[[2]]
  )
#> # A tibble: 21 × 14
#>    name  height  mass hair_…¹ skin_…² eye_c…³ birth…⁴ sex   gender homew…⁵
#>    <chr>  <int> <dbl> <chr>   <chr>   <chr>     <dbl> <chr> <chr>  <chr>  
#>  1 Dart…    202   136 none    white   yellow     41.9 male  mascu… Tatooi…
#>  2 Owen…    178   120 brown,… light   blue       52   male  mascu… Tatooi…
#>  3 Bigg…    183    84 black   light   brown      24   male  mascu… Tatooi…
#>  4 Anak…    188    84 blond   fair    blue       41.9 male  mascu… Tatooi…
#>  5 Chew…    228   112 brown   unknown blue      200   male  mascu… Kashyy…
#>  6 Jabb…    175  1358 NA      green-… orange    600   herm… mascu… Nal Hu…
#>  7 Jek …    180   110 brown   fair    blue       NA   male  mascu… Bestin…
#>  8 IG-88    200   140 none    metal   red        15   none  mascu… NA     
#>  9 Bossk    190   113 none    green   red        53   male  mascu… Trando…
#> 10 Ackb…    180    83 none    brown … orange     41   male  mascu… Mon Ca…
#> # … with 11 more rows, 4 more variables: species <chr>, films <list>,
#> #   vehicles <list>, starships <list>, and abbreviated variable names
#> #   ¹​hair_color, ²​skin_color, ³​eye_color, ⁴​birth_year, ⁵​homeworld
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
# Learn more in ?dplyr_data_masking