Subset rows using column values

The filter() function is used to subset a data frame, retaining all rows that satisfy your conditions. To be retained, the row must produce a value of TRUE for all conditions. Note that when a condition evaluates to NA the row will be dropped, unlike base subsetting with [.

filter(.data, ..., .preserve = FALSE)

Arguments

.data
A data frame, data frame extension (e.g. a tibble), or a lazy data frame (e.g. from dbplyr or dtplyr). See Methods, below, for more details.

...
<data-masking> Expressions that return a logical value, and are defined in terms of the variables in .data. If multiple expressions are included, they are combined with the & operator. Only rows for which all conditions evaluate to TRUE are kept.

.preserve
Relevant when the .data input is grouped. If .preserve = FALSE (the default), the grouping structure is recalculated based on the resulting data, otherwise the grouping is kept as is.

Value

An object of the same type as .data. The output has the following properties:

Rows are a subset of the input, but appear in the same order.

Columns are not modified.

The number of groups may be reduced (if .preserve is not TRUE).

Data frame attributes are preserved.

Details

The filter() function is used to subset the rows of .data, applying the expressions in ... to the column values to determine which rows should be retained. It can be applied to both grouped and ungrouped data (see group_by() and ungroup()). However, dplyr is not yet smart enough to optimise the filtering operation on grouped datasets that do not need grouped calculations. For this reason, filtering is often considerably faster on ungrouped data.

Useful filter functions

There are many functions and operators that are useful when constructing the expressions used to filter the data:

==, >, >= etc

&, |, !, xor()

is.na()

between(), near()

Grouped tibbles

Because filtering expressions are computed within groups, they may yield different results on grouped tibbles. This will be the case as soon as an aggregating, lagging, or ranking function is involved. Compare this ungrouped filtering:

starwars %>% filter(mass > mean(mass, na.rm = TRUE))

With the grouped equivalent:

starwars %>% group_by(gender) %>% filter(mass > mean(mass, na.rm = TRUE))

In the ungrouped version, filter() compares the value of mass in each row to the global average (taken over the whole data set), keeping only the rows with mass greater than this global average. In contrast, the grouped version calculates the average mass separately for each gender group, and keeps rows with mass greater than the relevant within-gender average.

Methods

This function is a generic, which means that packages can provide implementations (methods) for other classes. See the documentation of individual methods for extra arguments and differences in behaviour.

The following methods are currently available in loaded packages: dbplyr (tbl_lazy), dplyr (data.frame, ts) .

See also

Examples

# Filtering by one criterion
filter(starwars, species == "Human")

#> # A tibble: 35 x 14
#>    name    height  mass hair_color  skin_color eye_color birth_year sex   gender
#>    <chr>    <int> <dbl> <chr>       <chr>      <chr>          <dbl> <chr> <chr> 
#>  1 Luke S…    172    77 blond       fair       blue            19   male  mascu…
#>  2 Darth …    202   136 none        white      yellow          41.9 male  mascu…
#>  3 Leia O…    150    49 brown       light      brown           19   fema… femin…
#>  4 Owen L…    178   120 brown, grey light      blue            52   male  mascu…
#>  5 Beru W…    165    75 brown       light      blue            47   fema… femin…
#>  6 Biggs …    183    84 black       light      brown           24   male  mascu…
#>  7 Obi-Wa…    182    77 auburn, wh… fair       blue-gray       57   male  mascu…
#>  8 Anakin…    188    84 blond       fair       blue            41.9 male  mascu…
#>  9 Wilhuf…    180    NA auburn, gr… fair       blue            64   male  mascu…
#> 10 Han So…    180    80 brown       fair       brown           29   male  mascu…
#> # … with 25 more rows, and 5 more variables: homeworld <chr>, species <chr>,
#> #   films <list>, vehicles <list>, starships <list>

filter(starwars, mass > 1000)

#> # A tibble: 1 x 14
#>   name    height  mass hair_color skin_color  eye_color birth_year sex    gender
#>   <chr>    <int> <dbl> <chr>      <chr>       <chr>          <dbl> <chr>  <chr> 
#> 1 Jabba …    175  1358 NA         green-tan,… orange           600 herma… mascu…
#> # … with 5 more variables: homeworld <chr>, species <chr>, films <list>,
#> #   vehicles <list>, starships <list>

# Filtering by multiple criteria within a single logical expression
filter(starwars, hair_color == "none" & eye_color == "black")

#> # A tibble: 9 x 14
#>   name    height  mass hair_color skin_color   eye_color birth_year sex   gender
#>   <chr>    <int> <dbl> <chr>      <chr>        <chr>          <dbl> <chr> <chr> 
#> 1 Nien N…    160    68 none       grey         black             NA male  mascu…
#> 2 Gasgano    122    NA none       white, blue  black             NA male  mascu…
#> 3 Kit Fi…    196    87 none       green        black             NA male  mascu…
#> 4 Plo Ko…    188    80 none       orange       black             22 male  mascu…
#> 5 Lama Su    229    88 none       grey         black             NA male  mascu…
#> 6 Taun We    213    NA none       grey         black             NA fema… femin…
#> 7 Shaak …    178    57 none       red, blue, … black             NA fema… femin…
#> 8 Tion M…    206    80 none       grey         black             NA male  mascu…
#> 9 BB8         NA    NA none       none         black             NA none  mascu…
#> # … with 5 more variables: homeworld <chr>, species <chr>, films <list>,
#> #   vehicles <list>, starships <list>

filter(starwars, hair_color == "none" | eye_color == "black")

#> # A tibble: 38 x 14
#>    name    height  mass hair_color skin_color  eye_color birth_year sex   gender
#>    <chr>    <int> <dbl> <chr>      <chr>       <chr>          <dbl> <chr> <chr> 
#>  1 Darth …    202   136 none       white       yellow          41.9 male  mascu…
#>  2 Greedo     173    74 NA         green       black           44   male  mascu…
#>  3 IG-88      200   140 none       metal       red             15   none  mascu…
#>  4 Bossk      190   113 none       green       red             53   male  mascu…
#>  5 Lobot      175    79 none       light       blue            37   male  mascu…
#>  6 Ackbar     180    83 none       brown mott… orange          41   male  mascu…
#>  7 Nien N…    160    68 none       grey        black           NA   male  mascu…
#>  8 Nute G…    191    90 none       mottled gr… red             NA   male  mascu…
#>  9 Jar Ja…    196    66 none       orange      orange          52   male  mascu…
#> 10 Roos T…    224    82 none       grey        orange          NA   male  mascu…
#> # … with 28 more rows, and 5 more variables: homeworld <chr>, species <chr>,
#> #   films <list>, vehicles <list>, starships <list>

# When multiple expressions are used, they are combined using &
filter(starwars, hair_color == "none", eye_color == "black")

#> # A tibble: 9 x 14
#>   name    height  mass hair_color skin_color   eye_color birth_year sex   gender
#>   <chr>    <int> <dbl> <chr>      <chr>        <chr>          <dbl> <chr> <chr> 
#> 1 Nien N…    160    68 none       grey         black             NA male  mascu…
#> 2 Gasgano    122    NA none       white, blue  black             NA male  mascu…
#> 3 Kit Fi…    196    87 none       green        black             NA male  mascu…
#> 4 Plo Ko…    188    80 none       orange       black             22 male  mascu…
#> 5 Lama Su    229    88 none       grey         black             NA male  mascu…
#> 6 Taun We    213    NA none       grey         black             NA fema… femin…
#> 7 Shaak …    178    57 none       red, blue, … black             NA fema… femin…
#> 8 Tion M…    206    80 none       grey         black             NA male  mascu…
#> 9 BB8         NA    NA none       none         black             NA none  mascu…
#> # … with 5 more variables: homeworld <chr>, species <chr>, films <list>,
#> #   vehicles <list>, starships <list>

# The filtering operation may yield different results on grouped
# tibbles because the expressions are computed within groups.
#
# The following filters rows where `mass` is greater than the
# global average:
starwars %>% filter(mass > mean(mass, na.rm = TRUE))

#> # A tibble: 10 x 14
#>    name    height  mass hair_color  skin_color eye_color birth_year sex   gender
#>    <chr>    <int> <dbl> <chr>       <chr>      <chr>          <dbl> <chr> <chr> 
#>  1 Darth …    202   136 none        white      yellow          41.9 male  mascu…
#>  2 Owen L…    178   120 brown, grey light      blue            52   male  mascu…
#>  3 Chewba…    228   112 brown       unknown    blue           200   male  mascu…
#>  4 Jabba …    175  1358 NA          green-tan… orange         600   herm… mascu…
#>  5 Jek To…    180   110 brown       fair       blue            NA   male  mascu…
#>  6 IG-88      200   140 none        metal      red             15   none  mascu…
#>  7 Bossk      190   113 none        green      red             53   male  mascu…
#>  8 Dexter…    198   102 none        brown      yellow          NA   male  mascu…
#>  9 Grievo…    216   159 none        brown, wh… green, y…       NA   male  mascu…
#> 10 Tarfful    234   136 brown       brown      blue            NA   male  mascu…
#> # … with 5 more variables: homeworld <chr>, species <chr>, films <list>,
#> #   vehicles <list>, starships <list>

# Whereas this keeps rows with `mass` greater than the gender
# average:
starwars %>% group_by(gender) %>% filter(mass > mean(mass, na.rm = TRUE))

#> # A tibble: 14 x 14
#> # Groups:   gender [2]
#>    name   height   mass hair_color  skin_color eye_color birth_year sex   gender
#>    <chr>   <int>  <dbl> <chr>       <chr>      <chr>          <dbl> <chr> <chr> 
#>  1 Darth…    202  136   none        white      yellow          41.9 male  mascu…
#>  2 Owen …    178  120   brown, grey light      blue            52   male  mascu…
#>  3 Beru …    165   75   brown       light      blue            47   fema… femin…
#>  4 Chewb…    228  112   brown       unknown    blue           200   male  mascu…
#>  5 Jabba…    175 1358   NA          green-tan… orange         600   herm… mascu…
#>  6 Jek T…    180  110   brown       fair       blue            NA   male  mascu…
#>  7 IG-88     200  140   none        metal      red             15   none  mascu…
#>  8 Bossk     190  113   none        green      red             53   male  mascu…
#>  9 Ayla …    178   55   none        blue       hazel           48   fema… femin…
#> 10 Lumin…    170   56.2 black       yellow     blue            58   fema… femin…
#> 11 Zam W…    168   55   blonde      fair, gre… yellow          NA   fema… femin…
#> 12 Shaak…    178   57   none        red, blue… black           NA   fema… femin…
#> 13 Griev…    216  159   none        brown, wh… green, y…       NA   male  mascu…
#> 14 Tarff…    234  136   brown       brown      blue            NA   male  mascu…
#> # … with 5 more variables: homeworld <chr>, species <chr>, films <list>,
#> #   vehicles <list>, starships <list>

# To refer to column names that are stored as strings, use the `.data` pronoun:
vars <- c("mass", "height")
cond <- c(80, 150)
starwars %>%
  filter(
    .data[[vars[[1]]]] > cond[[1]],
    .data[[vars[[2]]]] > cond[[2]]
  )

#> # A tibble: 21 x 14
#>    name    height  mass hair_color  skin_color eye_color birth_year sex   gender
#>    <chr>    <int> <dbl> <chr>       <chr>      <chr>          <dbl> <chr> <chr> 
#>  1 Darth …    202   136 none        white      yellow          41.9 male  mascu…
#>  2 Owen L…    178   120 brown, grey light      blue            52   male  mascu…
#>  3 Biggs …    183    84 black       light      brown           24   male  mascu…
#>  4 Anakin…    188    84 blond       fair       blue            41.9 male  mascu…
#>  5 Chewba…    228   112 brown       unknown    blue           200   male  mascu…
#>  6 Jabba …    175  1358 NA          green-tan… orange         600   herm… mascu…
#>  7 Jek To…    180   110 brown       fair       blue            NA   male  mascu…
#>  8 IG-88      200   140 none        metal      red             15   none  mascu…
#>  9 Bossk      190   113 none        green      red             53   male  mascu…
#> 10 Ackbar     180    83 none        brown mot… orange          41   male  mascu…
#> # … with 11 more rows, and 5 more variables: homeworld <chr>, species <chr>,
#> #   films <list>, vehicles <list>, starships <list>

# Learn more in ?dplyr_data_masking