funs {dplyr}	R Documentation

Create a list of function calls

Description

funs() is deprecated; please use list() instead. We deprecated this function because it provided a unique way of specifying anonymous functions, rather than adopting the conventions used by purrr and other packages in the tidyverse.

Usage

funs(..., .args = list())

Arguments

...
<data-masking> A list of functions specified by:

Their name, "mean"

The function itself, mean

A call to the function with . as a dummy argument, mean(., na.rm = TRUE)

The following notations are not supported, see examples:

An anonymous function, function(x) mean(x, na.rm = TRUE)

An anonymous function in purrr notation, ~mean(., na.rm = TRUE)

.args, args
A named list of additional arguments to be added to all function calls. As funs() is being deprecated, use other methods to supply arguments: ... argument in scoped verbs or make own functions with purrr::partial().

Examples

funs("mean", mean(., na.rm = TRUE))
# ->
list(mean = mean, mean = ~ mean(.x, na.rm = TRUE))

funs(m1 = mean, m2 = "mean", m3 = mean(., na.rm = TRUE))
# ->
list(m1 = mean, m2 = "mean", m3 = ~ mean(.x, na.rm = TRUE))

[Package dplyr version 1.1.3 Index]