distinct_prepare {dplyr}	R Documentation

Same basic philosophy as group_by_prepare(): lazy_dots comes in, list of data and vars (character vector) comes out.

Description

*_prepare() performs standard manipulation that is needed prior to actual data processing. They are only be needed by packages that implement dplyr backends.

Usage

distinct_prepare(
  .data,
  vars,
  group_vars = character(),
  .keep_all = FALSE,
  caller_env = caller_env(2)
)

group_by_prepare(
  .data,
  ...,
  caller_env = caller_env(2),
  .add = FALSE,
  .dots = deprecated(),
  add = deprecated()
)

Value

A list

data
Modified tbl

groups
Modified groups

[Package dplyr version 1.0.7 Index]