grouped_df {dplyr}	R Documentation

A grouped data frame.

Description

The easiest way to create a grouped data frame is to call the group_by() method on a data frame or tbl: this will take care of capturing the unevaluated expressions for you.

These functions are designed for programmatic use. For data analysis purposes see group_data() for the accessor functions that retrieve various metadata from a grouped data frames.

Usage

grouped_df(data, vars, drop = group_by_drop_default(data))

is.grouped_df(x)

is_grouped_df(x)

Arguments

data
a tbl or data frame.

vars
A character vector.

drop
When .drop = TRUE, empty groups are dropped.

[Package dplyr version 1.0.7 Index]