A grouped data frame.

The easiest way to create a grouped data frame is to call the group_by() method on a data frame or tbl: this will take care of capturing the unevaluated expressions for you.

See group_data() for the accessor functions that retrieve various metadata from a grouped_df.

grouped_df(data, vars, drop = FALSE)

is.grouped_df(x)

is_grouped_df(x)

Arguments

data
a tbl or data frame.

vars
A character vector.

drop
When .drop = TRUE, empty groups are dropped.