location {dplyr}	R Documentation

Print the location in memory of a data frame

Description

This is useful for understand how and when dplyr makes copies of data frames

Usage

location(df)

changes(x, y)

Arguments

df
a data frame

x, y
two data frames to compare

Examples

location(mtcars)
# ->
lobstr::ref(mtcars)

mtcars2 <- mutate(mtcars, cyl2 = cyl * 2)
# ->
lobstr::ref(mtcars2)

changes(mtcars, mtcars2)
# ->
lobstr::ref(mtcars, mtcars2)

[Package dplyr version 1.0.7 Index]