Print the location in memory of a data frame

This is useful for understand how and when dplyr makes copies of data frames

Usage

location(df)

changes(x, y)

Arguments

df

a data frame

x, y

two data frames to compare

Examples

location(mtcars)
#> Warning: `location()` was deprecated in dplyr 1.0.0.
#> Please use `lobst::ref()` instead.
#> This warning is displayed once every 8 hours.
#> Call `lifecycle::last_lifecycle_warnings()` to see where this warning was generated.
#> <0x55558a43cd08>
#> Variables:
#>  * mpg:       <0x55558368df70>
#>  * cyl:       <0x55558a45f020>
#>  * disp:      <0x55558a556590>
#>  * hp:        <0x55558a6fd8a0>
#>  * drat:      <0x55558a62c1a0>
#>  * wt:        <0x55558a6354f0>
#>  * qsec:      <0x5555879fa2b0>
#>  * vs:        <0x5555876ecbc0>
#>  * am:        <0x5555875e7e80>
#>  * gear:      <0x55558964f7e0>
#>  * carb:      <0x555589197870>
#> Attributes:
#>  * names:     <0x55558a43caf8>
#>  * row.names: <0x555586edc060>
#>  * class:     <0x5555892b6ca8>
# ->
lobstr::ref(mtcars)
#> █ [1:0x55558a43cd08] <df[,11]> 
#> ├─mpg = [2:0x55558368df70] <dbl> 
#> ├─cyl = [3:0x55558a45f020] <dbl> 
#> ├─disp = [4:0x55558a556590] <dbl> 
#> ├─hp = [5:0x55558a6fd8a0] <dbl> 
#> ├─drat = [6:0x55558a62c1a0] <dbl> 
#> ├─wt = [7:0x55558a6354f0] <dbl> 
#> ├─qsec = [8:0x5555879fa2b0] <dbl> 
#> ├─vs = [9:0x5555876ecbc0] <dbl> 
#> ├─am = [10:0x5555875e7e80] <dbl> 
#> ├─gear = [11:0x55558964f7e0] <dbl> 
#> └─carb = [12:0x555589197870] <dbl> 

mtcars2 <- mutate(mtcars, cyl2 = cyl * 2)
# ->
lobstr::ref(mtcars2)
#> █ [1:0x55558a37b458] <df[,12]> 
#> ├─mpg = [2:0x55558368df70] <dbl> 
#> ├─cyl = [3:0x55558a45f020] <dbl> 
#> ├─disp = [4:0x55558a556590] <dbl> 
#> ├─hp = [5:0x55558a6fd8a0] <dbl> 
#> ├─drat = [6:0x55558a62c1a0] <dbl> 
#> ├─wt = [7:0x55558a6354f0] <dbl> 
#> ├─qsec = [8:0x5555879fa2b0] <dbl> 
#> ├─vs = [9:0x5555876ecbc0] <dbl> 
#> ├─am = [10:0x5555875e7e80] <dbl> 
#> ├─gear = [11:0x55558964f7e0] <dbl> 
#> ├─carb = [12:0x555589197870] <dbl> 
#> └─cyl2 = [13:0x55558ac104c0] <dbl> 

changes(mtcars, mtcars2)
#> Warning: `changes()` was deprecated in dplyr 1.0.0.
#> Please use `lobstr::ref()` instead.
#> This warning is displayed once every 8 hours.
#> Call `lifecycle::last_lifecycle_warnings()` to see where this warning was generated.
#> Changed variables:
#>       old     new           
#> cyl2  <added> 0x55558ac104c0
#> 
#> Changed attributes:
#>       old            new           
#> names 0x55558a43caf8 0x55558a37a538
# ->
lobstr::ref(mtcars, mtcars2)
#> █ [1:0x55558a43cd08] <df[,11]> 
#> ├─mpg = [2:0x55558368df70] <dbl> 
#> ├─cyl = [3:0x55558a45f020] <dbl> 
#> ├─disp = [4:0x55558a556590] <dbl> 
#> ├─hp = [5:0x55558a6fd8a0] <dbl> 
#> ├─drat = [6:0x55558a62c1a0] <dbl> 
#> ├─wt = [7:0x55558a6354f0] <dbl> 
#> ├─qsec = [8:0x5555879fa2b0] <dbl> 
#> ├─vs = [9:0x5555876ecbc0] <dbl> 
#> ├─am = [10:0x5555875e7e80] <dbl> 
#> ├─gear = [11:0x55558964f7e0] <dbl> 
#> └─carb = [12:0x555589197870] <dbl> 
#>  
#> █ [13:0x55558a37b458] <df[,12]> 
#> ├─mpg = [2:0x55558368df70] 
#> ├─cyl = [3:0x55558a45f020] 
#> ├─disp = [4:0x55558a556590] 
#> ├─hp = [5:0x55558a6fd8a0] 
#> ├─drat = [6:0x55558a62c1a0] 
#> ├─wt = [7:0x55558a6354f0] 
#> ├─qsec = [8:0x5555879fa2b0] 
#> ├─vs = [9:0x5555876ecbc0] 
#> ├─am = [10:0x5555875e7e80] 
#> ├─gear = [11:0x55558964f7e0] 
#> ├─carb = [12:0x555589197870] 
#> └─cyl2 = [14:0x55558ac104c0] <dbl>