Print the location in memory of a data frame

This is useful for understand how and when dplyr makes copies of data frames

location(df)

changes(x, y)

Arguments

df
a data frame

x, y
two data frames to compare

Examples

location(mtcars)

#> Warning: `location()` is deprecated as of dplyr 1.0.0.
#> Please use `lobst::ref()` instead.
#> This warning is displayed once every 8 hours.
#> Call `lifecycle::last_warnings()` to see where this warning was generated.

#> <0x7fb26258bab8>
#> Variables:
#>  * mpg:       <0x7fb261701df0>
#>  * cyl:       <0x7fb261702de0>
#>  * disp:      <0x7fb2617e6d40>
#>  * hp:        <0x7fb2617dfa20>
#>  * drat:      <0x7fb2617f7870>
#>  * wt:        <0x7fb2617fb2a0>
#>  * qsec:      <0x7fb26560c440>
#>  * vs:        <0x7fb261729ed0>
#>  * am:        <0x7fb26172a500>
#>  * gear:      <0x7fb2617116d0>
#>  * carb:      <0x7fb261713600>
#> Attributes:
#>  * names:     <0x7fb26258ba08>
#>  * row.names: <0x7fb261720ee0>
#>  * class:     <0x7fb263d2b560>

# ->
lobstr::ref(mtcars)

#> █ [1:0x7fb26258bab8] <df[,11]> 
#> ├─mpg = [2:0x7fb261701df0] <dbl> 
#> ├─cyl = [3:0x7fb261702de0] <dbl> 
#> ├─disp = [4:0x7fb2617e6d40] <dbl> 
#> ├─hp = [5:0x7fb2617dfa20] <dbl> 
#> ├─drat = [6:0x7fb2617f7870] <dbl> 
#> ├─wt = [7:0x7fb2617fb2a0] <dbl> 
#> ├─qsec = [8:0x7fb26560c440] <dbl> 
#> ├─vs = [9:0x7fb261729ed0] <dbl> 
#> ├─am = [10:0x7fb26172a500] <dbl> 
#> ├─gear = [11:0x7fb2617116d0] <dbl> 
#> └─carb = [12:0x7fb261713600] <dbl> 

mtcars2 <- mutate(mtcars, cyl2 = cyl * 2)
# ->
lobstr::ref(mtcars2)

#> █ [1:0x7fb269a57c88] <df[,12]> 
#> ├─mpg = [2:0x7fb261701df0] <dbl> 
#> ├─cyl = [3:0x7fb261702de0] <dbl> 
#> ├─disp = [4:0x7fb2617e6d40] <dbl> 
#> ├─hp = [5:0x7fb2617dfa20] <dbl> 
#> ├─drat = [6:0x7fb2617f7870] <dbl> 
#> ├─wt = [7:0x7fb2617fb2a0] <dbl> 
#> ├─qsec = [8:0x7fb26560c440] <dbl> 
#> ├─vs = [9:0x7fb261729ed0] <dbl> 
#> ├─am = [10:0x7fb26172a500] <dbl> 
#> ├─gear = [11:0x7fb2617116d0] <dbl> 
#> ├─carb = [12:0x7fb261713600] <dbl> 
#> └─cyl2 = [13:0x7fb2655df8c0] <dbl> 

changes(mtcars, mtcars2)

#> Warning: `changes()` is deprecated as of dplyr 1.0.0.
#> Please use `lobstr::ref()` instead.
#> This warning is displayed once every 8 hours.
#> Call `lifecycle::last_warnings()` to see where this warning was generated.

#> Changed variables:
#>       old     new           
#> cyl2  <added> 0x7fb2655df8c0
#> 
#> Changed attributes:
#>       old            new           
#> names 0x7fb26258ba08 0x7fb269a57bd8

# ->
lobstr::ref(mtcars, mtcars2)

#> █ [1:0x7fb26258bab8] <df[,11]> 
#> ├─mpg = [2:0x7fb261701df0] <dbl> 
#> ├─cyl = [3:0x7fb261702de0] <dbl> 
#> ├─disp = [4:0x7fb2617e6d40] <dbl> 
#> ├─hp = [5:0x7fb2617dfa20] <dbl> 
#> ├─drat = [6:0x7fb2617f7870] <dbl> 
#> ├─wt = [7:0x7fb2617fb2a0] <dbl> 
#> ├─qsec = [8:0x7fb26560c440] <dbl> 
#> ├─vs = [9:0x7fb261729ed0] <dbl> 
#> ├─am = [10:0x7fb26172a500] <dbl> 
#> ├─gear = [11:0x7fb2617116d0] <dbl> 
#> └─carb = [12:0x7fb261713600] <dbl> 
#>  
#> █ [13:0x7fb269a57c88] <df[,12]> 
#> ├─mpg = [2:0x7fb261701df0] 
#> ├─cyl = [3:0x7fb261702de0] 
#> ├─disp = [4:0x7fb2617e6d40] 
#> ├─hp = [5:0x7fb2617dfa20] 
#> ├─drat = [6:0x7fb2617f7870] 
#> ├─wt = [7:0x7fb2617fb2a0] 
#> ├─qsec = [8:0x7fb26560c440] 
#> ├─vs = [9:0x7fb261729ed0] 
#> ├─am = [10:0x7fb26172a500] 
#> ├─gear = [11:0x7fb2617116d0] 
#> ├─carb = [12:0x7fb261713600] 
#> └─cyl2 = [14:0x7fb2655df8c0] <dbl>