Print the location in memory of a data frame

This is useful for understand how and when dplyr makes copies of data frames

location(df)

changes(x, y)

Arguments

df
a data frame

x, y
two data frames to compare

Examples

location(mtcars)

#> Warning: `location()` is deprecated as of dplyr 1.0.0.
#> Please use `lobst::ref()` instead.
#> This warning is displayed once every 8 hours.
#> Call `lifecycle::last_warnings()` to see where this warning was generated.

#> <0x7fde2c636368>
#> Variables:
#>  * mpg:       <0x7fde27439420>
#>  * cyl:       <0x7fde274afb80>
#>  * disp:      <0x7fde274487f0>
#>  * hp:        <0x7fde2b91eca0>
#>  * drat:      <0x7fde274abb70>
#>  * wt:        <0x7fde274119c0>
#>  * qsec:      <0x7fde274a5dd0>
#>  * vs:        <0x7fde27462130>
#>  * am:        <0x7fde2b94ee40>
#>  * gear:      <0x7fde274381f0>
#>  * carb:      <0x7fde274fa0f0>
#> Attributes:
#>  * names:     <0x7fde2c636418>
#>  * row.names: <0x7fde274fc5b0>
#>  * class:     <0x7fde2dcbd600>

# ->
lobstr::ref(mtcars)

#> █ [1:0x7fde2c636368] <df[,11]> 
#> ├─mpg = [2:0x7fde27439420] <dbl> 
#> ├─cyl = [3:0x7fde274afb80] <dbl> 
#> ├─disp = [4:0x7fde274487f0] <dbl> 
#> ├─hp = [5:0x7fde2b91eca0] <dbl> 
#> ├─drat = [6:0x7fde274abb70] <dbl> 
#> ├─wt = [7:0x7fde274119c0] <dbl> 
#> ├─qsec = [8:0x7fde274a5dd0] <dbl> 
#> ├─vs = [9:0x7fde27462130] <dbl> 
#> ├─am = [10:0x7fde2b94ee40] <dbl> 
#> ├─gear = [11:0x7fde274381f0] <dbl> 
#> └─carb = [12:0x7fde274fa0f0] <dbl> 

mtcars2 <- mutate(mtcars, cyl2 = cyl * 2)
# ->
lobstr::ref(mtcars2)

#> █ [1:0x7fde2c572ab8] <df[,12]> 
#> ├─mpg = [2:0x7fde27439420] <dbl> 
#> ├─cyl = [3:0x7fde274afb80] <dbl> 
#> ├─disp = [4:0x7fde274487f0] <dbl> 
#> ├─hp = [5:0x7fde2b91eca0] <dbl> 
#> ├─drat = [6:0x7fde274abb70] <dbl> 
#> ├─wt = [7:0x7fde274119c0] <dbl> 
#> ├─qsec = [8:0x7fde274a5dd0] <dbl> 
#> ├─vs = [9:0x7fde27462130] <dbl> 
#> ├─am = [10:0x7fde2b94ee40] <dbl> 
#> ├─gear = [11:0x7fde274381f0] <dbl> 
#> ├─carb = [12:0x7fde274fa0f0] <dbl> 
#> └─cyl2 = [13:0x7fde277eb520] <dbl> 

changes(mtcars, mtcars2)

#> Warning: `changes()` is deprecated as of dplyr 1.0.0.
#> Please use `lobstr::ref()` instead.
#> This warning is displayed once every 8 hours.
#> Call `lifecycle::last_warnings()` to see where this warning was generated.

#> Changed variables:
#>       old     new           
#> cyl2  <added> 0x7fde277eb520
#> 
#> Changed attributes:
#>       old            new           
#> names 0x7fde2c636418 0x7fde2c660d08

# ->
lobstr::ref(mtcars, mtcars2)

#> █ [1:0x7fde2c636368] <df[,11]> 
#> ├─mpg = [2:0x7fde27439420] <dbl> 
#> ├─cyl = [3:0x7fde274afb80] <dbl> 
#> ├─disp = [4:0x7fde274487f0] <dbl> 
#> ├─hp = [5:0x7fde2b91eca0] <dbl> 
#> ├─drat = [6:0x7fde274abb70] <dbl> 
#> ├─wt = [7:0x7fde274119c0] <dbl> 
#> ├─qsec = [8:0x7fde274a5dd0] <dbl> 
#> ├─vs = [9:0x7fde27462130] <dbl> 
#> ├─am = [10:0x7fde2b94ee40] <dbl> 
#> ├─gear = [11:0x7fde274381f0] <dbl> 
#> └─carb = [12:0x7fde274fa0f0] <dbl> 
#>  
#> █ [13:0x7fde2c572ab8] <df[,12]> 
#> ├─mpg = [2:0x7fde27439420] 
#> ├─cyl = [3:0x7fde274afb80] 
#> ├─disp = [4:0x7fde274487f0] 
#> ├─hp = [5:0x7fde2b91eca0] 
#> ├─drat = [6:0x7fde274abb70] 
#> ├─wt = [7:0x7fde274119c0] 
#> ├─qsec = [8:0x7fde274a5dd0] 
#> ├─vs = [9:0x7fde27462130] 
#> ├─am = [10:0x7fde2b94ee40] 
#> ├─gear = [11:0x7fde274381f0] 
#> ├─carb = [12:0x7fde274fa0f0] 
#> └─cyl2 = [14:0x7fde277eb520] <dbl>