Create, modify, and delete columns

mutate() adds new variables and preserves existing ones; transmute() adds new variables and drops existing ones. New variables overwrite existing variables of the same name. Variables can be removed by setting their value to NULL.

mutate(.data, ...)

transmute(.data, ...)

Arguments

.data
A data frame, data frame extension (e.g. a tibble), or a lazy data frame (e.g. from dbplyr or dtplyr). See Methods, below, for more details.

...
<tidy-eval> Name-value pairs of expressions, each with length 1 or the same length as the number of rows in the group (if using group_by()) or in the entire input (if not using groups). The name of each argument will be the name of a new variable, and the value will be its corresponding value. Use a NULL value in mutate to drop a variable. New variables overwrite existing variables of the same name.

Value

An object of the same type as .data.

For mutate():

Rows are not affected.

Existing columns will be preserved unless explicitly modified.

New columns will be added to the right of existing columns.

Columns given value NULL will be removed

Groups will be recomputed if a grouping variable is mutated.

Data frame attributes are preserved.

For transmute():

Rows are not affected.

Apart from grouping variables, existing columns will be remove unless explicitly kept.

Column order matches order of expressions.

Groups will be recomputed if a grouping variable is mutated.

Data frame attributes are preserved.

Useful mutate functions

+, -, log(), etc., for their usual mathematical meanings

lead(), lag()

dense_rank(), min_rank(), percent_rank(), row_number(), cume_dist(), ntile()

cumsum(), cummean(), cummin(), cummax(), cumany(), cumall()

na_if(), coalesce()

if_else(), recode(), case_when()

Grouped tibbles

Because mutating expressions are computed within groups, they may yield different results on grouped tibbles. This will be the case as soon as an aggregating, lagging, or ranking function is involved. Compare this ungrouped mutate:

starwars %>%
  mutate(mass / mean(mass, na.rm = TRUE)) %>%
  pull()

With the grouped equivalent:

starwars %>%
  group_by(gender) %>%
  mutate(mass / mean(mass, na.rm = TRUE)) %>%
  pull()

The former normalises mass by the global average whereas the latter normalises by the averages within gender levels.

Scoped mutation and transmutation

The three scoped variants of mutate() (mutate_all(), mutate_if() and mutate_at()) and the three variants of transmute() (transmute_all(), transmute_if(), transmute_at()) make it easy to apply a transformation to a selection of variables.

Methods

These function are generics, which means that packages can provide implementations (methods) for other classes. See the documentation of individual methods for extra arguments and differences in behaviour.

Methods available in currently loaded packages:

mutate(): dbplyr (tbl_lazy), dplyr (data.frame, default) .

transmute(): dbplyr (tbl_lazy), dplyr (data.frame) .

See also

Examples

# Newly created variables are available immediately
mtcars %>% as_tibble() %>% mutate(
  cyl2 = cyl * 2,
  cyl4 = cyl2 * 2
)

#> # A tibble: 32 x 13
#>      mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb  cyl2  cyl4
#>    <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1  21       6  160    110  3.9   2.62  16.5     0     1     4     4    12    24
#>  2  21       6  160    110  3.9   2.88  17.0     0     1     4     4    12    24
#>  3  22.8     4  108     93  3.85  2.32  18.6     1     1     4     1     8    16
#>  4  21.4     6  258    110  3.08  3.22  19.4     1     0     3     1    12    24
#>  5  18.7     8  360    175  3.15  3.44  17.0     0     0     3     2    16    32
#>  6  18.1     6  225    105  2.76  3.46  20.2     1     0     3     1    12    24
#>  7  14.3     8  360    245  3.21  3.57  15.8     0     0     3     4    16    32
#>  8  24.4     4  147.    62  3.69  3.19  20       1     0     4     2     8    16
#>  9  22.8     4  141.    95  3.92  3.15  22.9     1     0     4     2     8    16
#> 10  19.2     6  168.   123  3.92  3.44  18.3     1     0     4     4    12    24
#> # … with 22 more rows

# As well as adding new variables, you can use mutate() to
# remove variables and modify existing variables.
mtcars %>% as_tibble() %>% mutate(
  mpg = NULL,
  disp = disp * 0.0163871 # convert to litres
)

#> # A tibble: 32 x 10
#>      cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#>    <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1     6  2.62   110  3.9   2.62  16.5     0     1     4     4
#>  2     6  2.62   110  3.9   2.88  17.0     0     1     4     4
#>  3     4  1.77    93  3.85  2.32  18.6     1     1     4     1
#>  4     6  4.23   110  3.08  3.22  19.4     1     0     3     1
#>  5     8  5.90   175  3.15  3.44  17.0     0     0     3     2
#>  6     6  3.69   105  2.76  3.46  20.2     1     0     3     1
#>  7     8  5.90   245  3.21  3.57  15.8     0     0     3     4
#>  8     4  2.40    62  3.69  3.19  20       1     0     4     2
#>  9     4  2.31    95  3.92  3.15  22.9     1     0     4     2
#> 10     6  2.75   123  3.92  3.44  18.3     1     0     4     4
#> # … with 22 more rows

# window functions are useful for grouped mutates
mtcars %>%
 group_by(cyl) %>%
 mutate(rank = min_rank(desc(mpg)))

#> # A tibble: 32 x 12
#> # Groups:   cyl [3]
#>      mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb  rank
#>    <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <int>
#>  1  21       6  160    110  3.9   2.62  16.5     0     1     4     4     2
#>  2  21       6  160    110  3.9   2.88  17.0     0     1     4     4     2
#>  3  22.8     4  108     93  3.85  2.32  18.6     1     1     4     1     8
#>  4  21.4     6  258    110  3.08  3.22  19.4     1     0     3     1     1
#>  5  18.7     8  360    175  3.15  3.44  17.0     0     0     3     2     2
#>  6  18.1     6  225    105  2.76  3.46  20.2     1     0     3     1     6
#>  7  14.3     8  360    245  3.21  3.57  15.8     0     0     3     4    11
#>  8  24.4     4  147.    62  3.69  3.19  20       1     0     4     2     7
#>  9  22.8     4  141.    95  3.92  3.15  22.9     1     0     4     2     8
#> 10  19.2     6  168.   123  3.92  3.44  18.3     1     0     4     4     5
#> # … with 22 more rows

# see `vignette("window-functions")` for more details

# mutate() vs transmute --------------------------
# mutate() keeps all existing variables
mtcars %>%
  mutate(displ_l = disp / 61.0237)

#>                      mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#> Mazda RX4           21.0   6 160.0 110 3.90 2.620 16.46  0  1    4    4
#> Mazda RX4 Wag       21.0   6 160.0 110 3.90 2.875 17.02  0  1    4    4
#> Datsun 710          22.8   4 108.0  93 3.85 2.320 18.61  1  1    4    1
#> Hornet 4 Drive      21.4   6 258.0 110 3.08 3.215 19.44  1  0    3    1
#> Hornet Sportabout   18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2
#> Valiant             18.1   6 225.0 105 2.76 3.460 20.22  1  0    3    1
#> Duster 360          14.3   8 360.0 245 3.21 3.570 15.84  0  0    3    4
#> Merc 240D           24.4   4 146.7  62 3.69 3.190 20.00  1  0    4    2
#> Merc 230            22.8   4 140.8  95 3.92 3.150 22.90  1  0    4    2
#> Merc 280            19.2   6 167.6 123 3.92 3.440 18.30  1  0    4    4
#> Merc 280C           17.8   6 167.6 123 3.92 3.440 18.90  1  0    4    4
#> Merc 450SE          16.4   8 275.8 180 3.07 4.070 17.40  0  0    3    3
#> Merc 450SL          17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3
#> Merc 450SLC         15.2   8 275.8 180 3.07 3.780 18.00  0  0    3    3
#> Cadillac Fleetwood  10.4   8 472.0 205 2.93 5.250 17.98  0  0    3    4
#> Lincoln Continental 10.4   8 460.0 215 3.00 5.424 17.82  0  0    3    4
#> Chrysler Imperial   14.7   8 440.0 230 3.23 5.345 17.42  0  0    3    4
#> Fiat 128            32.4   4  78.7  66 4.08 2.200 19.47  1  1    4    1
#> Honda Civic         30.4   4  75.7  52 4.93 1.615 18.52  1  1    4    2
#> Toyota Corolla      33.9   4  71.1  65 4.22 1.835 19.90  1  1    4    1
#> Toyota Corona       21.5   4 120.1  97 3.70 2.465 20.01  1  0    3    1
#> Dodge Challenger    15.5   8 318.0 150 2.76 3.520 16.87  0  0    3    2
#> AMC Javelin         15.2   8 304.0 150 3.15 3.435 17.30  0  0    3    2
#> Camaro Z28          13.3   8 350.0 245 3.73 3.840 15.41  0  0    3    4
#> Pontiac Firebird    19.2   8 400.0 175 3.08 3.845 17.05  0  0    3    2
#> Fiat X1-9           27.3   4  79.0  66 4.08 1.935 18.90  1  1    4    1
#> Porsche 914-2       26.0   4 120.3  91 4.43 2.140 16.70  0  1    5    2
#> Lotus Europa        30.4   4  95.1 113 3.77 1.513 16.90  1  1    5    2
#> Ford Pantera L      15.8   8 351.0 264 4.22 3.170 14.50  0  1    5    4
#> Ferrari Dino        19.7   6 145.0 175 3.62 2.770 15.50  0  1    5    6
#> Maserati Bora       15.0   8 301.0 335 3.54 3.570 14.60  0  1    5    8
#> Volvo 142E          21.4   4 121.0 109 4.11 2.780 18.60  1  1    4    2
#>                      displ_l
#> Mazda RX4           2.621932
#> Mazda RX4 Wag       2.621932
#> Datsun 710          1.769804
#> Hornet 4 Drive      4.227866
#> Hornet Sportabout   5.899347
#> Valiant             3.687092
#> Duster 360          5.899347
#> Merc 240D           2.403984
#> Merc 230            2.307300
#> Merc 280            2.746474
#> Merc 280C           2.746474
#> Merc 450SE          4.519556
#> Merc 450SL          4.519556
#> Merc 450SLC         4.519556
#> Cadillac Fleetwood  7.734700
#> Lincoln Continental 7.538055
#> Chrysler Imperial   7.210313
#> Fiat 128            1.289663
#> Honda Civic         1.240502
#> Toyota Corolla      1.165121
#> Toyota Corona       1.968088
#> Dodge Challenger    5.211090
#> AMC Javelin         4.981671
#> Camaro Z28          5.735477
#> Pontiac Firebird    6.554830
#> Fiat X1-9           1.294579
#> Porsche 914-2       1.971365
#> Lotus Europa        1.558411
#> Ford Pantera L      5.751864
#> Ferrari Dino        2.376126
#> Maserati Bora       4.932510
#> Volvo 142E          1.982836

# transmute keeps only the variables you create
mtcars %>%
  transmute(displ_l = disp / 61.0237)

#>                      displ_l
#> Mazda RX4           2.621932
#> Mazda RX4 Wag       2.621932
#> Datsun 710          1.769804
#> Hornet 4 Drive      4.227866
#> Hornet Sportabout   5.899347
#> Valiant             3.687092
#> Duster 360          5.899347
#> Merc 240D           2.403984
#> Merc 230            2.307300
#> Merc 280            2.746474
#> Merc 280C           2.746474
#> Merc 450SE          4.519556
#> Merc 450SL          4.519556
#> Merc 450SLC         4.519556
#> Cadillac Fleetwood  7.734700
#> Lincoln Continental 7.538055
#> Chrysler Imperial   7.210313
#> Fiat 128            1.289663
#> Honda Civic         1.240502
#> Toyota Corolla      1.165121
#> Toyota Corona       1.968088
#> Dodge Challenger    5.211090
#> AMC Javelin         4.981671
#> Camaro Z28          5.735477
#> Pontiac Firebird    6.554830
#> Fiat X1-9           1.294579
#> Porsche 914-2       1.971365
#> Lotus Europa        1.558411
#> Ford Pantera L      5.751864
#> Ferrari Dino        2.376126
#> Maserati Bora       4.932510
#> Volvo 142E          1.982836

# Grouping ----------------------------------------
# The mutate operation may yield different results on grouped
# tibbles because the expressions are computed within groups.
# The following normalises `mass` by the global average:
starwars %>%
  mutate(mass / mean(mass, na.rm = TRUE)) %>%
  pull()

#>  [1]  0.7912704  0.7707179  0.3288397  1.3975685  0.5035357  1.2331487
#>  [7]  0.7707179  0.3288397  0.8632041  0.7912704  0.8632041         NA
#> [13]  1.1509388  0.8220991  0.7604417 13.9551329  0.7912704  1.1303863
#> [19]  0.1746961  0.7707179  0.8036019  1.4386735  1.1612150  0.8118229
#> [25]  0.8118229  0.8529279         NA         NA  0.2055248  0.6987843
#> [31]  0.9145853  0.9248615         NA  0.6782318  0.8426516         NA
#> [37]         NA         NA  0.4110496         NA         NA  0.8220991
#> [43]         NA  0.5651932  0.4624308         NA  0.6679556  0.8632041
#> [49]  0.8426516  0.8940328         NA  0.5138120         NA         NA
#> [55]  0.8220991         NA  0.8734803         NA         NA  0.8220991
#> [61]  0.5775246  0.5138120         NA  0.8220991         NA  0.8118229
#> [67]  0.5651932  1.0481764  0.9043091         NA         NA  0.1541436
#> [73]         NA  0.4932595         NA  0.5857456  1.6339220  1.3975685
#> [79]  0.8118229  0.4932595  0.8220991         NA         NA         NA
#> [85]         NA         NA  0.4624308

# Whereas this normalises `mass` by the averages within gender
# levels:
starwars %>%
  group_by(gender) %>%
  mutate(mass / mean(mass, na.rm = TRUE)) %>%
  pull()

#>  [1]  0.7254095  0.7065677  0.3014689  1.2812428  0.8959772  1.1305083
#>  [7]  1.3713937  0.3014689  0.7913558  0.7254095  0.7913558         NA
#> [13]  1.0551411  0.7536722  0.6971468 12.7935861  0.7254095  1.0362993
#> [19]  0.1601553  0.7065677  0.7367146  1.3189264  1.0645620  0.7442513
#> [25]  0.7442513  0.7819349         NA         NA  0.1884181  0.6406214
#> [31]  0.8384604  0.8478813         NA  0.6217796  0.7725140         NA
#> [37]         NA         NA  0.3768361         NA         NA  0.7536722
#> [43]         NA  1.0056887  0.4239406         NA  0.6123587  0.7913558
#> [49]  0.7725140  0.8196185         NA  0.9142625         NA         NA
#> [55]  0.7536722         NA  0.8007767         NA         NA  0.7536722
#> [61]  1.0276310  0.9142625         NA  0.7536722         NA  0.7442513
#> [67]  1.0056887  0.9609321  0.8290395         NA         NA  0.1413135
#> [73]         NA  0.4522033         NA  1.0422592  1.4979236  1.2812428
#> [79]  0.7442513  1.0000000  0.7536722         NA         NA         NA
#> [85]         NA         NA  0.8228362

# Indirection ----------------------------------------
# Refer to column names stored as strings with the `.data` pronoun:
vars <- c("mass", "height")
mutate(starwars, prod = .data[[vars[[1]]]] * .data[[vars[[2]]]])

#> # A tibble: 87 x 15
#>    name  height  mass hair_color skin_color eye_color birth_year sex   gender
#>    <chr>  <int> <dbl> <chr>      <chr>      <chr>          <dbl> <chr> <chr> 
#>  1 Luke…    172    77 blond      fair       blue            19   male  mascu…
#>  2 C-3PO    167    75 <NA>       gold       yellow         112   none  mascu…
#>  3 R2-D2     96    32 <NA>       white, bl… red             33   none  mascu…
#>  4 Dart…    202   136 none       white      yellow          41.9 male  mascu…
#>  5 Leia…    150    49 brown      light      brown           19   fema… femin…
#>  6 Owen…    178   120 brown, gr… light      blue            52   male  mascu…
#>  7 Beru…    165    75 brown      light      blue            47   fema… femin…
#>  8 R5-D4     97    32 <NA>       white, red red             NA   none  mascu…
#>  9 Bigg…    183    84 black      light      brown           24   male  mascu…
#> 10 Obi-…    182    77 auburn, w… fair       blue-gray       57   male  mascu…
#> # … with 77 more rows, and 6 more variables: homeworld <chr>, species <chr>,
#> #   films <list>, vehicles <list>, starships <list>, prod <dbl>

# Learn more in ?dplyr_tidy_eval