n_distinct {dplyr}	R Documentation

Efficiently count the number of unique values in a set of vectors

Description

This is a faster and more concise equivalent of length(unique(x))

Usage

n_distinct(..., na.rm = FALSE)

Arguments

...
vectors of values

na.rm
if TRUE missing values don't count

Examples

x <- sample(1:10, 1e5, rep = TRUE)
length(unique(x))
n_distinct(x)

[Package dplyr version 1.0.7 Index]