NASA spatio-temporal data

This data comes from the ASA 2007 data expo, http://stat-computing.org/dataexpo/2006/. The data are geographic and atmospheric measures on a very coarse 24 by 24 grid covering Central America. The variables are: temperature (surface and air), ozone, air pressure, and cloud cover (low, mid, and high). All variables are monthly averages, with observations for Jan 1995 to Dec 2000. These data were obtained from the NASA Langley Research Center Atmospheric Sciences Data Center (with permission; see important copyright terms below).

nasa

Format

A tbl_cube with 41,472 observations.

Dimensions

lat, long: latitude and longitude

year, month: month and year

Measures

cloudlow, cloudmed, cloudhigh: cloud cover at three heights

ozone

surftemp and temperature

pressure

Examples

nasa

#> Source: local array [41,472 x 4]
#> D: lat [dbl, 24]
#> D: long [dbl, 24]
#> D: month [int, 12]
#> D: year [int, 6]
#> M: cloudhigh [dbl[,24,12,6]]
#> M: cloudlow [dbl[,24,12,6]]
#> M: cloudmid [dbl[,24,12,6]]
#> M: ozone [dbl[,24,12,6]]
#> M: pressure [dbl[,24,12,6]]
#> M: surftemp [dbl[,24,12,6]]
#> M: temperature [dbl[,24,12,6]]