Nest by one or more variables

nest_by() is closely related to group_by(). However, instead of storing the group structure in the metadata, it makes it explicit in the data, giving each group key a single row with a list-column of data frames that contain all the other data.

nest_by(.data, ..., .key = "data", .keep = FALSE, .add = FALSE)

Arguments

.data
A data frame, data frame extension (e.g. a tibble), or a lazy data frame (e.g. from dbplyr or dtplyr). See Methods, below, for more details.

...
In group_by(), variables or computations to group by. In ungroup(), variables to remove from the grouping.

.key
Name of the list column

.keep
Should the grouping columns be kept in the list column.

.add
When FALSE, the default, group_by() will override existing groups. To add to the existing groups, use .add = TRUE.

This argument was previously called add, but that prevented creating a new grouping variable called add, and conflicts with our naming conventions.

Value

A rowwise() data frame. The output has the following properties:

The rows come from the underlying group_keys().

The columns are the grouping keys plus one list-column of data frames.

Data frame attributes are not preserved, because nest_by() fundamentally creates a new data frame.

A tbl with one row per unique combination of the grouping variables. The first columns are the grouping variables, followed by a list column of tibbles with matching rows of the remaining columns.

Methods

This function is a generic, which means that packages can provide implementations (methods) for other classes. See the documentation of individual methods for extra arguments and differences in behaviour.

The following methods are currently available in loaded packages: dplyr (data.frame) .

Examples

# After nesting, you get one row per group
iris %>% nest_by(Species)

#> # A tibble: 3 x 2
#> # Rowwise:  Species
#>   Species              data
#>   <fct>      <list<df[,4]>>
#> 1 setosa           [50 × 4]
#> 2 versicolor       [50 × 4]
#> 3 virginica        [50 × 4]

starwars %>% nest_by(species)

#> # A tibble: 38 x 2
#> # Rowwise:  species
#>    species              data
#>    <chr>     <list<df[,13]>>
#>  1 Aleena           [1 × 13]
#>  2 Besalisk         [1 × 13]
#>  3 Cerean           [1 × 13]
#>  4 Chagrian         [1 × 13]
#>  5 Clawdite         [1 × 13]
#>  6 Droid            [6 × 13]
#>  7 Dug              [1 × 13]
#>  8 Ewok             [1 × 13]
#>  9 Geonosian        [1 × 13]
#> 10 Gungan           [3 × 13]
#> # … with 28 more rows

# The output is grouped by row, which makes modelling particularly easy
models <- mtcars %>%
  nest_by(cyl) %>%
  mutate(model = list(lm(mpg ~ wt, data = data)))
models

#> # A tibble: 3 x 3
#> # Rowwise:  cyl
#>     cyl            data model 
#>   <dbl> <list<df[,10]>> <list>
#> 1     4       [11 × 10] <lm>  
#> 2     6        [7 × 10] <lm>  
#> 3     8       [14 × 10] <lm>  

models %>% summarise(rsq = summary(model)$r.squared)

#> # A tibble: 3 x 2
#> # Rowwise:  cyl
#>     cyl   rsq
#>   <dbl> <dbl>
#> 1     4 0.509
#> 2     6 0.465
#> 3     8 0.423