Nest join

nest_join() returns all rows and columns in x with a new nested-df column that contains all matches from y. When there is no match, the list column is a 0-row tibble.

nest_join(x, y, by = NULL, copy = FALSE, keep = FALSE, name = NULL, ...)

# S3 method for data.frame
nest_join(x, y, by = NULL, copy = FALSE, keep = FALSE, name = NULL, ...)

Arguments

x, y
A pair of data frames, data frame extensions (e.g. a tibble), or lazy data frames (e.g. from dbplyr or dtplyr). See Methods, below, for more details.

by
A character vector of variables to join by.

If NULL, the default, *_join() will perform a natural join, using all variables in common across x and y. A message lists the variables so that you can check they're correct; suppress the message by supplying by explicitly.

To join by different variables on x and y, use a named vector. For example, by = c("a" = "b") will match x$a to y$b.

To join by multiple variables, use a vector with length > 1. For example, by = c("a", "b") will match x$a to y$a and x$b to y$b. Use a named vector to match different variables in x and y. For example, by = c("a" = "b", "c" = "d") will match x$a to y$b and x$c to y$d.

To perform a cross-join, generating all combinations of x and y, use by = character().

copy
If x and y are not from the same data source, and copy is TRUE, then y will be copied into the same src as x. This allows you to join tables across srcs, but it is a potentially expensive operation so you must opt into it.

keep
Should the join keys from both x and y be preserved in the output? Only applies to nest_join(), left_join(), right_join(), and full_join().

name
The name of the list column nesting joins create. If NULL the name of y is used.

...
Other parameters passed onto methods.

Details

In some sense, a nest_join() is the most fundamental join since you can recreate the other joins from it:

inner_join() is a nest_join() plus tidyr::unnest()

left_join() nest_join() plus unnest(.drop = FALSE).

semi_join() is a nest_join() plus a filter() where you check that every element of data has at least one row,

anti_join() is a nest_join() plus a filter() where you check every element has zero rows.

Methods

This function is a generic, which means that packages can provide implementations (methods) for other classes. See the documentation of individual methods for extra arguments and differences in behaviour.

The following methods are currently available in loaded packages: dplyr (data.frame) .

See also

Examples

band_members %>% nest_join(band_instruments)

#> Joining, by = "name"

#> # A tibble: 3 x 3
#>   name  band    band_instruments
#> * <chr> <chr>   <list>          
#> 1 Mick  Stones  <tibble [0 × 1]>
#> 2 John  Beatles <tibble [1 × 1]>
#> 3 Paul  Beatles <tibble [1 × 1]>