progress_estimated {dplyr}	R Documentation

Progress bar with estimated time.

Description

This progress bar has been deprecated since providing progress bars is not the responsibility of dplyr. Instead, you might try the more powerful progress package.

This reference class represents a text progress bar displayed estimated time remaining. When finished, it displays the total duration. The automatic progress bar can be disabled by setting option dplyr.show_progress to FALSE.

Usage

progress_estimated(n, min_time = 0)

Arguments

n
Total number of items

min_time
Progress bar will wait until at least min_time seconds have elapsed before displaying any results.

Value

A ref class with methods tick(), print(), pause(), and stop().

Examples

p <- progress_estimated(3)
p$tick()
p$tick()
p$tick()

p <- progress_estimated(3)
for (i in 1:3) p$pause(0.1)$tick()$print()

p <- progress_estimated(3)
p$tick()$print()$
 pause(1)$stop()

# If min_time is set, progress bar not shown until that many
# seconds have elapsed
p <- progress_estimated(3, min_time = 3)
for (i in 1:3) p$pause(0.1)$tick()$print()

## Not run: 
p <- progress_estimated(10, min_time = 3)
for (i in 1:10) p$pause(0.5)$tick()$print()

## End(Not run)

[Package dplyr version 1.0.7 Index]