Windowed rank functions.

Six variations on ranking functions, mimicking the ranking functions described in SQL2003. They are currently implemented using the built in rank function, and are provided mainly as a convenience when converting between R and SQL. All ranking functions map smallest inputs to smallest outputs. Use desc() to reverse the direction.

row_number(x)

ntile(x = row_number(), n)

min_rank(x)

dense_rank(x)

percent_rank(x)

cume_dist(x)

Arguments

x
a vector of values to rank. Missing values are left as is. If you want to treat them as the smallest or largest values, replace with Inf or -Inf before ranking.

n
number of groups to split up into.

Details

row_number(): equivalent to rank(ties.method = "first")

min_rank(): equivalent to rank(ties.method = "min")

dense_rank(): like min_rank(), but with no gaps between ranks

percent_rank(): a number between 0 and 1 computed by rescaling min_rank to [0, 1]

cume_dist(): a cumulative distribution function. Proportion of all values less than or equal to the current rank.

ntile(): a rough rank, which breaks the input vector into n buckets.

Examples

x <- c(5, 1, 3, 2, 2, NA)
row_number(x)

#> [1]  5  1  4  2  3 NA

min_rank(x)

#> [1]  5  1  4  2  2 NA

dense_rank(x)

#> [1]  4  1  3  2  2 NA

percent_rank(x)

#> [1] 1.00 0.00 0.75 0.25 0.25   NA

cume_dist(x)

#> [1] 1.0 0.2 0.8 0.6 0.6  NA

ntile(x, 2)

#> [1]  2  1  2  1  1 NA

ntile(runif(100), 10)

#>   [1] 10  7 10  7  7  9  8  9  1 10  1  4  6  3  6 10  6 10  9  2  8  2  7  2  1
#>  [26] 10  2  8  2  3  1  6  3  6  4  8  2  6  4  6  8  4  4  3  5  5  9  3  4  9
#>  [51]  1  5  5  8 10  7  7  1  9  2  5  9  7  1  3  1 10  6  8 10  5  5  7  3  8
#>  [76]  5  1 10  7  9  4  9  1  9  6  4  6  2  8  5  2  8  5  4  4  2  3  3  7  3

# row_number can be used with single table verbs without specifying x
# (for data frames and databases that support windowing)
mutate(mtcars, row_number() == 1L)

#>     mpg cyl  disp  hp drat    wt  qsec vs am gear carb row_number() == 1L
#> 1  21.0   6 160.0 110 3.90 2.620 16.46  0  1    4    4               TRUE
#> 2  21.0   6 160.0 110 3.90 2.875 17.02  0  1    4    4              FALSE
#> 3  22.8   4 108.0  93 3.85 2.320 18.61  1  1    4    1              FALSE
#> 4  21.4   6 258.0 110 3.08 3.215 19.44  1  0    3    1              FALSE
#> 5  18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2              FALSE
#> 6  18.1   6 225.0 105 2.76 3.460 20.22  1  0    3    1              FALSE
#> 7  14.3   8 360.0 245 3.21 3.570 15.84  0  0    3    4              FALSE
#> 8  24.4   4 146.7  62 3.69 3.190 20.00  1  0    4    2              FALSE
#> 9  22.8   4 140.8  95 3.92 3.150 22.90  1  0    4    2              FALSE
#> 10 19.2   6 167.6 123 3.92 3.440 18.30  1  0    4    4              FALSE
#> 11 17.8   6 167.6 123 3.92 3.440 18.90  1  0    4    4              FALSE
#> 12 16.4   8 275.8 180 3.07 4.070 17.40  0  0    3    3              FALSE
#> 13 17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3              FALSE
#> 14 15.2   8 275.8 180 3.07 3.780 18.00  0  0    3    3              FALSE
#> 15 10.4   8 472.0 205 2.93 5.250 17.98  0  0    3    4              FALSE
#> 16 10.4   8 460.0 215 3.00 5.424 17.82  0  0    3    4              FALSE
#> 17 14.7   8 440.0 230 3.23 5.345 17.42  0  0    3    4              FALSE
#> 18 32.4   4  78.7  66 4.08 2.200 19.47  1  1    4    1              FALSE
#> 19 30.4   4  75.7  52 4.93 1.615 18.52  1  1    4    2              FALSE
#> 20 33.9   4  71.1  65 4.22 1.835 19.90  1  1    4    1              FALSE
#> 21 21.5   4 120.1  97 3.70 2.465 20.01  1  0    3    1              FALSE
#> 22 15.5   8 318.0 150 2.76 3.520 16.87  0  0    3    2              FALSE
#> 23 15.2   8 304.0 150 3.15 3.435 17.30  0  0    3    2              FALSE
#> 24 13.3   8 350.0 245 3.73 3.840 15.41  0  0    3    4              FALSE
#> 25 19.2   8 400.0 175 3.08 3.845 17.05  0  0    3    2              FALSE
#> 26 27.3   4  79.0  66 4.08 1.935 18.90  1  1    4    1              FALSE
#> 27 26.0   4 120.3  91 4.43 2.140 16.70  0  1    5    2              FALSE
#> 28 30.4   4  95.1 113 3.77 1.513 16.90  1  1    5    2              FALSE
#> 29 15.8   8 351.0 264 4.22 3.170 14.50  0  1    5    4              FALSE
#> 30 19.7   6 145.0 175 3.62 2.770 15.50  0  1    5    6              FALSE
#> 31 15.0   8 301.0 335 3.54 3.570 14.60  0  1    5    8              FALSE
#> 32 21.4   4 121.0 109 4.11 2.780 18.60  1  1    4    2              FALSE

mtcars %>% filter(between(row_number(), 1, 10))

#>     mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#> 1  21.0   6 160.0 110 3.90 2.620 16.46  0  1    4    4
#> 2  21.0   6 160.0 110 3.90 2.875 17.02  0  1    4    4
#> 3  22.8   4 108.0  93 3.85 2.320 18.61  1  1    4    1
#> 4  21.4   6 258.0 110 3.08 3.215 19.44  1  0    3    1
#> 5  18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2
#> 6  18.1   6 225.0 105 2.76 3.460 20.22  1  0    3    1
#> 7  14.3   8 360.0 245 3.21 3.570 15.84  0  0    3    4
#> 8  24.4   4 146.7  62 3.69 3.190 20.00  1  0    4    2
#> 9  22.8   4 140.8  95 3.92 3.150 22.90  1  0    4    2
#> 10 19.2   6 167.6 123 3.92 3.440 18.30  1  0    4    4