Group input by rows

rowwise() is used for the results of do() when you create list-variables. It is also useful to support arbitrary complex operations that need to be applied to each row.

rowwise(data)

Arguments

data
Input data frame.

Details

Currently, rowwise grouping only works with data frames. Its main impact is to allow you to work with list-variables in summarise() and mutate() without having to use [[1]]. This makes summarise() on a rowwise tbl effectively equivalent to plyr::ldply().

Examples

df <- expand.grid(x = 1:3, y = 3:1)
df_done <- df %>% rowwise() %>% do(i = seq(.$x, .$y))
df_done

#> Source: local data frame [9 x 1]
#> Groups: <by row>
#> 
#> # A tibble: 9 x 1
#>   i        
#> * <list>   
#> 1 <int [3]>
#> 2 <int [2]>
#> 3 <int [1]>
#> 4 <int [2]>
#> 5 <int [1]>
#> 6 <int [2]>
#> 7 <int [1]>
#> 8 <int [2]>
#> 9 <int [3]>

df_done %>% summarise(n = length(i))

#> # A tibble: 9 x 1
#>       n
#>   <int>
#> 1     3
#> 2     2
#> 3     1
#> 4     2
#> 5     1
#> 6     2
#> 7     1
#> 8     2
#> 9     3