Sample n rows from a table

This is a wrapper around sample.int() to make it easy to select random rows from a table. It currently only works for local tbls.

sample_n(tbl, size, replace = FALSE, weight = NULL, .env = NULL)

sample_frac(tbl, size = 1, replace = FALSE, weight = NULL,
  .env = NULL)

Arguments

tbl
tbl of data.

size
For sample_n(), the number of rows to select. For sample_frac(), the fraction of rows to select. If tbl is grouped, size applies to each group.

replace
Sample with or without replacement?

weight
Sampling weights. This must evaluate to a vector of non-negative numbers the same length as the input. Weights are automatically standardised to sum to 1.

This argument is automatically quoted and later evaluated in the context of the data frame. It supports unquoting. See vignette("programming") for an introduction to these concepts.

.env
This variable is deprecated and no longer has any effect. To evaluate weight in a particular context, you can now unquote a quosure.

Examples

by_cyl <- mtcars %>% group_by(cyl)

# Sample fixed number per group
sample_n(mtcars, 10)

#>     mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#> 1  10.4   8 472.0 205 2.93 5.250 17.98  0  0    3    4
#> 2  21.0   6 160.0 110 3.90 2.620 16.46  0  1    4    4
#> 3  26.0   4 120.3  91 4.43 2.140 16.70  0  1    5    2
#> 4  13.3   8 350.0 245 3.73 3.840 15.41  0  0    3    4
#> 5  17.8   6 167.6 123 3.92 3.440 18.90  1  0    4    4
#> 6  15.0   8 301.0 335 3.54 3.570 14.60  0  1    5    8
#> 7  15.5   8 318.0 150 2.76 3.520 16.87  0  0    3    2
#> 8  14.7   8 440.0 230 3.23 5.345 17.42  0  0    3    4
#> 9  19.2   8 400.0 175 3.08 3.845 17.05  0  0    3    2
#> 10 22.8   4 140.8  95 3.92 3.150 22.90  1  0    4    2

sample_n(mtcars, 50, replace = TRUE)

#>     mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#> 1  18.1   6 225.0 105 2.76 3.460 20.22  1  0    3    1
#> 2  30.4   4  95.1 113 3.77 1.513 16.90  1  1    5    2
#> 3  14.7   8 440.0 230 3.23 5.345 17.42  0  0    3    4
#> 4  17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3
#> 5  17.8   6 167.6 123 3.92 3.440 18.90  1  0    4    4
#> 6  21.5   4 120.1  97 3.70 2.465 20.01  1  0    3    1
#> 7  18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2
#> 8  15.2   8 275.8 180 3.07 3.780 18.00  0  0    3    3
#> 9  22.8   4 140.8  95 3.92 3.150 22.90  1  0    4    2
#> 10 30.4   4  75.7  52 4.93 1.615 18.52  1  1    4    2
#> 11 21.4   6 258.0 110 3.08 3.215 19.44  1  0    3    1
#> 12 17.8   6 167.6 123 3.92 3.440 18.90  1  0    4    4
#> 13 15.2   8 304.0 150 3.15 3.435 17.30  0  0    3    2
#> 14 17.8   6 167.6 123 3.92 3.440 18.90  1  0    4    4
#> 15 14.7   8 440.0 230 3.23 5.345 17.42  0  0    3    4
#> 16 26.0   4 120.3  91 4.43 2.140 16.70  0  1    5    2
#> 17 10.4   8 460.0 215 3.00 5.424 17.82  0  0    3    4
#> 18 16.4   8 275.8 180 3.07 4.070 17.40  0  0    3    3
#> 19 22.8   4 140.8  95 3.92 3.150 22.90  1  0    4    2
#> 20 10.4   8 472.0 205 2.93 5.250 17.98  0  0    3    4
#> 21 10.4   8 460.0 215 3.00 5.424 17.82  0  0    3    4
#> 22 15.8   8 351.0 264 4.22 3.170 14.50  0  1    5    4
#> 23 30.4   4  95.1 113 3.77 1.513 16.90  1  1    5    2
#> 24 24.4   4 146.7  62 3.69 3.190 20.00  1  0    4    2
#> 25 30.4   4  95.1 113 3.77 1.513 16.90  1  1    5    2
#> 26 21.5   4 120.1  97 3.70 2.465 20.01  1  0    3    1
#> 27 17.8   6 167.6 123 3.92 3.440 18.90  1  0    4    4
#> 28 21.5   4 120.1  97 3.70 2.465 20.01  1  0    3    1
#> 29 15.5   8 318.0 150 2.76 3.520 16.87  0  0    3    2
#> 30 17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3
#> 31 30.4   4  95.1 113 3.77 1.513 16.90  1  1    5    2
#> 32 32.4   4  78.7  66 4.08 2.200 19.47  1  1    4    1
#> 33 13.3   8 350.0 245 3.73 3.840 15.41  0  0    3    4
#> 34 21.4   4 121.0 109 4.11 2.780 18.60  1  1    4    2
#> 35 15.2   8 275.8 180 3.07 3.780 18.00  0  0    3    3
#> 36 21.4   4 121.0 109 4.11 2.780 18.60  1  1    4    2
#> 37 17.8   6 167.6 123 3.92 3.440 18.90  1  0    4    4
#> 38 15.8   8 351.0 264 4.22 3.170 14.50  0  1    5    4
#> 39 14.3   8 360.0 245 3.21 3.570 15.84  0  0    3    4
#> 40 22.8   4 108.0  93 3.85 2.320 18.61  1  1    4    1
#> 41 21.5   4 120.1  97 3.70 2.465 20.01  1  0    3    1
#> 42 15.2   8 275.8 180 3.07 3.780 18.00  0  0    3    3
#> 43 10.4   8 472.0 205 2.93 5.250 17.98  0  0    3    4
#> 44 18.1   6 225.0 105 2.76 3.460 20.22  1  0    3    1
#> 45 27.3   4  79.0  66 4.08 1.935 18.90  1  1    4    1
#> 46 27.3   4  79.0  66 4.08 1.935 18.90  1  1    4    1
#> 47 14.3   8 360.0 245 3.21 3.570 15.84  0  0    3    4
#> 48 21.4   6 258.0 110 3.08 3.215 19.44  1  0    3    1
#> 49 10.4   8 460.0 215 3.00 5.424 17.82  0  0    3    4
#> 50 24.4   4 146.7  62 3.69 3.190 20.00  1  0    4    2

sample_n(mtcars, 10, weight = mpg)

#>     mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#> 1  30.4   4  95.1 113 3.77 1.513 16.90  1  1    5    2
#> 2  19.2   8 400.0 175 3.08 3.845 17.05  0  0    3    2
#> 3  21.0   6 160.0 110 3.90 2.875 17.02  0  1    4    4
#> 4  30.4   4  75.7  52 4.93 1.615 18.52  1  1    4    2
#> 5  32.4   4  78.7  66 4.08 2.200 19.47  1  1    4    1
#> 6  17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3
#> 7  15.0   8 301.0 335 3.54 3.570 14.60  0  1    5    8
#> 8  26.0   4 120.3  91 4.43 2.140 16.70  0  1    5    2
#> 9  19.2   6 167.6 123 3.92 3.440 18.30  1  0    4    4
#> 10 19.7   6 145.0 175 3.62 2.770 15.50  0  1    5    6

sample_n(by_cyl, 3)

#> # A tibble: 9 x 11
#> # Groups:   cyl [3]
#>     mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#>   <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1  22.8     4 141.     95  3.92  3.15  22.9     1     0     4     2
#> 2  22.8     4 108      93  3.85  2.32  18.6     1     1     4     1
#> 3  30.4     4  95.1   113  3.77  1.51  16.9     1     1     5     2
#> 4  17.8     6 168.    123  3.92  3.44  18.9     1     0     4     4
#> 5  19.2     6 168.    123  3.92  3.44  18.3     1     0     4     4
#> 6  21       6 160     110  3.9   2.62  16.5     0     1     4     4
#> 7  10.4     8 472     205  2.93  5.25  18.0     0     0     3     4
#> 8  18.7     8 360     175  3.15  3.44  17.0     0     0     3     2
#> 9  13.3     8 350     245  3.73  3.84  15.4     0     0     3     4

sample_n(by_cyl, 10, replace = TRUE)

#> # A tibble: 30 x 11
#> # Groups:   cyl [3]
#>      mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#>    <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1  22.8     4 141.     95  3.92  3.15  22.9     1     0     4     2
#>  2  27.3     4  79      66  4.08  1.94  18.9     1     1     4     1
#>  3  30.4     4  75.7    52  4.93  1.62  18.5     1     1     4     2
#>  4  27.3     4  79      66  4.08  1.94  18.9     1     1     4     1
#>  5  32.4     4  78.7    66  4.08  2.2   19.5     1     1     4     1
#>  6  30.4     4  95.1   113  3.77  1.51  16.9     1     1     5     2
#>  7  22.8     4 141.     95  3.92  3.15  22.9     1     0     4     2
#>  8  22.8     4 141.     95  3.92  3.15  22.9     1     0     4     2
#>  9  32.4     4  78.7    66  4.08  2.2   19.5     1     1     4     1
#> 10  32.4     4  78.7    66  4.08  2.2   19.5     1     1     4     1
#> # ... with 20 more rows

sample_n(by_cyl, 3, weight = mpg / mean(mpg))

#> # A tibble: 9 x 11
#> # Groups:   cyl [3]
#>     mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#>   <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1  22.8     4  141.    95  3.92  3.15  22.9     1     0     4     2
#> 2  27.3     4   79     66  4.08  1.94  18.9     1     1     4     1
#> 3  21.5     4  120.    97  3.7   2.46  20.0     1     0     3     1
#> 4  21       6  160    110  3.9   2.62  16.5     0     1     4     4
#> 5  17.8     6  168.   123  3.92  3.44  18.9     1     0     4     4
#> 6  21.4     6  258    110  3.08  3.22  19.4     1     0     3     1
#> 7  15.8     8  351    264  4.22  3.17  14.5     0     1     5     4
#> 8  14.7     8  440    230  3.23  5.34  17.4     0     0     3     4
#> 9  16.4     8  276.   180  3.07  4.07  17.4     0     0     3     3

# Sample fixed fraction per group
# Default is to sample all data = randomly resample rows
sample_frac(mtcars)

#>     mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#> 1  19.2   6 167.6 123 3.92 3.440 18.30  1  0    4    4
#> 2  32.4   4  78.7  66 4.08 2.200 19.47  1  1    4    1
#> 3  22.8   4 140.8  95 3.92 3.150 22.90  1  0    4    2
#> 4  15.5   8 318.0 150 2.76 3.520 16.87  0  0    3    2
#> 5  26.0   4 120.3  91 4.43 2.140 16.70  0  1    5    2
#> 6  27.3   4  79.0  66 4.08 1.935 18.90  1  1    4    1
#> 7  13.3   8 350.0 245 3.73 3.840 15.41  0  0    3    4
#> 8  14.7   8 440.0 230 3.23 5.345 17.42  0  0    3    4
#> 9  21.4   4 121.0 109 4.11 2.780 18.60  1  1    4    2
#> 10 14.3   8 360.0 245 3.21 3.570 15.84  0  0    3    4
#> 11 21.5   4 120.1  97 3.70 2.465 20.01  1  0    3    1
#> 12 33.9   4  71.1  65 4.22 1.835 19.90  1  1    4    1
#> 13 21.0   6 160.0 110 3.90 2.620 16.46  0  1    4    4
#> 14 15.2   8 275.8 180 3.07 3.780 18.00  0  0    3    3
#> 15 17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3
#> 16 24.4   4 146.7  62 3.69 3.190 20.00  1  0    4    2
#> 17 30.4   4  95.1 113 3.77 1.513 16.90  1  1    5    2
#> 18 21.4   6 258.0 110 3.08 3.215 19.44  1  0    3    1
#> 19 15.8   8 351.0 264 4.22 3.170 14.50  0  1    5    4
#> 20 30.4   4  75.7  52 4.93 1.615 18.52  1  1    4    2
#> 21 18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2
#> 22 15.2   8 304.0 150 3.15 3.435 17.30  0  0    3    2
#> 23 21.0   6 160.0 110 3.90 2.875 17.02  0  1    4    4
#> 24 15.0   8 301.0 335 3.54 3.570 14.60  0  1    5    8
#> 25 18.1   6 225.0 105 2.76 3.460 20.22  1  0    3    1
#> 26 19.7   6 145.0 175 3.62 2.770 15.50  0  1    5    6
#> 27 22.8   4 108.0  93 3.85 2.320 18.61  1  1    4    1
#> 28 16.4   8 275.8 180 3.07 4.070 17.40  0  0    3    3
#> 29 17.8   6 167.6 123 3.92 3.440 18.90  1  0    4    4
#> 30 19.2   8 400.0 175 3.08 3.845 17.05  0  0    3    2
#> 31 10.4   8 472.0 205 2.93 5.250 17.98  0  0    3    4
#> 32 10.4   8 460.0 215 3.00 5.424 17.82  0  0    3    4

sample_frac(mtcars, 0.1)

#>    mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#> 1 15.8   8 351.0 264 4.22 3.170 14.50  0  1    5    4
#> 2 30.4   4  75.7  52 4.93 1.615 18.52  1  1    4    2
#> 3 19.2   8 400.0 175 3.08 3.845 17.05  0  0    3    2

sample_frac(mtcars, 1.5, replace = TRUE)

#>     mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#> 1  19.2   8 400.0 175 3.08 3.845 17.05  0  0    3    2
#> 2  15.2   8 304.0 150 3.15 3.435 17.30  0  0    3    2
#> 3  15.8   8 351.0 264 4.22 3.170 14.50  0  1    5    4
#> 4  17.8   6 167.6 123 3.92 3.440 18.90  1  0    4    4
#> 5  26.0   4 120.3  91 4.43 2.140 16.70  0  1    5    2
#> 6  19.7   6 145.0 175 3.62 2.770 15.50  0  1    5    6
#> 7  32.4   4  78.7  66 4.08 2.200 19.47  1  1    4    1
#> 8  17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3
#> 9  18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2
#> 10 21.0   6 160.0 110 3.90 2.620 16.46  0  1    4    4
#> 11 21.5   4 120.1  97 3.70 2.465 20.01  1  0    3    1
#> 12 15.8   8 351.0 264 4.22 3.170 14.50  0  1    5    4
#> 13 24.4   4 146.7  62 3.69 3.190 20.00  1  0    4    2
#> 14 14.3   8 360.0 245 3.21 3.570 15.84  0  0    3    4
#> 15 21.5   4 120.1  97 3.70 2.465 20.01  1  0    3    1
#> 16 17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3
#> 17 30.4   4  75.7  52 4.93 1.615 18.52  1  1    4    2
#> 18 18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2
#> 19 15.2   8 275.8 180 3.07 3.780 18.00  0  0    3    3
#> 20 32.4   4  78.7  66 4.08 2.200 19.47  1  1    4    1
#> 21 24.4   4 146.7  62 3.69 3.190 20.00  1  0    4    2
#> 22 10.4   8 460.0 215 3.00 5.424 17.82  0  0    3    4
#> 23 22.8   4 108.0  93 3.85 2.320 18.61  1  1    4    1
#> 24 19.2   6 167.6 123 3.92 3.440 18.30  1  0    4    4
#> 25 14.7   8 440.0 230 3.23 5.345 17.42  0  0    3    4
#> 26 22.8   4 108.0  93 3.85 2.320 18.61  1  1    4    1
#> 27 21.0   6 160.0 110 3.90 2.875 17.02  0  1    4    4
#> 28 17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3
#> 29 18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2
#> 30 18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2
#> 31 17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3
#> 32 30.4   4  95.1 113 3.77 1.513 16.90  1  1    5    2
#> 33 21.4   6 258.0 110 3.08 3.215 19.44  1  0    3    1
#> 34 21.4   6 258.0 110 3.08 3.215 19.44  1  0    3    1
#> 35 18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2
#> 36 26.0   4 120.3  91 4.43 2.140 16.70  0  1    5    2
#> 37 18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2
#> 38 22.8   4 140.8  95 3.92 3.150 22.90  1  0    4    2
#> 39 15.2   8 304.0 150 3.15 3.435 17.30  0  0    3    2
#> 40 19.7   6 145.0 175 3.62 2.770 15.50  0  1    5    6
#> 41 10.4   8 472.0 205 2.93 5.250 17.98  0  0    3    4
#> 42 30.4   4  95.1 113 3.77 1.513 16.90  1  1    5    2
#> 43 15.2   8 275.8 180 3.07 3.780 18.00  0  0    3    3
#> 44 32.4   4  78.7  66 4.08 2.200 19.47  1  1    4    1
#> 45 15.0   8 301.0 335 3.54 3.570 14.60  0  1    5    8
#> 46 15.5   8 318.0 150 2.76 3.520 16.87  0  0    3    2
#> 47 21.0   6 160.0 110 3.90 2.875 17.02  0  1    4    4
#> 48 24.4   4 146.7  62 3.69 3.190 20.00  1  0    4    2

sample_frac(mtcars, 0.1, weight = 1 / mpg)

#>    mpg cyl  disp  hp drat   wt qsec vs am gear carb
#> 1 21.4   4 121.0 109 4.11 2.78 18.6  1  1    4    2
#> 2 24.4   4 146.7  62 3.69 3.19 20.0  1  0    4    2
#> 3 16.4   8 275.8 180 3.07 4.07 17.4  0  0    3    3

sample_frac(by_cyl, 0.2)

#> # A tibble: 6 x 11
#> # Groups:   cyl [3]
#>     mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#>   <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1  21.5     4  120.    97  3.7   2.46  20.0     1     0     3     1
#> 2  26       4  120.    91  4.43  2.14  16.7     0     1     5     2
#> 3  21.4     6  258    110  3.08  3.22  19.4     1     0     3     1
#> 4  10.4     8  460    215  3     5.42  17.8     0     0     3     4
#> 5  10.4     8  472    205  2.93  5.25  18.0     0     0     3     4
#> 6  16.4     8  276.   180  3.07  4.07  17.4     0     0     3     3

sample_frac(by_cyl, 1, replace = TRUE)

#> # A tibble: 32 x 11
#> # Groups:   cyl [3]
#>      mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#>    <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1  26       4 120.     91  4.43  2.14  16.7     0     1     5     2
#>  2  27.3     4  79      66  4.08  1.94  18.9     1     1     4     1
#>  3  30.4     4  95.1   113  3.77  1.51  16.9     1     1     5     2
#>  4  33.9     4  71.1    65  4.22  1.84  19.9     1     1     4     1
#>  5  30.4     4  75.7    52  4.93  1.62  18.5     1     1     4     2
#>  6  21.4     4 121     109  4.11  2.78  18.6     1     1     4     2
#>  7  30.4     4  75.7    52  4.93  1.62  18.5     1     1     4     2
#>  8  33.9     4  71.1    65  4.22  1.84  19.9     1     1     4     1
#>  9  21.5     4 120.     97  3.7   2.46  20.0     1     0     3     1
#> 10  30.4     4  95.1   113  3.77  1.51  16.9     1     1     5     2
#> # ... with 22 more rows