Sample n rows from a table

sample_n() and sample_frac() have been superseded in favour of slice_sample(). While they will not be deprecated in the near future, retirement means that we will only perform critical bug fixes, so we recommend moving to the newer alternative.

These functions were superseded because we realised it was more convenient to have two mutually exclusive arguments to one function, rather than two separate functions. This also made it to clean up a few other smaller design issues with sample_n()/sample_frac:

The connection to slice() was not obvious.

The name of the first argument, tbl, is inconsistent with other single table verbs which use .data.

The size argument uses tidy evaluation, which is surprising and undocumented.

It was easier to remove the deprecated .env argument.

... was in a suboptimal position.

Usage

sample_n(tbl, size, replace = FALSE, weight = NULL, .env = NULL, ...)

sample_frac(tbl, size = 1, replace = FALSE, weight = NULL, .env = NULL, ...)

Arguments

tbl

A data.frame.

size

<tidy-select> For sample_n(), the number of rows to select. For sample_frac(), the fraction of rows to select. If tbl is grouped, size applies to each group.

replace

Sample with or without replacement?

weight

<tidy-select> Sampling weights. This must evaluate to a vector of non-negative numbers the same length as the input. Weights are automatically standardised to sum to 1.

.env

DEPRECATED.

...

ignored

Examples

by_cyl <- mtcars %>% group_by(cyl)

# sample_n() -> slice_sample() ----------------------------------------------
sample_n(mtcars, 10)
#>                      mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#> Merc 450SL          17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3
#> AMC Javelin         15.2   8 304.0 150 3.15 3.435 17.30  0  0    3    2
#> Cadillac Fleetwood  10.4   8 472.0 205 2.93 5.250 17.98  0  0    3    4
#> Toyota Corolla      33.9   4  71.1  65 4.22 1.835 19.90  1  1    4    1
#> Lincoln Continental 10.4   8 460.0 215 3.00 5.424 17.82  0  0    3    4
#> Chrysler Imperial   14.7   8 440.0 230 3.23 5.345 17.42  0  0    3    4
#> Merc 450SE          16.4   8 275.8 180 3.07 4.070 17.40  0  0    3    3
#> Datsun 710          22.8   4 108.0  93 3.85 2.320 18.61  1  1    4    1
#> Mazda RX4 Wag       21.0   6 160.0 110 3.90 2.875 17.02  0  1    4    4
#> Lotus Europa        30.4   4  95.1 113 3.77 1.513 16.90  1  1    5    2
sample_n(mtcars, 50, replace = TRUE)
#>                           mpg cyl  disp  hp drat    wt  qsec vs am gear
#> Merc 240D...1            24.4   4 146.7  62 3.69 3.190 20.00  1  0    4
#> Dodge Challenger         15.5   8 318.0 150 2.76 3.520 16.87  0  0    3
#> Lincoln Continental...3  10.4   8 460.0 215 3.00 5.424 17.82  0  0    3
#> Mazda RX4 Wag            21.0   6 160.0 110 3.90 2.875 17.02  0  1    4
#> Hornet 4 Drive...5       21.4   6 258.0 110 3.08 3.215 19.44  1  0    3
#> AMC Javelin              15.2   8 304.0 150 3.15 3.435 17.30  0  0    3
#> Mazda RX4...7            21.0   6 160.0 110 3.90 2.620 16.46  0  1    4
#> Valiant...8              18.1   6 225.0 105 2.76 3.460 20.22  1  0    3
#> Duster 360...9           14.3   8 360.0 245 3.21 3.570 15.84  0  0    3
#> Hornet Sportabout        18.7   8 360.0 175 3.15 3.440 17.02  0  0    3
#> Merc 450SLC...11         15.2   8 275.8 180 3.07 3.780 18.00  0  0    3
#> Fiat X1-9...12           27.3   4  79.0  66 4.08 1.935 18.90  1  1    4
#> Chrysler Imperial...13   14.7   8 440.0 230 3.23 5.345 17.42  0  0    3
#> Hornet 4 Drive...14      21.4   6 258.0 110 3.08 3.215 19.44  1  0    3
#> Merc 230...15            22.8   4 140.8  95 3.92 3.150 22.90  1  0    4
#> Fiat 128...16            32.4   4  78.7  66 4.08 2.200 19.47  1  1    4
#> Hornet 4 Drive...17      21.4   6 258.0 110 3.08 3.215 19.44  1  0    3
#> Toyota Corolla...18      33.9   4  71.1  65 4.22 1.835 19.90  1  1    4
#> Merc 280                 19.2   6 167.6 123 3.92 3.440 18.30  1  0    4
#> Toyota Corolla...20      33.9   4  71.1  65 4.22 1.835 19.90  1  1    4
#> Merc 450SE               16.4   8 275.8 180 3.07 4.070 17.40  0  0    3
#> Toyota Corona...22       21.5   4 120.1  97 3.70 2.465 20.01  1  0    3
#> Honda Civic...23         30.4   4  75.7  52 4.93 1.615 18.52  1  1    4
#> Merc 240D...24           24.4   4 146.7  62 3.69 3.190 20.00  1  0    4
#> Merc 450SL...25          17.3   8 275.8 180 3.07 3.730 17.60  0  0    3
#> Ferrari Dino             19.7   6 145.0 175 3.62 2.770 15.50  0  1    5
#> Mazda RX4...27           21.0   6 160.0 110 3.90 2.620 16.46  0  1    4
#> Fiat X1-9...28           27.3   4  79.0  66 4.08 1.935 18.90  1  1    4
#> Duster 360...29          14.3   8 360.0 245 3.21 3.570 15.84  0  0    3
#> Volvo 142E...30          21.4   4 121.0 109 4.11 2.780 18.60  1  1    4
#> Merc 450SLC...31         15.2   8 275.8 180 3.07 3.780 18.00  0  0    3
#> Merc 240D...32           24.4   4 146.7  62 3.69 3.190 20.00  1  0    4
#> Toyota Corolla...33      33.9   4  71.1  65 4.22 1.835 19.90  1  1    4
#> Mazda RX4...34           21.0   6 160.0 110 3.90 2.620 16.46  0  1    4
#> Merc 230...35            22.8   4 140.8  95 3.92 3.150 22.90  1  0    4
#> Toyota Corona...36       21.5   4 120.1  97 3.70 2.465 20.01  1  0    3
#> Lincoln Continental...37 10.4   8 460.0 215 3.00 5.424 17.82  0  0    3
#> Ford Pantera L...38      15.8   8 351.0 264 4.22 3.170 14.50  0  1    5
#> Chrysler Imperial...39   14.7   8 440.0 230 3.23 5.345 17.42  0  0    3
#> Merc 450SL...40          17.3   8 275.8 180 3.07 3.730 17.60  0  0    3
#> Merc 280C...41           17.8   6 167.6 123 3.92 3.440 18.90  1  0    4
#> Honda Civic...42         30.4   4  75.7  52 4.93 1.615 18.52  1  1    4
#> Mazda RX4...43           21.0   6 160.0 110 3.90 2.620 16.46  0  1    4
#> Valiant...44             18.1   6 225.0 105 2.76 3.460 20.22  1  0    3
#> Pontiac Firebird         19.2   8 400.0 175 3.08 3.845 17.05  0  0    3
#> Merc 280C...46           17.8   6 167.6 123 3.92 3.440 18.90  1  0    4
#> Fiat 128...47            32.4   4  78.7  66 4.08 2.200 19.47  1  1    4
#> Toyota Corolla...48      33.9   4  71.1  65 4.22 1.835 19.90  1  1    4
#> Volvo 142E...49          21.4   4 121.0 109 4.11 2.780 18.60  1  1    4
#> Ford Pantera L...50      15.8   8 351.0 264 4.22 3.170 14.50  0  1    5
#>                          carb
#> Merc 240D...1               2
#> Dodge Challenger            2
#> Lincoln Continental...3     4
#> Mazda RX4 Wag               4
#> Hornet 4 Drive...5          1
#> AMC Javelin                 2
#> Mazda RX4...7               4
#> Valiant...8                 1
#> Duster 360...9              4
#> Hornet Sportabout           2
#> Merc 450SLC...11            3
#> Fiat X1-9...12              1
#> Chrysler Imperial...13      4
#> Hornet 4 Drive...14         1
#> Merc 230...15               2
#> Fiat 128...16               1
#> Hornet 4 Drive...17         1
#> Toyota Corolla...18         1
#> Merc 280                    4
#> Toyota Corolla...20         1
#> Merc 450SE                  3
#> Toyota Corona...22          1
#> Honda Civic...23            2
#> Merc 240D...24              2
#> Merc 450SL...25             3
#> Ferrari Dino                6
#> Mazda RX4...27              4
#> Fiat X1-9...28              1
#> Duster 360...29             4
#> Volvo 142E...30             2
#> Merc 450SLC...31            3
#> Merc 240D...32              2
#> Toyota Corolla...33         1
#> Mazda RX4...34              4
#> Merc 230...35               2
#> Toyota Corona...36          1
#> Lincoln Continental...37    4
#> Ford Pantera L...38         4
#> Chrysler Imperial...39      4
#> Merc 450SL...40             3
#> Merc 280C...41              4
#> Honda Civic...42            2
#> Mazda RX4...43              4
#> Valiant...44                1
#> Pontiac Firebird            2
#> Merc 280C...46              4
#> Fiat 128...47               1
#> Toyota Corolla...48         1
#> Volvo 142E...49             2
#> Ford Pantera L...50         4
sample_n(mtcars, 10, weight = mpg)
#>                 mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#> Hornet 4 Drive 21.4   6 258.0 110 3.08 3.215 19.44  1  0    3    1
#> Merc 230       22.8   4 140.8  95 3.92 3.150 22.90  1  0    4    2
#> Toyota Corona  21.5   4 120.1  97 3.70 2.465 20.01  1  0    3    1
#> Datsun 710     22.8   4 108.0  93 3.85 2.320 18.61  1  1    4    1
#> Merc 240D      24.4   4 146.7  62 3.69 3.190 20.00  1  0    4    2
#> Merc 280       19.2   6 167.6 123 3.92 3.440 18.30  1  0    4    4
#> Ford Pantera L 15.8   8 351.0 264 4.22 3.170 14.50  0  1    5    4
#> Merc 450SL     17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3
#> Fiat 128       32.4   4  78.7  66 4.08 2.200 19.47  1  1    4    1
#> Merc 450SE     16.4   8 275.8 180 3.07 4.070 17.40  0  0    3    3

# Changes:
# * explicitly name the `n` argument,
# * the `weight` argument is now `weight_by`.

slice_sample(mtcars, n = 10)
#>                     mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#> Merc 240D          24.4   4 146.7  62 3.69 3.190 20.00  1  0    4    2
#> Fiat X1-9          27.3   4  79.0  66 4.08 1.935 18.90  1  1    4    1
#> Lotus Europa       30.4   4  95.1 113 3.77 1.513 16.90  1  1    5    2
#> Merc 450SL         17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3
#> Hornet Sportabout  18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2
#> Mazda RX4 Wag      21.0   6 160.0 110 3.90 2.875 17.02  0  1    4    4
#> Mazda RX4          21.0   6 160.0 110 3.90 2.620 16.46  0  1    4    4
#> Ford Pantera L     15.8   8 351.0 264 4.22 3.170 14.50  0  1    5    4
#> Honda Civic        30.4   4  75.7  52 4.93 1.615 18.52  1  1    4    2
#> Cadillac Fleetwood 10.4   8 472.0 205 2.93 5.250 17.98  0  0    3    4
slice_sample(mtcars, n = 50, replace = TRUE)
#>                          mpg cyl  disp  hp drat    wt  qsec vs am gear
#> Merc 280C...1           17.8   6 167.6 123 3.92 3.440 18.90  1  0    4
#> Lincoln Continental...2 10.4   8 460.0 215 3.00 5.424 17.82  0  0    3
#> Chrysler Imperial...3   14.7   8 440.0 230 3.23 5.345 17.42  0  0    3
#> Cadillac Fleetwood...4  10.4   8 472.0 205 2.93 5.250 17.98  0  0    3
#> Lincoln Continental...5 10.4   8 460.0 215 3.00 5.424 17.82  0  0    3
#> AMC Javelin             15.2   8 304.0 150 3.15 3.435 17.30  0  0    3
#> Porsche 914-2...7       26.0   4 120.3  91 4.43 2.140 16.70  0  1    5
#> Cadillac Fleetwood...8  10.4   8 472.0 205 2.93 5.250 17.98  0  0    3
#> Fiat X1-9...9           27.3   4  79.0  66 4.08 1.935 18.90  1  1    4
#> Porsche 914-2...10      26.0   4 120.3  91 4.43 2.140 16.70  0  1    5
#> Cadillac Fleetwood...11 10.4   8 472.0 205 2.93 5.250 17.98  0  0    3
#> Merc 280C...12          17.8   6 167.6 123 3.92 3.440 18.90  1  0    4
#> Honda Civic...13        30.4   4  75.7  52 4.93 1.615 18.52  1  1    4
#> Pontiac Firebird...14   19.2   8 400.0 175 3.08 3.845 17.05  0  0    3
#> Fiat X1-9...15          27.3   4  79.0  66 4.08 1.935 18.90  1  1    4
#> Chrysler Imperial...16  14.7   8 440.0 230 3.23 5.345 17.42  0  0    3
#> Merc 450SLC...17        15.2   8 275.8 180 3.07 3.780 18.00  0  0    3
#> Valiant...18            18.1   6 225.0 105 2.76 3.460 20.22  1  0    3
#> Ferrari Dino            19.7   6 145.0 175 3.62 2.770 15.50  0  1    5
#> Hornet 4 Drive          21.4   6 258.0 110 3.08 3.215 19.44  1  0    3
#> Merc 450SL              17.3   8 275.8 180 3.07 3.730 17.60  0  0    3
#> Hornet Sportabout...22  18.7   8 360.0 175 3.15 3.440 17.02  0  0    3
#> Fiat X1-9...23          27.3   4  79.0  66 4.08 1.935 18.90  1  1    4
#> Fiat 128...24           32.4   4  78.7  66 4.08 2.200 19.47  1  1    4
#> Fiat X1-9...25          27.3   4  79.0  66 4.08 1.935 18.90  1  1    4
#> Pontiac Firebird...26   19.2   8 400.0 175 3.08 3.845 17.05  0  0    3
#> Merc 240D...27          24.4   4 146.7  62 3.69 3.190 20.00  1  0    4
#> Merc 450SLC...28        15.2   8 275.8 180 3.07 3.780 18.00  0  0    3
#> Valiant...29            18.1   6 225.0 105 2.76 3.460 20.22  1  0    3
#> Mazda RX4               21.0   6 160.0 110 3.90 2.620 16.46  0  1    4
#> Merc 240D...31          24.4   4 146.7  62 3.69 3.190 20.00  1  0    4
#> Valiant...32            18.1   6 225.0 105 2.76 3.460 20.22  1  0    3
#> Pontiac Firebird...33   19.2   8 400.0 175 3.08 3.845 17.05  0  0    3
#> Merc 280                19.2   6 167.6 123 3.92 3.440 18.30  1  0    4
#> Merc 230...35           22.8   4 140.8  95 3.92 3.150 22.90  1  0    4
#> Honda Civic...36        30.4   4  75.7  52 4.93 1.615 18.52  1  1    4
#> Toyota Corolla...37     33.9   4  71.1  65 4.22 1.835 19.90  1  1    4
#> Lotus Europa            30.4   4  95.1 113 3.77 1.513 16.90  1  1    5
#> Toyota Corolla...39     33.9   4  71.1  65 4.22 1.835 19.90  1  1    4
#> Pontiac Firebird...40   19.2   8 400.0 175 3.08 3.845 17.05  0  0    3
#> Toyota Corona           21.5   4 120.1  97 3.70 2.465 20.01  1  0    3
#> Chrysler Imperial...42  14.7   8 440.0 230 3.23 5.345 17.42  0  0    3
#> Fiat X1-9...43          27.3   4  79.0  66 4.08 1.935 18.90  1  1    4
#> Porsche 914-2...44      26.0   4 120.3  91 4.43 2.140 16.70  0  1    5
#> Merc 240D...45          24.4   4 146.7  62 3.69 3.190 20.00  1  0    4
#> Fiat X1-9...46          27.3   4  79.0  66 4.08 1.935 18.90  1  1    4
#> Hornet Sportabout...47  18.7   8 360.0 175 3.15 3.440 17.02  0  0    3
#> Merc 450SE              16.4   8 275.8 180 3.07 4.070 17.40  0  0    3
#> Fiat 128...49           32.4   4  78.7  66 4.08 2.200 19.47  1  1    4
#> Merc 230...50           22.8   4 140.8  95 3.92 3.150 22.90  1  0    4
#>                         carb
#> Merc 280C...1              4
#> Lincoln Continental...2    4
#> Chrysler Imperial...3      4
#> Cadillac Fleetwood...4     4
#> Lincoln Continental...5    4
#> AMC Javelin                2
#> Porsche 914-2...7          2
#> Cadillac Fleetwood...8     4
#> Fiat X1-9...9              1
#> Porsche 914-2...10         2
#> Cadillac Fleetwood...11    4
#> Merc 280C...12             4
#> Honda Civic...13           2
#> Pontiac Firebird...14      2
#> Fiat X1-9...15             1
#> Chrysler Imperial...16     4
#> Merc 450SLC...17           3
#> Valiant...18               1
#> Ferrari Dino               6
#> Hornet 4 Drive             1
#> Merc 450SL                 3
#> Hornet Sportabout...22     2
#> Fiat X1-9...23             1
#> Fiat 128...24              1
#> Fiat X1-9...25             1
#> Pontiac Firebird...26      2
#> Merc 240D...27             2
#> Merc 450SLC...28           3
#> Valiant...29               1
#> Mazda RX4                  4
#> Merc 240D...31             2
#> Valiant...32               1
#> Pontiac Firebird...33      2
#> Merc 280                   4
#> Merc 230...35              2
#> Honda Civic...36           2
#> Toyota Corolla...37        1
#> Lotus Europa               2
#> Toyota Corolla...39        1
#> Pontiac Firebird...40      2
#> Toyota Corona              1
#> Chrysler Imperial...42     4
#> Fiat X1-9...43             1
#> Porsche 914-2...44         2
#> Merc 240D...45             2
#> Fiat X1-9...46             1
#> Hornet Sportabout...47     2
#> Merc 450SE                 3
#> Fiat 128...49              1
#> Merc 230...50              2
slice_sample(mtcars, n = 10, weight_by = mpg)
#>                     mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#> Mazda RX4 Wag      21.0   6 160.0 110 3.90 2.875 17.02  0  1    4    4
#> Lotus Europa       30.4   4  95.1 113 3.77 1.513 16.90  1  1    5    2
#> Cadillac Fleetwood 10.4   8 472.0 205 2.93 5.250 17.98  0  0    3    4
#> Merc 240D          24.4   4 146.7  62 3.69 3.190 20.00  1  0    4    2
#> Ford Pantera L     15.8   8 351.0 264 4.22 3.170 14.50  0  1    5    4
#> Fiat 128           32.4   4  78.7  66 4.08 2.200 19.47  1  1    4    1
#> Chrysler Imperial  14.7   8 440.0 230 3.23 5.345 17.42  0  0    3    4
#> Volvo 142E         21.4   4 121.0 109 4.11 2.780 18.60  1  1    4    2
#> AMC Javelin        15.2   8 304.0 150 3.15 3.435 17.30  0  0    3    2
#> Valiant            18.1   6 225.0 105 2.76 3.460 20.22  1  0    3    1

# Note that sample_n() would error if n was bigger than the group size
# slice_sample() will just use the available rows for consistency with
# the other slice helpers like slice_head()

# sample_frac() -> slice_sample() -------------------------------------------
sample_frac(mtcars)
#>                      mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#> Chrysler Imperial   14.7   8 440.0 230 3.23 5.345 17.42  0  0    3    4
#> Porsche 914-2       26.0   4 120.3  91 4.43 2.140 16.70  0  1    5    2
#> Camaro Z28          13.3   8 350.0 245 3.73 3.840 15.41  0  0    3    4
#> Merc 450SLC         15.2   8 275.8 180 3.07 3.780 18.00  0  0    3    3
#> Merc 280            19.2   6 167.6 123 3.92 3.440 18.30  1  0    4    4
#> Pontiac Firebird    19.2   8 400.0 175 3.08 3.845 17.05  0  0    3    2
#> Valiant             18.1   6 225.0 105 2.76 3.460 20.22  1  0    3    1
#> Merc 240D           24.4   4 146.7  62 3.69 3.190 20.00  1  0    4    2
#> Merc 280C           17.8   6 167.6 123 3.92 3.440 18.90  1  0    4    4
#> Merc 450SE          16.4   8 275.8 180 3.07 4.070 17.40  0  0    3    3
#> Maserati Bora       15.0   8 301.0 335 3.54 3.570 14.60  0  1    5    8
#> Volvo 142E          21.4   4 121.0 109 4.11 2.780 18.60  1  1    4    2
#> Mazda RX4           21.0   6 160.0 110 3.90 2.620 16.46  0  1    4    4
#> Fiat 128            32.4   4  78.7  66 4.08 2.200 19.47  1  1    4    1
#> Ford Pantera L      15.8   8 351.0 264 4.22 3.170 14.50  0  1    5    4
#> Datsun 710          22.8   4 108.0  93 3.85 2.320 18.61  1  1    4    1
#> Cadillac Fleetwood  10.4   8 472.0 205 2.93 5.250 17.98  0  0    3    4
#> Honda Civic         30.4   4  75.7  52 4.93 1.615 18.52  1  1    4    2
#> Fiat X1-9           27.3   4  79.0  66 4.08 1.935 18.90  1  1    4    1
#> Toyota Corolla      33.9   4  71.1  65 4.22 1.835 19.90  1  1    4    1
#> Dodge Challenger    15.5   8 318.0 150 2.76 3.520 16.87  0  0    3    2
#> AMC Javelin         15.2   8 304.0 150 3.15 3.435 17.30  0  0    3    2
#> Merc 450SL          17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3
#> Hornet Sportabout   18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2
#> Merc 230            22.8   4 140.8  95 3.92 3.150 22.90  1  0    4    2
#> Lincoln Continental 10.4   8 460.0 215 3.00 5.424 17.82  0  0    3    4
#> Hornet 4 Drive      21.4   6 258.0 110 3.08 3.215 19.44  1  0    3    1
#> Lotus Europa        30.4   4  95.1 113 3.77 1.513 16.90  1  1    5    2
#> Toyota Corona       21.5   4 120.1  97 3.70 2.465 20.01  1  0    3    1
#> Ferrari Dino        19.7   6 145.0 175 3.62 2.770 15.50  0  1    5    6
#> Mazda RX4 Wag       21.0   6 160.0 110 3.90 2.875 17.02  0  1    4    4
#> Duster 360          14.3   8 360.0 245 3.21 3.570 15.84  0  0    3    4
sample_frac(mtcars, replace = TRUE)
#>                      mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#> Hornet 4 Drive...1  21.4   6 258.0 110 3.08 3.215 19.44  1  0    3    1
#> Toyota Corona...2   21.5   4 120.1  97 3.70 2.465 20.01  1  0    3    1
#> Duster 360...3      14.3   8 360.0 245 3.21 3.570 15.84  0  0    3    4
#> Mazda RX4 Wag...4   21.0   6 160.0 110 3.90 2.875 17.02  0  1    4    4
#> AMC Javelin...5     15.2   8 304.0 150 3.15 3.435 17.30  0  0    3    2
#> Duster 360...6      14.3   8 360.0 245 3.21 3.570 15.84  0  0    3    4
#> Duster 360...7      14.3   8 360.0 245 3.21 3.570 15.84  0  0    3    4
#> Fiat 128            32.4   4  78.7  66 4.08 2.200 19.47  1  1    4    1
#> Maserati Bora       15.0   8 301.0 335 3.54 3.570 14.60  0  1    5    8
#> Toyota Corona...10  21.5   4 120.1  97 3.70 2.465 20.01  1  0    3    1
#> Datsun 710...11     22.8   4 108.0  93 3.85 2.320 18.61  1  1    4    1
#> Merc 280C...12      17.8   6 167.6 123 3.92 3.440 18.90  1  0    4    4
#> Porsche 914-2...13  26.0   4 120.3  91 4.43 2.140 16.70  0  1    5    2
#> Toyota Corolla      33.9   4  71.1  65 4.22 1.835 19.90  1  1    4    1
#> Pontiac Firebird    19.2   8 400.0 175 3.08 3.845 17.05  0  0    3    2
#> Merc 280            19.2   6 167.6 123 3.92 3.440 18.30  1  0    4    4
#> AMC Javelin...17    15.2   8 304.0 150 3.15 3.435 17.30  0  0    3    2
#> Duster 360...18     14.3   8 360.0 245 3.21 3.570 15.84  0  0    3    4
#> Toyota Corona...19  21.5   4 120.1  97 3.70 2.465 20.01  1  0    3    1
#> Merc 230...20       22.8   4 140.8  95 3.92 3.150 22.90  1  0    4    2
#> Ferrari Dino        19.7   6 145.0 175 3.62 2.770 15.50  0  1    5    6
#> Merc 280C...22      17.8   6 167.6 123 3.92 3.440 18.90  1  0    4    4
#> Mazda RX4 Wag...23  21.0   6 160.0 110 3.90 2.875 17.02  0  1    4    4
#> Dodge Challenger    15.5   8 318.0 150 2.76 3.520 16.87  0  0    3    2
#> Merc 230...25       22.8   4 140.8  95 3.92 3.150 22.90  1  0    4    2
#> Ford Pantera L      15.8   8 351.0 264 4.22 3.170 14.50  0  1    5    4
#> Merc 450SL          17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3
#> Hornet 4 Drive...28 21.4   6 258.0 110 3.08 3.215 19.44  1  0    3    1
#> Porsche 914-2...29  26.0   4 120.3  91 4.43 2.140 16.70  0  1    5    2
#> Datsun 710...30     22.8   4 108.0  93 3.85 2.320 18.61  1  1    4    1
#> AMC Javelin...31    15.2   8 304.0 150 3.15 3.435 17.30  0  0    3    2
#> Hornet Sportabout   18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2

# Changes:
# * use prop = 1 to randomly sample all rows

slice_sample(mtcars, prop = 1)
#>                      mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#> Mazda RX4           21.0   6 160.0 110 3.90 2.620 16.46  0  1    4    4
#> Valiant             18.1   6 225.0 105 2.76 3.460 20.22  1  0    3    1
#> Cadillac Fleetwood  10.4   8 472.0 205 2.93 5.250 17.98  0  0    3    4
#> Ferrari Dino        19.7   6 145.0 175 3.62 2.770 15.50  0  1    5    6
#> Hornet Sportabout   18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2
#> Volvo 142E          21.4   4 121.0 109 4.11 2.780 18.60  1  1    4    2
#> AMC Javelin         15.2   8 304.0 150 3.15 3.435 17.30  0  0    3    2
#> Merc 450SLC         15.2   8 275.8 180 3.07 3.780 18.00  0  0    3    3
#> Maserati Bora       15.0   8 301.0 335 3.54 3.570 14.60  0  1    5    8
#> Mazda RX4 Wag       21.0   6 160.0 110 3.90 2.875 17.02  0  1    4    4
#> Merc 280C           17.8   6 167.6 123 3.92 3.440 18.90  1  0    4    4
#> Merc 450SE          16.4   8 275.8 180 3.07 4.070 17.40  0  0    3    3
#> Merc 450SL          17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3
#> Hornet 4 Drive      21.4   6 258.0 110 3.08 3.215 19.44  1  0    3    1
#> Datsun 710          22.8   4 108.0  93 3.85 2.320 18.61  1  1    4    1
#> Dodge Challenger    15.5   8 318.0 150 2.76 3.520 16.87  0  0    3    2
#> Honda Civic         30.4   4  75.7  52 4.93 1.615 18.52  1  1    4    2
#> Toyota Corona       21.5   4 120.1  97 3.70 2.465 20.01  1  0    3    1
#> Merc 240D           24.4   4 146.7  62 3.69 3.190 20.00  1  0    4    2
#> Lotus Europa        30.4   4  95.1 113 3.77 1.513 16.90  1  1    5    2
#> Lincoln Continental 10.4   8 460.0 215 3.00 5.424 17.82  0  0    3    4
#> Fiat X1-9           27.3   4  79.0  66 4.08 1.935 18.90  1  1    4    1
#> Porsche 914-2       26.0   4 120.3  91 4.43 2.140 16.70  0  1    5    2
#> Pontiac Firebird    19.2   8 400.0 175 3.08 3.845 17.05  0  0    3    2
#> Duster 360          14.3   8 360.0 245 3.21 3.570 15.84  0  0    3    4
#> Merc 230            22.8   4 140.8  95 3.92 3.150 22.90  1  0    4    2
#> Fiat 128            32.4   4  78.7  66 4.08 2.200 19.47  1  1    4    1
#> Merc 280            19.2   6 167.6 123 3.92 3.440 18.30  1  0    4    4
#> Toyota Corolla      33.9   4  71.1  65 4.22 1.835 19.90  1  1    4    1
#> Ford Pantera L      15.8   8 351.0 264 4.22 3.170 14.50  0  1    5    4
#> Camaro Z28          13.3   8 350.0 245 3.73 3.840 15.41  0  0    3    4
#> Chrysler Imperial   14.7   8 440.0 230 3.23 5.345 17.42  0  0    3    4
slice_sample(mtcars, prop = 1, replace = TRUE)
#>                          mpg cyl  disp  hp drat    wt  qsec vs am gear
#> Merc 450SL...1          17.3   8 275.8 180 3.07 3.730 17.60  0  0    3
#> Porsche 914-2           26.0   4 120.3  91 4.43 2.140 16.70  0  1    5
#> Mazda RX4 Wag...3       21.0   6 160.0 110 3.90 2.875 17.02  0  1    4
#> Lincoln Continental...4 10.4   8 460.0 215 3.00 5.424 17.82  0  0    3
#> Honda Civic...5         30.4   4  75.7  52 4.93 1.615 18.52  1  1    4
#> Ford Pantera L...6      15.8   8 351.0 264 4.22 3.170 14.50  0  1    5
#> Lincoln Continental...7 10.4   8 460.0 215 3.00 5.424 17.82  0  0    3
#> Fiat 128...8            32.4   4  78.7  66 4.08 2.200 19.47  1  1    4
#> Datsun 710              22.8   4 108.0  93 3.85 2.320 18.61  1  1    4
#> Dodge Challenger...10   15.5   8 318.0 150 2.76 3.520 16.87  0  0    3
#> Toyota Corolla...11     33.9   4  71.1  65 4.22 1.835 19.90  1  1    4
#> Merc 280C               17.8   6 167.6 123 3.92 3.440 18.90  1  0    4
#> AMC Javelin...13        15.2   8 304.0 150 3.15 3.435 17.30  0  0    3
#> Merc 450SL...14         17.3   8 275.8 180 3.07 3.730 17.60  0  0    3
#> Toyota Corona...15      21.5   4 120.1  97 3.70 2.465 20.01  1  0    3
#> Mazda RX4 Wag...16      21.0   6 160.0 110 3.90 2.875 17.02  0  1    4
#> Dodge Challenger...17   15.5   8 318.0 150 2.76 3.520 16.87  0  0    3
#> Ford Pantera L...18     15.8   8 351.0 264 4.22 3.170 14.50  0  1    5
#> Ford Pantera L...19     15.8   8 351.0 264 4.22 3.170 14.50  0  1    5
#> Merc 230                22.8   4 140.8  95 3.92 3.150 22.90  1  0    4
#> Ferrari Dino            19.7   6 145.0 175 3.62 2.770 15.50  0  1    5
#> Fiat 128...22           32.4   4  78.7  66 4.08 2.200 19.47  1  1    4
#> Cadillac Fleetwood      10.4   8 472.0 205 2.93 5.250 17.98  0  0    3
#> AMC Javelin...24        15.2   8 304.0 150 3.15 3.435 17.30  0  0    3
#> Maserati Bora           15.0   8 301.0 335 3.54 3.570 14.60  0  1    5
#> AMC Javelin...26        15.2   8 304.0 150 3.15 3.435 17.30  0  0    3
#> Toyota Corona...27      21.5   4 120.1  97 3.70 2.465 20.01  1  0    3
#> Mazda RX4 Wag...28      21.0   6 160.0 110 3.90 2.875 17.02  0  1    4
#> AMC Javelin...29        15.2   8 304.0 150 3.15 3.435 17.30  0  0    3
#> Honda Civic...30        30.4   4  75.7  52 4.93 1.615 18.52  1  1    4
#> Toyota Corolla...31     33.9   4  71.1  65 4.22 1.835 19.90  1  1    4
#> Duster 360              14.3   8 360.0 245 3.21 3.570 15.84  0  0    3
#>                         carb
#> Merc 450SL...1             3
#> Porsche 914-2              2
#> Mazda RX4 Wag...3          4
#> Lincoln Continental...4    4
#> Honda Civic...5            2
#> Ford Pantera L...6         4
#> Lincoln Continental...7    4
#> Fiat 128...8               1
#> Datsun 710                 1
#> Dodge Challenger...10      2
#> Toyota Corolla...11        1
#> Merc 280C                  4
#> AMC Javelin...13           2
#> Merc 450SL...14            3
#> Toyota Corona...15         1
#> Mazda RX4 Wag...16         4
#> Dodge Challenger...17      2
#> Ford Pantera L...18        4
#> Ford Pantera L...19        4
#> Merc 230                   2
#> Ferrari Dino               6
#> Fiat 128...22              1
#> Cadillac Fleetwood         4
#> AMC Javelin...24           2
#> Maserati Bora              8
#> AMC Javelin...26           2
#> Toyota Corona...27         1
#> Mazda RX4 Wag...28         4
#> AMC Javelin...29           2
#> Honda Civic...30           2
#> Toyota Corolla...31        1
#> Duster 360                 4