Sample n rows from a table

sample_n() and sample_frac() have been retired in favour of slice_sample(). While they will not be deprecated in the near future, retirement means that we will only perform critical bug fixes, so we recommend moving to the newer alternative.

These functions were retired because we realised it was more convenient to have two mutually exclusive arguments to one function, rather than two separate functions. This also made it to clean up a few other smaller design issues with sample_n()/sample_frac:

The connection to slice() was not obvious.

The name of the first argument, tbl, is inconsistent with other single table verbs which use .data.

The size argument uses tidy evaluation, which is surprising and undocumented.

It was easier to remove the deprecated .env argument.

... was in a suboptimal position.

sample_n(tbl, size, replace = FALSE, weight = NULL, .env = NULL, ...)

sample_frac(tbl, size = 1, replace = FALSE, weight = NULL, .env = NULL, ...)

Arguments

tbl
A data.frame.

size
<tidy-select> For sample_n(), the number of rows to select. For sample_frac(), the fraction of rows to select. If tbl is grouped, size applies to each group.

replace
Sample with or without replacement?

weight
<tidy-select> Sampling weights. This must evaluate to a vector of non-negative numbers the same length as the input. Weights are automatically standardised to sum to 1.

.env
DEPRECATED.

...
ignored

Examples

by_cyl <- mtcars %>% group_by(cyl)

# sample_n() -> slice_sample() ----------------------------------------------
sample_n(mtcars, 10)

#>     mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#> 1  22.8   4 108.0  93 3.85 2.320 18.61  1  1    4    1
#> 2  14.7   8 440.0 230 3.23 5.345 17.42  0  0    3    4
#> 3  10.4   8 472.0 205 2.93 5.250 17.98  0  0    3    4
#> 4  16.4   8 275.8 180 3.07 4.070 17.40  0  0    3    3
#> 5  19.2   6 167.6 123 3.92 3.440 18.30  1  0    4    4
#> 6  19.2   8 400.0 175 3.08 3.845 17.05  0  0    3    2
#> 7  21.0   6 160.0 110 3.90 2.620 16.46  0  1    4    4
#> 8  24.4   4 146.7  62 3.69 3.190 20.00  1  0    4    2
#> 9  33.9   4  71.1  65 4.22 1.835 19.90  1  1    4    1
#> 10 17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3

sample_n(mtcars, 50, replace = TRUE)

#>     mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#> 1  13.3   8 350.0 245 3.73 3.840 15.41  0  0    3    4
#> 2  14.7   8 440.0 230 3.23 5.345 17.42  0  0    3    4
#> 3  15.2   8 304.0 150 3.15 3.435 17.30  0  0    3    2
#> 4  17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3
#> 5  17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3
#> 6  22.8   4 140.8  95 3.92 3.150 22.90  1  0    4    2
#> 7  22.8   4 140.8  95 3.92 3.150 22.90  1  0    4    2
#> 8  15.0   8 301.0 335 3.54 3.570 14.60  0  1    5    8
#> 9  17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3
#> 10 33.9   4  71.1  65 4.22 1.835 19.90  1  1    4    1
#> 11 15.2   8 304.0 150 3.15 3.435 17.30  0  0    3    2
#> 12 15.2   8 275.8 180 3.07 3.780 18.00  0  0    3    3
#> 13 24.4   4 146.7  62 3.69 3.190 20.00  1  0    4    2
#> 14 33.9   4  71.1  65 4.22 1.835 19.90  1  1    4    1
#> 15 14.3   8 360.0 245 3.21 3.570 15.84  0  0    3    4
#> 16 33.9   4  71.1  65 4.22 1.835 19.90  1  1    4    1
#> 17 16.4   8 275.8 180 3.07 4.070 17.40  0  0    3    3
#> 18 17.8   6 167.6 123 3.92 3.440 18.90  1  0    4    4
#> 19 19.2   8 400.0 175 3.08 3.845 17.05  0  0    3    2
#> 20 19.2   6 167.6 123 3.92 3.440 18.30  1  0    4    4
#> 21 21.5   4 120.1  97 3.70 2.465 20.01  1  0    3    1
#> 22 13.3   8 350.0 245 3.73 3.840 15.41  0  0    3    4
#> 23 15.2   8 304.0 150 3.15 3.435 17.30  0  0    3    2
#> 24 19.2   8 400.0 175 3.08 3.845 17.05  0  0    3    2
#> 25 21.5   4 120.1  97 3.70 2.465 20.01  1  0    3    1
#> 26 30.4   4  95.1 113 3.77 1.513 16.90  1  1    5    2
#> 27 15.5   8 318.0 150 2.76 3.520 16.87  0  0    3    2
#> 28 14.7   8 440.0 230 3.23 5.345 17.42  0  0    3    4
#> 29 21.5   4 120.1  97 3.70 2.465 20.01  1  0    3    1
#> 30 26.0   4 120.3  91 4.43 2.140 16.70  0  1    5    2
#> 31 15.2   8 304.0 150 3.15 3.435 17.30  0  0    3    2
#> 32 13.3   8 350.0 245 3.73 3.840 15.41  0  0    3    4
#> 33 19.7   6 145.0 175 3.62 2.770 15.50  0  1    5    6
#> 34 30.4   4  95.1 113 3.77 1.513 16.90  1  1    5    2
#> 35 30.4   4  75.7  52 4.93 1.615 18.52  1  1    4    2
#> 36 15.5   8 318.0 150 2.76 3.520 16.87  0  0    3    2
#> 37 15.5   8 318.0 150 2.76 3.520 16.87  0  0    3    2
#> 38 21.4   6 258.0 110 3.08 3.215 19.44  1  0    3    1
#> 39 17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3
#> 40 10.4   8 472.0 205 2.93 5.250 17.98  0  0    3    4
#> 41 17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3
#> 42 32.4   4  78.7  66 4.08 2.200 19.47  1  1    4    1
#> 43 32.4   4  78.7  66 4.08 2.200 19.47  1  1    4    1
#> 44 26.0   4 120.3  91 4.43 2.140 16.70  0  1    5    2
#> 45 30.4   4  75.7  52 4.93 1.615 18.52  1  1    4    2
#> 46 17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3
#> 47 22.8   4 140.8  95 3.92 3.150 22.90  1  0    4    2
#> 48 22.8   4 140.8  95 3.92 3.150 22.90  1  0    4    2
#> 49 21.0   6 160.0 110 3.90 2.620 16.46  0  1    4    4
#> 50 21.4   4 121.0 109 4.11 2.780 18.60  1  1    4    2

sample_n(mtcars, 10, weight = mpg)

#>     mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#> 1  22.8   4 108.0  93 3.85 2.320 18.61  1  1    4    1
#> 2  18.1   6 225.0 105 2.76 3.460 20.22  1  0    3    1
#> 3  18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2
#> 4  15.0   8 301.0 335 3.54 3.570 14.60  0  1    5    8
#> 5  15.8   8 351.0 264 4.22 3.170 14.50  0  1    5    4
#> 6  16.4   8 275.8 180 3.07 4.070 17.40  0  0    3    3
#> 7  22.8   4 140.8  95 3.92 3.150 22.90  1  0    4    2
#> 8  30.4   4  75.7  52 4.93 1.615 18.52  1  1    4    2
#> 9  17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3
#> 10 21.0   6 160.0 110 3.90 2.875 17.02  0  1    4    4

# Changes:
# * explicitly name the `n` argument,
# * the `weight` argument is now `weight_by`.

slice_sample(mtcars, n = 10)

#>     mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#> 1  15.8   8 351.0 264 4.22 3.170 14.50  0  1    5    4
#> 2  18.1   6 225.0 105 2.76 3.460 20.22  1  0    3    1
#> 3  26.0   4 120.3  91 4.43 2.140 16.70  0  1    5    2
#> 4  19.2   6 167.6 123 3.92 3.440 18.30  1  0    4    4
#> 5  15.5   8 318.0 150 2.76 3.520 16.87  0  0    3    2
#> 6  15.0   8 301.0 335 3.54 3.570 14.60  0  1    5    8
#> 7  30.4   4  95.1 113 3.77 1.513 16.90  1  1    5    2
#> 8  15.2   8 275.8 180 3.07 3.780 18.00  0  0    3    3
#> 9  21.4   6 258.0 110 3.08 3.215 19.44  1  0    3    1
#> 10 17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3

slice_sample(mtcars, n = 50, replace = TRUE)

#>     mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#> 1  21.0   6 160.0 110 3.90 2.620 16.46  0  1    4    4
#> 2  15.2   8 304.0 150 3.15 3.435 17.30  0  0    3    2
#> 3  21.4   4 121.0 109 4.11 2.780 18.60  1  1    4    2
#> 4  18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2
#> 5  16.4   8 275.8 180 3.07 4.070 17.40  0  0    3    3
#> 6  17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3
#> 7  21.4   4 121.0 109 4.11 2.780 18.60  1  1    4    2
#> 8  10.4   8 472.0 205 2.93 5.250 17.98  0  0    3    4
#> 9  27.3   4  79.0  66 4.08 1.935 18.90  1  1    4    1
#> 10 15.2   8 304.0 150 3.15 3.435 17.30  0  0    3    2
#> 11 17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3
#> 12 21.4   6 258.0 110 3.08 3.215 19.44  1  0    3    1
#> 13 15.0   8 301.0 335 3.54 3.570 14.60  0  1    5    8
#> 14 14.7   8 440.0 230 3.23 5.345 17.42  0  0    3    4
#> 15 16.4   8 275.8 180 3.07 4.070 17.40  0  0    3    3
#> 16 21.4   4 121.0 109 4.11 2.780 18.60  1  1    4    2
#> 17 33.9   4  71.1  65 4.22 1.835 19.90  1  1    4    1
#> 18 18.1   6 225.0 105 2.76 3.460 20.22  1  0    3    1
#> 19 13.3   8 350.0 245 3.73 3.840 15.41  0  0    3    4
#> 20 15.0   8 301.0 335 3.54 3.570 14.60  0  1    5    8
#> 21 21.5   4 120.1  97 3.70 2.465 20.01  1  0    3    1
#> 22 19.2   6 167.6 123 3.92 3.440 18.30  1  0    4    4
#> 23 17.8   6 167.6 123 3.92 3.440 18.90  1  0    4    4
#> 24 18.1   6 225.0 105 2.76 3.460 20.22  1  0    3    1
#> 25 17.8   6 167.6 123 3.92 3.440 18.90  1  0    4    4
#> 26 30.4   4  95.1 113 3.77 1.513 16.90  1  1    5    2
#> 27 14.7   8 440.0 230 3.23 5.345 17.42  0  0    3    4
#> 28 15.8   8 351.0 264 4.22 3.170 14.50  0  1    5    4
#> 29 24.4   4 146.7  62 3.69 3.190 20.00  1  0    4    2
#> 30 13.3   8 350.0 245 3.73 3.840 15.41  0  0    3    4
#> 31 15.0   8 301.0 335 3.54 3.570 14.60  0  1    5    8
#> 32 14.7   8 440.0 230 3.23 5.345 17.42  0  0    3    4
#> 33 21.0   6 160.0 110 3.90 2.875 17.02  0  1    4    4
#> 34 18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2
#> 35 19.2   8 400.0 175 3.08 3.845 17.05  0  0    3    2
#> 36 14.3   8 360.0 245 3.21 3.570 15.84  0  0    3    4
#> 37 17.8   6 167.6 123 3.92 3.440 18.90  1  0    4    4
#> 38 16.4   8 275.8 180 3.07 4.070 17.40  0  0    3    3
#> 39 10.4   8 472.0 205 2.93 5.250 17.98  0  0    3    4
#> 40 30.4   4  75.7  52 4.93 1.615 18.52  1  1    4    2
#> 41 18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2
#> 42 17.8   6 167.6 123 3.92 3.440 18.90  1  0    4    4
#> 43 15.2   8 304.0 150 3.15 3.435 17.30  0  0    3    2
#> 44 15.2   8 275.8 180 3.07 3.780 18.00  0  0    3    3
#> 45 21.0   6 160.0 110 3.90 2.875 17.02  0  1    4    4
#> 46 32.4   4  78.7  66 4.08 2.200 19.47  1  1    4    1
#> 47 21.4   6 258.0 110 3.08 3.215 19.44  1  0    3    1
#> 48 13.3   8 350.0 245 3.73 3.840 15.41  0  0    3    4
#> 49 27.3   4  79.0  66 4.08 1.935 18.90  1  1    4    1
#> 50 19.2   8 400.0 175 3.08 3.845 17.05  0  0    3    2

slice_sample(mtcars, n = 10, weight_by = mpg)

#>     mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#> 1  13.3   8 350.0 245 3.73 3.840 15.41  0  0    3    4
#> 2  18.1   6 225.0 105 2.76 3.460 20.22  1  0    3    1
#> 3  17.8   6 167.6 123 3.92 3.440 18.90  1  0    4    4
#> 4  14.3   8 360.0 245 3.21 3.570 15.84  0  0    3    4
#> 5  21.4   6 258.0 110 3.08 3.215 19.44  1  0    3    1
#> 6  17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3
#> 7  15.5   8 318.0 150 2.76 3.520 16.87  0  0    3    2
#> 8  15.0   8 301.0 335 3.54 3.570 14.60  0  1    5    8
#> 9  21.0   6 160.0 110 3.90 2.620 16.46  0  1    4    4
#> 10 10.4   8 472.0 205 2.93 5.250 17.98  0  0    3    4

# Note that sample_n() would error if n was bigger than the group size
# slice_sample() will just use the available rows for consistency with
# the other slice helpers like slice_head()

# sample_frac() -> slice_sample() -------------------------------------------
sample_frac(mtcars)

#>     mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#> 1  26.0   4 120.3  91 4.43 2.140 16.70  0  1    5    2
#> 2  21.0   6 160.0 110 3.90 2.875 17.02  0  1    4    4
#> 3  17.8   6 167.6 123 3.92 3.440 18.90  1  0    4    4
#> 4  14.3   8 360.0 245 3.21 3.570 15.84  0  0    3    4
#> 5  30.4   4  75.7  52 4.93 1.615 18.52  1  1    4    2
#> 6  18.1   6 225.0 105 2.76 3.460 20.22  1  0    3    1
#> 7  22.8   4 108.0  93 3.85 2.320 18.61  1  1    4    1
#> 8  16.4   8 275.8 180 3.07 4.070 17.40  0  0    3    3
#> 9  21.4   4 121.0 109 4.11 2.780 18.60  1  1    4    2
#> 10 13.3   8 350.0 245 3.73 3.840 15.41  0  0    3    4
#> 11 19.2   8 400.0 175 3.08 3.845 17.05  0  0    3    2
#> 12 10.4   8 472.0 205 2.93 5.250 17.98  0  0    3    4
#> 13 30.4   4  95.1 113 3.77 1.513 16.90  1  1    5    2
#> 14 14.7   8 440.0 230 3.23 5.345 17.42  0  0    3    4
#> 15 21.0   6 160.0 110 3.90 2.620 16.46  0  1    4    4
#> 16 32.4   4  78.7  66 4.08 2.200 19.47  1  1    4    1
#> 17 19.7   6 145.0 175 3.62 2.770 15.50  0  1    5    6
#> 18 22.8   4 140.8  95 3.92 3.150 22.90  1  0    4    2
#> 19 27.3   4  79.0  66 4.08 1.935 18.90  1  1    4    1
#> 20 21.5   4 120.1  97 3.70 2.465 20.01  1  0    3    1
#> 21 21.4   6 258.0 110 3.08 3.215 19.44  1  0    3    1
#> 22 17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3
#> 23 15.2   8 304.0 150 3.15 3.435 17.30  0  0    3    2
#> 24 33.9   4  71.1  65 4.22 1.835 19.90  1  1    4    1
#> 25 15.5   8 318.0 150 2.76 3.520 16.87  0  0    3    2
#> 26 10.4   8 460.0 215 3.00 5.424 17.82  0  0    3    4
#> 27 15.8   8 351.0 264 4.22 3.170 14.50  0  1    5    4
#> 28 18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2
#> 29 15.0   8 301.0 335 3.54 3.570 14.60  0  1    5    8
#> 30 24.4   4 146.7  62 3.69 3.190 20.00  1  0    4    2
#> 31 19.2   6 167.6 123 3.92 3.440 18.30  1  0    4    4
#> 32 15.2   8 275.8 180 3.07 3.780 18.00  0  0    3    3

sample_frac(mtcars, replace = TRUE)

#>     mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#> 1  21.0   6 160.0 110 3.90 2.875 17.02  0  1    4    4
#> 2  30.4   4  95.1 113 3.77 1.513 16.90  1  1    5    2
#> 3  30.4   4  75.7  52 4.93 1.615 18.52  1  1    4    2
#> 4  15.0   8 301.0 335 3.54 3.570 14.60  0  1    5    8
#> 5  19.7   6 145.0 175 3.62 2.770 15.50  0  1    5    6
#> 6  32.4   4  78.7  66 4.08 2.200 19.47  1  1    4    1
#> 7  21.4   6 258.0 110 3.08 3.215 19.44  1  0    3    1
#> 8  30.4   4  75.7  52 4.93 1.615 18.52  1  1    4    2
#> 9  17.8   6 167.6 123 3.92 3.440 18.90  1  0    4    4
#> 10 13.3   8 350.0 245 3.73 3.840 15.41  0  0    3    4
#> 11 30.4   4  95.1 113 3.77 1.513 16.90  1  1    5    2
#> 12 14.3   8 360.0 245 3.21 3.570 15.84  0  0    3    4
#> 13 10.4   8 472.0 205 2.93 5.250 17.98  0  0    3    4
#> 14 21.4   6 258.0 110 3.08 3.215 19.44  1  0    3    1
#> 15 21.0   6 160.0 110 3.90 2.620 16.46  0  1    4    4
#> 16 18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2
#> 17 18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2
#> 18 15.2   8 304.0 150 3.15 3.435 17.30  0  0    3    2
#> 19 13.3   8 350.0 245 3.73 3.840 15.41  0  0    3    4
#> 20 17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3
#> 21 18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2
#> 22 19.7   6 145.0 175 3.62 2.770 15.50  0  1    5    6
#> 23 21.4   4 121.0 109 4.11 2.780 18.60  1  1    4    2
#> 24 27.3   4  79.0  66 4.08 1.935 18.90  1  1    4    1
#> 25 17.8   6 167.6 123 3.92 3.440 18.90  1  0    4    4
#> 26 19.2   6 167.6 123 3.92 3.440 18.30  1  0    4    4
#> 27 14.7   8 440.0 230 3.23 5.345 17.42  0  0    3    4
#> 28 16.4   8 275.8 180 3.07 4.070 17.40  0  0    3    3
#> 29 21.0   6 160.0 110 3.90 2.620 16.46  0  1    4    4
#> 30 30.4   4  75.7  52 4.93 1.615 18.52  1  1    4    2
#> 31 15.2   8 275.8 180 3.07 3.780 18.00  0  0    3    3
#> 32 33.9   4  71.1  65 4.22 1.835 19.90  1  1    4    1

# Changes:
# * use prop = 1 to randomly sample all rows

slice_sample(mtcars, prop = 1)

#>     mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#> 1  10.4   8 472.0 205 2.93 5.250 17.98  0  0    3    4
#> 2  21.0   6 160.0 110 3.90 2.875 17.02  0  1    4    4
#> 3  27.3   4  79.0  66 4.08 1.935 18.90  1  1    4    1
#> 4  15.0   8 301.0 335 3.54 3.570 14.60  0  1    5    8
#> 5  18.1   6 225.0 105 2.76 3.460 20.22  1  0    3    1
#> 6  21.4   6 258.0 110 3.08 3.215 19.44  1  0    3    1
#> 7  19.2   8 400.0 175 3.08 3.845 17.05  0  0    3    2
#> 8  16.4   8 275.8 180 3.07 4.070 17.40  0  0    3    3
#> 9  15.2   8 304.0 150 3.15 3.435 17.30  0  0    3    2
#> 10 33.9   4  71.1  65 4.22 1.835 19.90  1  1    4    1
#> 11 32.4   4  78.7  66 4.08 2.200 19.47  1  1    4    1
#> 12 21.0   6 160.0 110 3.90 2.620 16.46  0  1    4    4
#> 13 13.3   8 350.0 245 3.73 3.840 15.41  0  0    3    4
#> 14 15.2   8 275.8 180 3.07 3.780 18.00  0  0    3    3
#> 15 30.4   4  75.7  52 4.93 1.615 18.52  1  1    4    2
#> 16 22.8   4 108.0  93 3.85 2.320 18.61  1  1    4    1
#> 17 19.2   6 167.6 123 3.92 3.440 18.30  1  0    4    4
#> 18 14.7   8 440.0 230 3.23 5.345 17.42  0  0    3    4
#> 19 15.5   8 318.0 150 2.76 3.520 16.87  0  0    3    2
#> 20 21.4   4 121.0 109 4.11 2.780 18.60  1  1    4    2
#> 21 17.8   6 167.6 123 3.92 3.440 18.90  1  0    4    4
#> 22 14.3   8 360.0 245 3.21 3.570 15.84  0  0    3    4
#> 23 17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3
#> 24 24.4   4 146.7  62 3.69 3.190 20.00  1  0    4    2
#> 25 26.0   4 120.3  91 4.43 2.140 16.70  0  1    5    2
#> 26 10.4   8 460.0 215 3.00 5.424 17.82  0  0    3    4
#> 27 30.4   4  95.1 113 3.77 1.513 16.90  1  1    5    2
#> 28 21.5   4 120.1  97 3.70 2.465 20.01  1  0    3    1
#> 29 15.8   8 351.0 264 4.22 3.170 14.50  0  1    5    4
#> 30 19.7   6 145.0 175 3.62 2.770 15.50  0  1    5    6
#> 31 22.8   4 140.8  95 3.92 3.150 22.90  1  0    4    2
#> 32 18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2

slice_sample(mtcars, prop = 1, replace = TRUE)

#>     mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#> 1  21.4   4 121.0 109 4.11 2.780 18.60  1  1    4    2
#> 2  33.9   4  71.1  65 4.22 1.835 19.90  1  1    4    1
#> 3  15.0   8 301.0 335 3.54 3.570 14.60  0  1    5    8
#> 4  10.4   8 472.0 205 2.93 5.250 17.98  0  0    3    4
#> 5  15.5   8 318.0 150 2.76 3.520 16.87  0  0    3    2
#> 6  21.0   6 160.0 110 3.90 2.875 17.02  0  1    4    4
#> 7  15.0   8 301.0 335 3.54 3.570 14.60  0  1    5    8
#> 8  10.4   8 460.0 215 3.00 5.424 17.82  0  0    3    4
#> 9  18.1   6 225.0 105 2.76 3.460 20.22  1  0    3    1
#> 10 16.4   8 275.8 180 3.07 4.070 17.40  0  0    3    3
#> 11 21.4   4 121.0 109 4.11 2.780 18.60  1  1    4    2
#> 12 18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2
#> 13 17.8   6 167.6 123 3.92 3.440 18.90  1  0    4    4
#> 14 15.0   8 301.0 335 3.54 3.570 14.60  0  1    5    8
#> 15 15.8   8 351.0 264 4.22 3.170 14.50  0  1    5    4
#> 16 15.0   8 301.0 335 3.54 3.570 14.60  0  1    5    8
#> 17 27.3   4  79.0  66 4.08 1.935 18.90  1  1    4    1
#> 18 18.1   6 225.0 105 2.76 3.460 20.22  1  0    3    1
#> 19 22.8   4 108.0  93 3.85 2.320 18.61  1  1    4    1
#> 20 21.0   6 160.0 110 3.90 2.875 17.02  0  1    4    4
#> 21 19.7   6 145.0 175 3.62 2.770 15.50  0  1    5    6
#> 22 30.4   4  95.1 113 3.77 1.513 16.90  1  1    5    2
#> 23 16.4   8 275.8 180 3.07 4.070 17.40  0  0    3    3
#> 24 19.7   6 145.0 175 3.62 2.770 15.50  0  1    5    6
#> 25 22.8   4 140.8  95 3.92 3.150 22.90  1  0    4    2
#> 26 26.0   4 120.3  91 4.43 2.140 16.70  0  1    5    2
#> 27 30.4   4  75.7  52 4.93 1.615 18.52  1  1    4    2
#> 28 13.3   8 350.0 245 3.73 3.840 15.41  0  0    3    4
#> 29 19.2   8 400.0 175 3.08 3.845 17.05  0  0    3    2
#> 30 18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2
#> 31 18.1   6 225.0 105 2.76 3.460 20.22  1  0    3    1
#> 32 15.5   8 318.0 150 2.76 3.520 16.87  0  0    3    2