Sample n rows from a table

sample_n() and sample_frac() have been superseded in favour of slice_sample(). While they will not be deprecated in the near future, retirement means that we will only perform critical bug fixes, so we recommend moving to the newer alternative.

These functions were superseded because we realised it was more convenient to have two mutually exclusive arguments to one function, rather than two separate functions. This also made it to clean up a few other smaller design issues with sample_n()/sample_frac:

The connection to slice() was not obvious.

The name of the first argument, tbl, is inconsistent with other single table verbs which use .data.

The size argument uses tidy evaluation, which is surprising and undocumented.

It was easier to remove the deprecated .env argument.

... was in a suboptimal position.

sample_n(tbl, size, replace = FALSE, weight = NULL, .env = NULL, ...)

sample_frac(tbl, size = 1, replace = FALSE, weight = NULL, .env = NULL, ...)

Arguments

tbl
A data.frame.

size
<tidy-select> For sample_n(), the number of rows to select. For sample_frac(), the fraction of rows to select. If tbl is grouped, size applies to each group.

replace
Sample with or without replacement?

weight
<tidy-select> Sampling weights. This must evaluate to a vector of non-negative numbers the same length as the input. Weights are automatically standardised to sum to 1.

.env
DEPRECATED.

...
ignored

Examples

by_cyl <- mtcars %>% group_by(cyl)

# sample_n() -> slice_sample() ----------------------------------------------
sample_n(mtcars, 10)

#>                     mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#> AMC Javelin        15.2   8 304.0 150 3.15 3.435 17.30  0  0    3    2
#> Camaro Z28         13.3   8 350.0 245 3.73 3.840 15.41  0  0    3    4
#> Ferrari Dino       19.7   6 145.0 175 3.62 2.770 15.50  0  1    5    6
#> Lotus Europa       30.4   4  95.1 113 3.77 1.513 16.90  1  1    5    2
#> Honda Civic        30.4   4  75.7  52 4.93 1.615 18.52  1  1    4    2
#> Dodge Challenger   15.5   8 318.0 150 2.76 3.520 16.87  0  0    3    2
#> Porsche 914-2      26.0   4 120.3  91 4.43 2.140 16.70  0  1    5    2
#> Hornet 4 Drive     21.4   6 258.0 110 3.08 3.215 19.44  1  0    3    1
#> Merc 450SL         17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3
#> Cadillac Fleetwood 10.4   8 472.0 205 2.93 5.250 17.98  0  0    3    4

sample_n(mtcars, 50, replace = TRUE)

#>                        mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#> Merc 450SL...1        17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3
#> Fiat 128...2          32.4   4  78.7  66 4.08 2.200 19.47  1  1    4    1
#> Fiat 128...3          32.4   4  78.7  66 4.08 2.200 19.47  1  1    4    1
#> Porsche 914-2...4     26.0   4 120.3  91 4.43 2.140 16.70  0  1    5    2
#> Honda Civic...5       30.4   4  75.7  52 4.93 1.615 18.52  1  1    4    2
#> Merc 450SL...6        17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3
#> Merc 230...7          22.8   4 140.8  95 3.92 3.150 22.90  1  0    4    2
#> Merc 230...8          22.8   4 140.8  95 3.92 3.150 22.90  1  0    4    2
#> Mazda RX4...9         21.0   6 160.0 110 3.90 2.620 16.46  0  1    4    4
#> Volvo 142E...10       21.4   4 121.0 109 4.11 2.780 18.60  1  1    4    2
#> Lincoln Continental   10.4   8 460.0 215 3.00 5.424 17.82  0  0    3    4
#> AMC Javelin...12      15.2   8 304.0 150 3.15 3.435 17.30  0  0    3    2
#> Fiat X1-9...13        27.3   4  79.0  66 4.08 1.935 18.90  1  1    4    1
#> Ford Pantera L...14   15.8   8 351.0 264 4.22 3.170 14.50  0  1    5    4
#> Volvo 142E...15       21.4   4 121.0 109 4.11 2.780 18.60  1  1    4    2
#> Ford Pantera L...16   15.8   8 351.0 264 4.22 3.170 14.50  0  1    5    4
#> Merc 240D             24.4   4 146.7  62 3.69 3.190 20.00  1  0    4    2
#> AMC Javelin...18      15.2   8 304.0 150 3.15 3.435 17.30  0  0    3    2
#> Honda Civic...19      30.4   4  75.7  52 4.93 1.615 18.52  1  1    4    2
#> Fiat 128...20         32.4   4  78.7  66 4.08 2.200 19.47  1  1    4    1
#> Ford Pantera L...21   15.8   8 351.0 264 4.22 3.170 14.50  0  1    5    4
#> Valiant...22          18.1   6 225.0 105 2.76 3.460 20.22  1  0    3    1
#> Porsche 914-2...23    26.0   4 120.3  91 4.43 2.140 16.70  0  1    5    2
#> Merc 280              19.2   6 167.6 123 3.92 3.440 18.30  1  0    4    4
#> Dodge Challenger...25 15.5   8 318.0 150 2.76 3.520 16.87  0  0    3    2
#> Valiant...26          18.1   6 225.0 105 2.76 3.460 20.22  1  0    3    1
#> Dodge Challenger...27 15.5   8 318.0 150 2.76 3.520 16.87  0  0    3    2
#> Merc 450SLC           15.2   8 275.8 180 3.07 3.780 18.00  0  0    3    3
#> Hornet 4 Drive...29   21.4   6 258.0 110 3.08 3.215 19.44  1  0    3    1
#> Maserati Bora...30    15.0   8 301.0 335 3.54 3.570 14.60  0  1    5    8
#> Merc 450SL...31       17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3
#> Mazda RX4...32        21.0   6 160.0 110 3.90 2.620 16.46  0  1    4    4
#> AMC Javelin...33      15.2   8 304.0 150 3.15 3.435 17.30  0  0    3    2
#> Volvo 142E...34       21.4   4 121.0 109 4.11 2.780 18.60  1  1    4    2
#> Hornet Sportabout     18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2
#> Merc 450SE...36       16.4   8 275.8 180 3.07 4.070 17.40  0  0    3    3
#> Merc 450SL...37       17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3
#> Volvo 142E...38       21.4   4 121.0 109 4.11 2.780 18.60  1  1    4    2
#> Cadillac Fleetwood    10.4   8 472.0 205 2.93 5.250 17.98  0  0    3    4
#> Fiat X1-9...40        27.3   4  79.0  66 4.08 1.935 18.90  1  1    4    1
#> AMC Javelin...41      15.2   8 304.0 150 3.15 3.435 17.30  0  0    3    2
#> Merc 450SL...42       17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3
#> Hornet 4 Drive...43   21.4   6 258.0 110 3.08 3.215 19.44  1  0    3    1
#> Maserati Bora...44    15.0   8 301.0 335 3.54 3.570 14.60  0  1    5    8
#> Chrysler Imperial     14.7   8 440.0 230 3.23 5.345 17.42  0  0    3    4
#> Merc 450SE...46       16.4   8 275.8 180 3.07 4.070 17.40  0  0    3    3
#> Volvo 142E...47       21.4   4 121.0 109 4.11 2.780 18.60  1  1    4    2
#> Toyota Corolla        33.9   4  71.1  65 4.22 1.835 19.90  1  1    4    1
#> Valiant...49          18.1   6 225.0 105 2.76 3.460 20.22  1  0    3    1
#> Camaro Z28            13.3   8 350.0 245 3.73 3.840 15.41  0  0    3    4

sample_n(mtcars, 10, weight = mpg)

#>                   mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#> Merc 450SL       17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3
#> Merc 230         22.8   4 140.8  95 3.92 3.150 22.90  1  0    4    2
#> Lotus Europa     30.4   4  95.1 113 3.77 1.513 16.90  1  1    5    2
#> Datsun 710       22.8   4 108.0  93 3.85 2.320 18.61  1  1    4    1
#> Pontiac Firebird 19.2   8 400.0 175 3.08 3.845 17.05  0  0    3    2
#> Mazda RX4 Wag    21.0   6 160.0 110 3.90 2.875 17.02  0  1    4    4
#> Toyota Corona    21.5   4 120.1  97 3.70 2.465 20.01  1  0    3    1
#> Merc 450SE       16.4   8 275.8 180 3.07 4.070 17.40  0  0    3    3
#> Mazda RX4        21.0   6 160.0 110 3.90 2.620 16.46  0  1    4    4
#> Valiant          18.1   6 225.0 105 2.76 3.460 20.22  1  0    3    1

# Changes:
# * explicitly name the `n` argument,
# * the `weight` argument is now `weight_by`.

slice_sample(mtcars, n = 10)

#>                     mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#> Camaro Z28         13.3   8 350.0 245 3.73 3.840 15.41  0  0    3    4
#> Maserati Bora      15.0   8 301.0 335 3.54 3.570 14.60  0  1    5    8
#> Chrysler Imperial  14.7   8 440.0 230 3.23 5.345 17.42  0  0    3    4
#> Mazda RX4 Wag      21.0   6 160.0 110 3.90 2.875 17.02  0  1    4    4
#> Hornet Sportabout  18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2
#> Pontiac Firebird   19.2   8 400.0 175 3.08 3.845 17.05  0  0    3    2
#> Duster 360         14.3   8 360.0 245 3.21 3.570 15.84  0  0    3    4
#> Merc 280C          17.8   6 167.6 123 3.92 3.440 18.90  1  0    4    4
#> Merc 450SE         16.4   8 275.8 180 3.07 4.070 17.40  0  0    3    3
#> Cadillac Fleetwood 10.4   8 472.0 205 2.93 5.250 17.98  0  0    3    4

slice_sample(mtcars, n = 50, replace = TRUE)

#>                         mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#> Honda Civic...1        30.4   4  75.7  52 4.93 1.615 18.52  1  1    4    2
#> Hornet Sportabout...2  18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2
#> Merc 280C...3          17.8   6 167.6 123 3.92 3.440 18.90  1  0    4    4
#> AMC Javelin            15.2   8 304.0 150 3.15 3.435 17.30  0  0    3    2
#> Merc 450SLC...5        15.2   8 275.8 180 3.07 3.780 18.00  0  0    3    3
#> Mazda RX4 Wag...6      21.0   6 160.0 110 3.90 2.875 17.02  0  1    4    4
#> Fiat 128...7           32.4   4  78.7  66 4.08 2.200 19.47  1  1    4    1
#> Hornet 4 Drive         21.4   6 258.0 110 3.08 3.215 19.44  1  0    3    1
#> Camaro Z28             13.3   8 350.0 245 3.73 3.840 15.41  0  0    3    4
#> Fiat X1-9...10         27.3   4  79.0  66 4.08 1.935 18.90  1  1    4    1
#> Pontiac Firebird...11  19.2   8 400.0 175 3.08 3.845 17.05  0  0    3    2
#> Mazda RX4 Wag...12     21.0   6 160.0 110 3.90 2.875 17.02  0  1    4    4
#> Merc 240D              24.4   4 146.7  62 3.69 3.190 20.00  1  0    4    2
#> Ford Pantera L...14    15.8   8 351.0 264 4.22 3.170 14.50  0  1    5    4
#> Fiat 128...15          32.4   4  78.7  66 4.08 2.200 19.47  1  1    4    1
#> Datsun 710...16        22.8   4 108.0  93 3.85 2.320 18.61  1  1    4    1
#> Lincoln Continental    10.4   8 460.0 215 3.00 5.424 17.82  0  0    3    4
#> Merc 230...18          22.8   4 140.8  95 3.92 3.150 22.90  1  0    4    2
#> Maserati Bora          15.0   8 301.0 335 3.54 3.570 14.60  0  1    5    8
#> Hornet Sportabout...20 18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2
#> Ford Pantera L...21    15.8   8 351.0 264 4.22 3.170 14.50  0  1    5    4
#> Porsche 914-2...22     26.0   4 120.3  91 4.43 2.140 16.70  0  1    5    2
#> Mazda RX4 Wag...23     21.0   6 160.0 110 3.90 2.875 17.02  0  1    4    4
#> Merc 280C...24         17.8   6 167.6 123 3.92 3.440 18.90  1  0    4    4
#> Duster 360             14.3   8 360.0 245 3.21 3.570 15.84  0  0    3    4
#> Honda Civic...26       30.4   4  75.7  52 4.93 1.615 18.52  1  1    4    2
#> Valiant...27           18.1   6 225.0 105 2.76 3.460 20.22  1  0    3    1
#> Ford Pantera L...28    15.8   8 351.0 264 4.22 3.170 14.50  0  1    5    4
#> Datsun 710...29        22.8   4 108.0  93 3.85 2.320 18.61  1  1    4    1
#> Merc 450SE...30        16.4   8 275.8 180 3.07 4.070 17.40  0  0    3    3
#> Valiant...31           18.1   6 225.0 105 2.76 3.460 20.22  1  0    3    1
#> Porsche 914-2...32     26.0   4 120.3  91 4.43 2.140 16.70  0  1    5    2
#> Porsche 914-2...33     26.0   4 120.3  91 4.43 2.140 16.70  0  1    5    2
#> Porsche 914-2...34     26.0   4 120.3  91 4.43 2.140 16.70  0  1    5    2
#> Valiant...35           18.1   6 225.0 105 2.76 3.460 20.22  1  0    3    1
#> Merc 450SE...36        16.4   8 275.8 180 3.07 4.070 17.40  0  0    3    3
#> Cadillac Fleetwood     10.4   8 472.0 205 2.93 5.250 17.98  0  0    3    4
#> Honda Civic...38       30.4   4  75.7  52 4.93 1.615 18.52  1  1    4    2
#> Chrysler Imperial      14.7   8 440.0 230 3.23 5.345 17.42  0  0    3    4
#> Mazda RX4...40         21.0   6 160.0 110 3.90 2.620 16.46  0  1    4    4
#> Fiat X1-9...41         27.3   4  79.0  66 4.08 1.935 18.90  1  1    4    1
#> Mazda RX4...42         21.0   6 160.0 110 3.90 2.620 16.46  0  1    4    4
#> Porsche 914-2...43     26.0   4 120.3  91 4.43 2.140 16.70  0  1    5    2
#> Pontiac Firebird...44  19.2   8 400.0 175 3.08 3.845 17.05  0  0    3    2
#> Datsun 710...45        22.8   4 108.0  93 3.85 2.320 18.61  1  1    4    1
#> Merc 230...46          22.8   4 140.8  95 3.92 3.150 22.90  1  0    4    2
#> Merc 450SLC...47       15.2   8 275.8 180 3.07 3.780 18.00  0  0    3    3
#> Toyota Corolla         33.9   4  71.1  65 4.22 1.835 19.90  1  1    4    1
#> Pontiac Firebird...49  19.2   8 400.0 175 3.08 3.845 17.05  0  0    3    2
#> Ford Pantera L...50    15.8   8 351.0 264 4.22 3.170 14.50  0  1    5    4

slice_sample(mtcars, n = 10, weight_by = mpg)

#>                 mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#> Merc 280       19.2   6 167.6 123 3.92 3.440 18.30  1  0    4    4
#> Mazda RX4      21.0   6 160.0 110 3.90 2.620 16.46  0  1    4    4
#> Honda Civic    30.4   4  75.7  52 4.93 1.615 18.52  1  1    4    2
#> Lotus Europa   30.4   4  95.1 113 3.77 1.513 16.90  1  1    5    2
#> Ford Pantera L 15.8   8 351.0 264 4.22 3.170 14.50  0  1    5    4
#> Porsche 914-2  26.0   4 120.3  91 4.43 2.140 16.70  0  1    5    2
#> Mazda RX4 Wag  21.0   6 160.0 110 3.90 2.875 17.02  0  1    4    4
#> Fiat 128       32.4   4  78.7  66 4.08 2.200 19.47  1  1    4    1
#> Merc 230       22.8   4 140.8  95 3.92 3.150 22.90  1  0    4    2
#> Valiant        18.1   6 225.0 105 2.76 3.460 20.22  1  0    3    1

# Note that sample_n() would error if n was bigger than the group size
# slice_sample() will just use the available rows for consistency with
# the other slice helpers like slice_head()

# sample_frac() -> slice_sample() -------------------------------------------
sample_frac(mtcars)

#>                      mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#> Camaro Z28          13.3   8 350.0 245 3.73 3.840 15.41  0  0    3    4
#> Valiant             18.1   6 225.0 105 2.76 3.460 20.22  1  0    3    1
#> Honda Civic         30.4   4  75.7  52 4.93 1.615 18.52  1  1    4    2
#> Dodge Challenger    15.5   8 318.0 150 2.76 3.520 16.87  0  0    3    2
#> Mazda RX4 Wag       21.0   6 160.0 110 3.90 2.875 17.02  0  1    4    4
#> Ferrari Dino        19.7   6 145.0 175 3.62 2.770 15.50  0  1    5    6
#> Fiat 128            32.4   4  78.7  66 4.08 2.200 19.47  1  1    4    1
#> Hornet 4 Drive      21.4   6 258.0 110 3.08 3.215 19.44  1  0    3    1
#> Porsche 914-2       26.0   4 120.3  91 4.43 2.140 16.70  0  1    5    2
#> Merc 280C           17.8   6 167.6 123 3.92 3.440 18.90  1  0    4    4
#> Duster 360          14.3   8 360.0 245 3.21 3.570 15.84  0  0    3    4
#> Cadillac Fleetwood  10.4   8 472.0 205 2.93 5.250 17.98  0  0    3    4
#> Pontiac Firebird    19.2   8 400.0 175 3.08 3.845 17.05  0  0    3    2
#> Mazda RX4           21.0   6 160.0 110 3.90 2.620 16.46  0  1    4    4
#> Hornet Sportabout   18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2
#> Fiat X1-9           27.3   4  79.0  66 4.08 1.935 18.90  1  1    4    1
#> Ford Pantera L      15.8   8 351.0 264 4.22 3.170 14.50  0  1    5    4
#> Merc 240D           24.4   4 146.7  62 3.69 3.190 20.00  1  0    4    2
#> Merc 450SL          17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3
#> Chrysler Imperial   14.7   8 440.0 230 3.23 5.345 17.42  0  0    3    4
#> Merc 280            19.2   6 167.6 123 3.92 3.440 18.30  1  0    4    4
#> AMC Javelin         15.2   8 304.0 150 3.15 3.435 17.30  0  0    3    2
#> Merc 450SE          16.4   8 275.8 180 3.07 4.070 17.40  0  0    3    3
#> Volvo 142E          21.4   4 121.0 109 4.11 2.780 18.60  1  1    4    2
#> Toyota Corolla      33.9   4  71.1  65 4.22 1.835 19.90  1  1    4    1
#> Merc 230            22.8   4 140.8  95 3.92 3.150 22.90  1  0    4    2
#> Datsun 710          22.8   4 108.0  93 3.85 2.320 18.61  1  1    4    1
#> Toyota Corona       21.5   4 120.1  97 3.70 2.465 20.01  1  0    3    1
#> Maserati Bora       15.0   8 301.0 335 3.54 3.570 14.60  0  1    5    8
#> Lotus Europa        30.4   4  95.1 113 3.77 1.513 16.90  1  1    5    2
#> Merc 450SLC         15.2   8 275.8 180 3.07 3.780 18.00  0  0    3    3
#> Lincoln Continental 10.4   8 460.0 215 3.00 5.424 17.82  0  0    3    4

sample_frac(mtcars, replace = TRUE)

#>                        mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#> Valiant...1           18.1   6 225.0 105 2.76 3.460 20.22  1  0    3    1
#> Hornet 4 Drive...2    21.4   6 258.0 110 3.08 3.215 19.44  1  0    3    1
#> Pontiac Firebird...3  19.2   8 400.0 175 3.08 3.845 17.05  0  0    3    2
#> Merc 450SE            16.4   8 275.8 180 3.07 4.070 17.40  0  0    3    3
#> AMC Javelin           15.2   8 304.0 150 3.15 3.435 17.30  0  0    3    2
#> Toyota Corolla...6    33.9   4  71.1  65 4.22 1.835 19.90  1  1    4    1
#> Fiat 128...7          32.4   4  78.7  66 4.08 2.200 19.47  1  1    4    1
#> Mazda RX4             21.0   6 160.0 110 3.90 2.620 16.46  0  1    4    4
#> Toyota Corolla...9    33.9   4  71.1  65 4.22 1.835 19.90  1  1    4    1
#> Pontiac Firebird...10 19.2   8 400.0 175 3.08 3.845 17.05  0  0    3    2
#> Merc 450SLC...11      15.2   8 275.8 180 3.07 3.780 18.00  0  0    3    3
#> Merc 450SLC...12      15.2   8 275.8 180 3.07 3.780 18.00  0  0    3    3
#> Datsun 710...13       22.8   4 108.0  93 3.85 2.320 18.61  1  1    4    1
#> Merc 280              19.2   6 167.6 123 3.92 3.440 18.30  1  0    4    4
#> Datsun 710...15       22.8   4 108.0  93 3.85 2.320 18.61  1  1    4    1
#> Ferrari Dino          19.7   6 145.0 175 3.62 2.770 15.50  0  1    5    6
#> Honda Civic...17      30.4   4  75.7  52 4.93 1.615 18.52  1  1    4    2
#> Porsche 914-2         26.0   4 120.3  91 4.43 2.140 16.70  0  1    5    2
#> Duster 360            14.3   8 360.0 245 3.21 3.570 15.84  0  0    3    4
#> Datsun 710...20       22.8   4 108.0  93 3.85 2.320 18.61  1  1    4    1
#> Merc 240D             24.4   4 146.7  62 3.69 3.190 20.00  1  0    4    2
#> Hornet 4 Drive...22   21.4   6 258.0 110 3.08 3.215 19.44  1  0    3    1
#> Volvo 142E...23       21.4   4 121.0 109 4.11 2.780 18.60  1  1    4    2
#> Honda Civic...24      30.4   4  75.7  52 4.93 1.615 18.52  1  1    4    2
#> Dodge Challenger      15.5   8 318.0 150 2.76 3.520 16.87  0  0    3    2
#> Pontiac Firebird...26 19.2   8 400.0 175 3.08 3.845 17.05  0  0    3    2
#> Fiat 128...27         32.4   4  78.7  66 4.08 2.200 19.47  1  1    4    1
#> Cadillac Fleetwood    10.4   8 472.0 205 2.93 5.250 17.98  0  0    3    4
#> Valiant...29          18.1   6 225.0 105 2.76 3.460 20.22  1  0    3    1
#> Camaro Z28            13.3   8 350.0 245 3.73 3.840 15.41  0  0    3    4
#> Volvo 142E...31       21.4   4 121.0 109 4.11 2.780 18.60  1  1    4    2
#> Toyota Corolla...32   33.9   4  71.1  65 4.22 1.835 19.90  1  1    4    1

# Changes:
# * use prop = 1 to randomly sample all rows

slice_sample(mtcars, prop = 1)

#>                      mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#> Maserati Bora       15.0   8 301.0 335 3.54 3.570 14.60  0  1    5    8
#> Cadillac Fleetwood  10.4   8 472.0 205 2.93 5.250 17.98  0  0    3    4
#> Dodge Challenger    15.5   8 318.0 150 2.76 3.520 16.87  0  0    3    2
#> Mazda RX4 Wag       21.0   6 160.0 110 3.90 2.875 17.02  0  1    4    4
#> Lincoln Continental 10.4   8 460.0 215 3.00 5.424 17.82  0  0    3    4
#> Valiant             18.1   6 225.0 105 2.76 3.460 20.22  1  0    3    1
#> Merc 450SE          16.4   8 275.8 180 3.07 4.070 17.40  0  0    3    3
#> Hornet Sportabout   18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2
#> Merc 280C           17.8   6 167.6 123 3.92 3.440 18.90  1  0    4    4
#> Porsche 914-2       26.0   4 120.3  91 4.43 2.140 16.70  0  1    5    2
#> Datsun 710          22.8   4 108.0  93 3.85 2.320 18.61  1  1    4    1
#> Ford Pantera L      15.8   8 351.0 264 4.22 3.170 14.50  0  1    5    4
#> Fiat X1-9           27.3   4  79.0  66 4.08 1.935 18.90  1  1    4    1
#> Merc 230            22.8   4 140.8  95 3.92 3.150 22.90  1  0    4    2
#> Pontiac Firebird    19.2   8 400.0 175 3.08 3.845 17.05  0  0    3    2
#> AMC Javelin         15.2   8 304.0 150 3.15 3.435 17.30  0  0    3    2
#> Chrysler Imperial   14.7   8 440.0 230 3.23 5.345 17.42  0  0    3    4
#> Duster 360          14.3   8 360.0 245 3.21 3.570 15.84  0  0    3    4
#> Merc 450SL          17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3
#> Hornet 4 Drive      21.4   6 258.0 110 3.08 3.215 19.44  1  0    3    1
#> Toyota Corona       21.5   4 120.1  97 3.70 2.465 20.01  1  0    3    1
#> Merc 450SLC         15.2   8 275.8 180 3.07 3.780 18.00  0  0    3    3
#> Ferrari Dino        19.7   6 145.0 175 3.62 2.770 15.50  0  1    5    6
#> Volvo 142E          21.4   4 121.0 109 4.11 2.780 18.60  1  1    4    2
#> Mazda RX4           21.0   6 160.0 110 3.90 2.620 16.46  0  1    4    4
#> Honda Civic         30.4   4  75.7  52 4.93 1.615 18.52  1  1    4    2
#> Merc 280            19.2   6 167.6 123 3.92 3.440 18.30  1  0    4    4
#> Merc 240D           24.4   4 146.7  62 3.69 3.190 20.00  1  0    4    2
#> Camaro Z28          13.3   8 350.0 245 3.73 3.840 15.41  0  0    3    4
#> Toyota Corolla      33.9   4  71.1  65 4.22 1.835 19.90  1  1    4    1
#> Fiat 128            32.4   4  78.7  66 4.08 2.200 19.47  1  1    4    1
#> Lotus Europa        30.4   4  95.1 113 3.77 1.513 16.90  1  1    5    2

slice_sample(mtcars, prop = 1, replace = TRUE)

#>                          mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#> Valiant                 18.1   6 225.0 105 2.76 3.460 20.22  1  0    3    1
#> Merc 230                22.8   4 140.8  95 3.92 3.150 22.90  1  0    4    2
#> Merc 280C               17.8   6 167.6 123 3.92 3.440 18.90  1  0    4    4
#> Fiat 128...4            32.4   4  78.7  66 4.08 2.200 19.47  1  1    4    1
#> Hornet Sportabout...5   18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2
#> Lotus Europa...6        30.4   4  95.1 113 3.77 1.513 16.90  1  1    5    2
#> Cadillac Fleetwood...7  10.4   8 472.0 205 2.93 5.250 17.98  0  0    3    4
#> Duster 360              14.3   8 360.0 245 3.21 3.570 15.84  0  0    3    4
#> Cadillac Fleetwood...9  10.4   8 472.0 205 2.93 5.250 17.98  0  0    3    4
#> Mazda RX4 Wag           21.0   6 160.0 110 3.90 2.875 17.02  0  1    4    4
#> Cadillac Fleetwood...11 10.4   8 472.0 205 2.93 5.250 17.98  0  0    3    4
#> Pontiac Firebird...12   19.2   8 400.0 175 3.08 3.845 17.05  0  0    3    2
#> Ford Pantera L...13     15.8   8 351.0 264 4.22 3.170 14.50  0  1    5    4
#> Ferrari Dino...14       19.7   6 145.0 175 3.62 2.770 15.50  0  1    5    6
#> Pontiac Firebird...15   19.2   8 400.0 175 3.08 3.845 17.05  0  0    3    2
#> Volvo 142E              21.4   4 121.0 109 4.11 2.780 18.60  1  1    4    2
#> Lotus Europa...17       30.4   4  95.1 113 3.77 1.513 16.90  1  1    5    2
#> Ford Pantera L...18     15.8   8 351.0 264 4.22 3.170 14.50  0  1    5    4
#> Merc 280...19           19.2   6 167.6 123 3.92 3.440 18.30  1  0    4    4
#> Ferrari Dino...20       19.7   6 145.0 175 3.62 2.770 15.50  0  1    5    6
#> Datsun 710              22.8   4 108.0  93 3.85 2.320 18.61  1  1    4    1
#> Ford Pantera L...22     15.8   8 351.0 264 4.22 3.170 14.50  0  1    5    4
#> Lincoln Continental     10.4   8 460.0 215 3.00 5.424 17.82  0  0    3    4
#> Toyota Corona           21.5   4 120.1  97 3.70 2.465 20.01  1  0    3    1
#> Camaro Z28              13.3   8 350.0 245 3.73 3.840 15.41  0  0    3    4
#> Fiat X1-9...26          27.3   4  79.0  66 4.08 1.935 18.90  1  1    4    1
#> Hornet Sportabout...27  18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2
#> Mazda RX4               21.0   6 160.0 110 3.90 2.620 16.46  0  1    4    4
#> Merc 280...29           19.2   6 167.6 123 3.92 3.440 18.30  1  0    4    4
#> Honda Civic             30.4   4  75.7  52 4.93 1.615 18.52  1  1    4    2
#> Fiat 128...31           32.4   4  78.7  66 4.08 2.200 19.47  1  1    4    1
#> Fiat X1-9...32          27.3   4  79.0  66 4.08 1.935 18.90  1  1    4    1