Select variables

Select (and optionally rename) variables in a data frame, using a concise mini-language that makes it easy to refer to variables based on their name (e.g. a:f selects all columns between a on the left to f on the right). You can also use predicate functions like is.numeric to select variables based on their properties.

select(.data, ...)

Arguments

.data
A tbl. All main verbs are S3 generics and provide methods for tbl_df(), dtplyr::tbl_dt() and dbplyr::tbl_dbi().

...
<tidy-select> One or more unquoted expressions separated by commas. You can treat variable names like they are positions, so you can use expressions like x:y to select ranges of variables.

Value

An object of the same type as .data.

Rows are not affected.

Output columns are a subset of input columns, potentially with a different order. Columns will be renamed if new_name = old_name form is used.

Data frame attributes are preserved.

Groups are maintained; you can not select off grouping variables.

Useful functions

As well as using existing functions like : and c(), there are a number of special functions that only work inside select():

any_of(), all_of().

starts_with(), ends_with(), contains(), matches().

num_range().

group_cols(), last_col().

everything().

You can also use predicate functions (functions that return a single TRUE or FALSE) like is.numeric, is.character, and is.factor to select variables with specific types.

Selections can be combined using Boolean algebra like:

starts_with("a") & ends_with("x"): start with "a" and end with "x"

starts_with("a") | starts_with("b"): start with "a" or "b"

!starts_with("a"): doesn't start with "a"

To remove variables from a selection, use -:

starts_with("a") - ends_width("x"): start with "a" and doesn't end with "x"

is.numeric - c(a, b, c): numeric variables except for a, b, c.

See select helpers for more details and examples.

Note that except for :, - and c(), all complex expressions are evaluated outside the data frame context. This is to prevent accidental matching of data frame variables when you refer to variables from the calling context.

See also

Examples

iris <- as_tibble(iris) # so it prints a little nicer
select(iris, starts_with("Petal"))

#> # A tibble: 150 x 2
#>    Petal.Length Petal.Width
#>           <dbl>       <dbl>
#>  1          1.4         0.2
#>  2          1.4         0.2
#>  3          1.3         0.2
#>  4          1.5         0.2
#>  5          1.4         0.2
#>  6          1.7         0.4
#>  7          1.4         0.3
#>  8          1.5         0.2
#>  9          1.4         0.2
#> 10          1.5         0.1
#> # … with 140 more rows

select(iris, ends_with("Width"))

#> # A tibble: 150 x 2
#>    Sepal.Width Petal.Width
#>          <dbl>       <dbl>
#>  1         3.5         0.2
#>  2         3           0.2
#>  3         3.2         0.2
#>  4         3.1         0.2
#>  5         3.6         0.2
#>  6         3.9         0.4
#>  7         3.4         0.3
#>  8         3.4         0.2
#>  9         2.9         0.2
#> 10         3.1         0.1
#> # … with 140 more rows

select(iris, !starts_with("Petal"))

#> # A tibble: 150 x 3
#>    Sepal.Length Sepal.Width Species
#>           <dbl>       <dbl> <fct>  
#>  1          5.1         3.5 setosa 
#>  2          4.9         3   setosa 
#>  3          4.7         3.2 setosa 
#>  4          4.6         3.1 setosa 
#>  5          5           3.6 setosa 
#>  6          5.4         3.9 setosa 
#>  7          4.6         3.4 setosa 
#>  8          5           3.4 setosa 
#>  9          4.4         2.9 setosa 
#> 10          4.9         3.1 setosa 
#> # … with 140 more rows

select(iris, starts_with("Petal") & ends_with("Width"))

#> # A tibble: 150 x 1
#>    Petal.Width
#>          <dbl>
#>  1         0.2
#>  2         0.2
#>  3         0.2
#>  4         0.2
#>  5         0.2
#>  6         0.4
#>  7         0.3
#>  8         0.2
#>  9         0.2
#> 10         0.1
#> # … with 140 more rows

select(iris, is.numeric)

#> # A tibble: 150 x 4
#>    Sepal.Length Sepal.Width Petal.Length Petal.Width
#>           <dbl>       <dbl>        <dbl>       <dbl>
#>  1          5.1         3.5          1.4         0.2
#>  2          4.9         3            1.4         0.2
#>  3          4.7         3.2          1.3         0.2
#>  4          4.6         3.1          1.5         0.2
#>  5          5           3.6          1.4         0.2
#>  6          5.4         3.9          1.7         0.4
#>  7          4.6         3.4          1.4         0.3
#>  8          5           3.4          1.5         0.2
#>  9          4.4         2.9          1.4         0.2
#> 10          4.9         3.1          1.5         0.1
#> # … with 140 more rows

# If desired, you can rename individual variables as you select them
select(iris, species = Species, width = Petal.Width, length = Petal.Length)

#> # A tibble: 150 x 3
#>    species width length
#>    <fct>   <dbl>  <dbl>
#>  1 setosa    0.2    1.4
#>  2 setosa    0.2    1.4
#>  3 setosa    0.2    1.3
#>  4 setosa    0.2    1.5
#>  5 setosa    0.2    1.4
#>  6 setosa    0.4    1.7
#>  7 setosa    0.3    1.4
#>  8 setosa    0.2    1.5
#>  9 setosa    0.2    1.4
#> 10 setosa    0.1    1.5
#> # … with 140 more rows

df <- as.data.frame(matrix(runif(100), nrow = 10))
df <- as_tibble(df[c(3, 4, 7, 1, 9, 8, 5, 2, 6, 10)])
select(df, V4:V6)

#> # A tibble: 10 x 8
#>         V4    V7    V1    V9    V8     V5     V2     V6
#>      <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl>  <dbl>  <dbl>
#>  1 0.568   0.753 0.932 0.890 0.623 0.0512 0.577  0.410 
#>  2 0.614   0.222 0.197 0.412 0.214 0.727  0.249  0.493 
#>  3 0.00233 0.262 0.181 0.578 0.175 0.226  0.692  0.251 
#>  4 0.129   0.698 0.564 0.382 0.772 0.427  0.330  0.603 
#>  5 0.261   0.649 0.685 0.841 0.882 0.0613 0.241  0.465 
#>  6 0.971   0.671 0.475 0.774 0.612 0.515  0.602  0.876 
#>  7 0.665   0.997 0.677 0.947 0.441 0.713  0.0220 0.987 
#>  8 0.874   0.973 0.849 0.285 0.759 0.673  0.426  0.540 
#>  9 0.149   0.387 0.432 0.141 0.178 0.152  0.104  0.0661
#> 10 0.536   0.636 0.389 0.665 0.297 0.288  0.378  0.614 

select(df, num_range("V", 4:6))

#> # A tibble: 10 x 3
#>         V4     V5     V6
#>      <dbl>  <dbl>  <dbl>
#>  1 0.568   0.0512 0.410 
#>  2 0.614   0.727  0.493 
#>  3 0.00233 0.226  0.251 
#>  4 0.129   0.427  0.603 
#>  5 0.261   0.0613 0.465 
#>  6 0.971   0.515  0.876 
#>  7 0.665   0.713  0.987 
#>  8 0.874   0.673  0.540 
#>  9 0.149   0.152  0.0661
#> 10 0.536   0.288  0.614 

# Select the grouping variables:
starwars %>% group_by(gender) %>% select(group_cols())

#> Adding missing grouping variables: `gender`

#> # A tibble: 87 x 1
#> # Groups:   gender [3]
#>    gender   
#>    <chr>    
#>  1 masculine
#>  2 masculine
#>  3 masculine
#>  4 masculine
#>  5 feminine 
#>  6 masculine
#>  7 feminine 
#>  8 masculine
#>  9 masculine
#> 10 masculine
#> # … with 77 more rows

# Moving variables around --------------------------
# Move `Species` to the front
select(iris, Species, everything())

#> # A tibble: 150 x 5
#>    Species Sepal.Length Sepal.Width Petal.Length Petal.Width
#>    <fct>          <dbl>       <dbl>        <dbl>       <dbl>
#>  1 setosa           5.1         3.5          1.4         0.2
#>  2 setosa           4.9         3            1.4         0.2
#>  3 setosa           4.7         3.2          1.3         0.2
#>  4 setosa           4.6         3.1          1.5         0.2
#>  5 setosa           5           3.6          1.4         0.2
#>  6 setosa           5.4         3.9          1.7         0.4
#>  7 setosa           4.6         3.4          1.4         0.3
#>  8 setosa           5           3.4          1.5         0.2
#>  9 setosa           4.4         2.9          1.4         0.2
#> 10 setosa           4.9         3.1          1.5         0.1
#> # … with 140 more rows

# Move `Sepal.Length` to the back: first select everything except
# `Sepal.Length`, then select `Sepal.Length`
select(iris, !Sepal.Length, Sepal.Length)

#> # A tibble: 150 x 5
#>    Sepal.Width Petal.Length Petal.Width Species Sepal.Length
#>          <dbl>        <dbl>       <dbl> <fct>          <dbl>
#>  1         3.5          1.4         0.2 setosa           5.1
#>  2         3            1.4         0.2 setosa           4.9
#>  3         3.2          1.3         0.2 setosa           4.7
#>  4         3.1          1.5         0.2 setosa           4.6
#>  5         3.6          1.4         0.2 setosa           5  
#>  6         3.9          1.7         0.4 setosa           5.4
#>  7         3.4          1.4         0.3 setosa           4.6
#>  8         3.4          1.5         0.2 setosa           5  
#>  9         2.9          1.4         0.2 setosa           4.4
#> 10         3.1          1.5         0.1 setosa           4.9
#> # … with 140 more rows