Subset columns using their names and types

Select (and optionally rename) variables in a data frame, using a concise mini-language that makes it easy to refer to variables based on their name (e.g. a:f selects all columns from a on the left to f on the right). You can also use predicate functions like is.numeric to select variables based on their properties.

select(.data, ...)

Arguments

.data
A data frame, data frame extension (e.g. a tibble), or a lazy data frame (e.g. from dbplyr or dtplyr). See Methods, below, for more details.

...
<tidy-select> One or more unquoted expressions separated by commas. Variable names can be used as if they were positions in the data frame, so expressions like x:y can be used to select a range of variables.

Value

An object of the same type as .data. The output has the following properties:

Rows are not affected.

Output columns are a subset of input columns, potentially with a different order. Columns will be renamed if new_name = old_name form is used.

Data frame attributes are preserved.

Groups are maintained; you can't select off grouping variables.

Useful functions

As well as using existing functions like : and c(), there are a number of special functions that only work inside select():

any_of(), all_of().

starts_with(), ends_with(), contains(), matches().

num_range().

group_cols(), last_col().

everything().

You can also use predicate functions (functions that return a single TRUE or FALSE) like is.numeric, is.character, and is.factor to select variables of specific types.

Selections can be combined using Boolean algebra:

starts_with("a") & ends_with("x"): variables with names that start with "a" and end with "x"

starts_with("a") | starts_with("b"): variables with names that start with "a" or "b"

!starts_with("a"): variables with names that do not start with "a"

To remove variables from a selection, use & and !:

starts_with("a") & !ends_width("x"): variables with names that start with "a" and do not end with "x"

is.numeric & !c(a, b, c): numeric variables except, for a, b, c.

See select helpers for more details and examples.

Note that except for :, - and c(), all complex expressions are evaluated outside the data frame context. This is to prevent accidental matching of data frame variables when you refer to variables from the calling environment.

Methods

This function is a generic, which means that packages can provide implementations (methods) for other classes. See the documentation of individual methods for extra arguments and differences in behaviour.

The following methods are currently available in loaded packages: dbplyr (tbl_lazy), dplyr (data.frame, default), utils (list) .

See also

Examples

select(starwars, starts_with("h"))

#> # A tibble: 87 x 3
#>    height hair_color    homeworld
#>     <int> <chr>         <chr>    
#>  1    172 blond         Tatooine 
#>  2    167 NA            Tatooine 
#>  3     96 NA            Naboo    
#>  4    202 none          Tatooine 
#>  5    150 brown         Alderaan 
#>  6    178 brown, grey   Tatooine 
#>  7    165 brown         Tatooine 
#>  8     97 NA            Tatooine 
#>  9    183 black         Tatooine 
#> 10    182 auburn, white Stewjon  
#> # … with 77 more rows

select(starwars, ends_with("color"))

#> # A tibble: 87 x 3
#>    hair_color    skin_color  eye_color
#>    <chr>         <chr>       <chr>    
#>  1 blond         fair        blue     
#>  2 NA            gold        yellow   
#>  3 NA            white, blue red      
#>  4 none          white       yellow   
#>  5 brown         light       brown    
#>  6 brown, grey   light       blue     
#>  7 brown         light       blue     
#>  8 NA            white, red  red      
#>  9 black         light       brown    
#> 10 auburn, white fair        blue-gray
#> # … with 77 more rows

select(starwars, !contains("s"))

#> # A tibble: 87 x 7
#>    name             height hair_color    eye_color birth_year gender   homeworld
#>    <chr>             <int> <chr>         <chr>          <dbl> <chr>    <chr>    
#>  1 Luke Skywalker      172 blond         blue            19   masculi… Tatooine 
#>  2 C-3PO               167 NA            yellow         112   masculi… Tatooine 
#>  3 R2-D2                96 NA            red             33   masculi… Naboo    
#>  4 Darth Vader         202 none          yellow          41.9 masculi… Tatooine 
#>  5 Leia Organa         150 brown         brown           19   feminine Alderaan 
#>  6 Owen Lars           178 brown, grey   blue            52   masculi… Tatooine 
#>  7 Beru Whitesun l…    165 brown         blue            47   feminine Tatooine 
#>  8 R5-D4                97 NA            red             NA   masculi… Tatooine 
#>  9 Biggs Darklight…    183 black         brown           24   masculi… Tatooine 
#> 10 Obi-Wan Kenobi      182 auburn, white blue-gray       57   masculi… Stewjon  
#> # … with 77 more rows

select(starwars, starts_with("h") & ends_with("color"))

#> # A tibble: 87 x 1
#>    hair_color   
#>    <chr>        
#>  1 blond        
#>  2 NA           
#>  3 NA           
#>  4 none         
#>  5 brown        
#>  6 brown, grey  
#>  7 brown        
#>  8 NA           
#>  9 black        
#> 10 auburn, white
#> # … with 77 more rows

select(starwars, is.numeric)

#> # A tibble: 87 x 3
#>    height  mass birth_year
#>     <int> <dbl>      <dbl>
#>  1    172    77       19  
#>  2    167    75      112  
#>  3     96    32       33  
#>  4    202   136       41.9
#>  5    150    49       19  
#>  6    178   120       52  
#>  7    165    75       47  
#>  8     97    32       NA  
#>  9    183    84       24  
#> 10    182    77       57  
#> # … with 77 more rows

# Optionally, rename individual variables as they are selected,
# in the format `new_name = old_name`
select(starwars, character_name = name, character_height = height)

#> # A tibble: 87 x 2
#>    character_name     character_height
#>    <chr>                         <int>
#>  1 Luke Skywalker                  172
#>  2 C-3PO                           167
#>  3 R2-D2                            96
#>  4 Darth Vader                     202
#>  5 Leia Organa                     150
#>  6 Owen Lars                       178
#>  7 Beru Whitesun lars              165
#>  8 R5-D4                            97
#>  9 Biggs Darklighter               183
#> 10 Obi-Wan Kenobi                  182
#> # … with 77 more rows

# Use num_range() to select variables with numeric suffixes
df <- as.data.frame(matrix(runif(100), nrow = 10))
select(df, V4:V6) # Specify variable names explicitly

#>            V4         V5         V6
#> 1  0.09347633 0.58883843 0.61910649
#> 2  0.45980079 0.08148829 0.33753305
#> 3  0.56838655 0.74753070 0.65079497
#> 4  0.35348889 0.32998415 0.30790391
#> 5  0.63402214 0.24618046 0.01695481
#> 6  0.17635642 0.04398338 0.03748926
#> 7  0.92547428 0.21633754 0.37799359
#> 8  0.23290282 0.76802358 0.29402637
#> 9  0.33601364 0.82610630 0.78646655
#> 10 0.14908573 0.97283678 0.93246912

select(df, num_range(prefix = "V", range = 4:6)) # Or, specify the prefix used on a numeric range

#>            V4         V5         V6
#> 1  0.09347633 0.58883843 0.61910649
#> 2  0.45980079 0.08148829 0.33753305
#> 3  0.56838655 0.74753070 0.65079497
#> 4  0.35348889 0.32998415 0.30790391
#> 5  0.63402214 0.24618046 0.01695481
#> 6  0.17635642 0.04398338 0.03748926
#> 7  0.92547428 0.21633754 0.37799359
#> 8  0.23290282 0.76802358 0.29402637
#> 9  0.33601364 0.82610630 0.78646655
#> 10 0.14908573 0.97283678 0.93246912

# Select the existing grouping variables:
starwars %>% group_by(gender, eye_color) %>% select(group_cols())

#> # A tibble: 87 x 2
#> # Groups:   gender, eye_color [23]
#>    gender    eye_color
#>    <chr>     <chr>    
#>  1 masculine blue     
#>  2 masculine yellow   
#>  3 masculine red      
#>  4 masculine yellow   
#>  5 feminine  brown    
#>  6 masculine blue     
#>  7 feminine  blue     
#>  8 masculine red      
#>  9 masculine brown    
#> 10 masculine blue-gray
#> # … with 77 more rows

# Using select() semantics in across()
starwars %>% summarise(across(.cols = height:mass, .fns = ~mean(.x, na.rm = TRUE)))

#> # A tibble: 1 x 2
#>   height  mass
#>    <dbl> <dbl>
#> 1   174.  97.3

# Use `{{ }}` inside functions to tunnel data-variables through
# function arguments. See ?dplyr_tidy_eval for more information.
averages <- function(data, vars) {
  data %>%
    select({{ vars }}) %>%
    lapply(mean, na.rm = TRUE)
}
starwars %>% averages(height)

#> $height
#> [1] 174.358
#> 

starwars %>% averages(c(height, mass))

#> $height
#> [1] 174.358
#> 
#> $mass
#> [1] 97.31186
#> 

# Modifying the order of variables --------------------------
# As of dplyr 1.0.0, use relocate(), not select():
starwars %>% select(name:birth_year) %>% relocate(birth_year, .before = 1)

#> # A tibble: 87 x 7
#>    birth_year name               height  mass hair_color    skin_color eye_color
#>         <dbl> <chr>               <int> <dbl> <chr>         <chr>      <chr>    
#>  1       19   Luke Skywalker        172    77 blond         fair       blue     
#>  2      112   C-3PO                 167    75 NA            gold       yellow   
#>  3       33   R2-D2                  96    32 NA            white, bl… red      
#>  4       41.9 Darth Vader           202   136 none          white      yellow   
#>  5       19   Leia Organa           150    49 brown         light      brown    
#>  6       52   Owen Lars             178   120 brown, grey   light      blue     
#>  7       47   Beru Whitesun lars    165    75 brown         light      blue     
#>  8       NA   R5-D4                  97    32 NA            white, red red      
#>  9       24   Biggs Darklighter     183    84 black         light      brown    
#> 10       57   Obi-Wan Kenobi        182    77 auburn, white fair       blue-gray
#> # … with 77 more rows

starwars %>% select(name:birth_year) %>% relocate(name, .after = last_col())

#> # A tibble: 87 x 7
#>    height  mass hair_color    skin_color  eye_color birth_year name             
#>     <int> <dbl> <chr>         <chr>       <chr>          <dbl> <chr>            
#>  1    172    77 blond         fair        blue            19   Luke Skywalker   
#>  2    167    75 NA            gold        yellow         112   C-3PO            
#>  3     96    32 NA            white, blue red             33   R2-D2            
#>  4    202   136 none          white       yellow          41.9 Darth Vader      
#>  5    150    49 brown         light       brown           19   Leia Organa      
#>  6    178   120 brown, grey   light       blue            52   Owen Lars        
#>  7    165    75 brown         light       blue            47   Beru Whitesun la…
#>  8     97    32 NA            white, red  red             NA   R5-D4            
#>  9    183    84 black         light       brown           24   Biggs Darklighter
#> 10    182    77 auburn, white fair        blue-gray       57   Obi-Wan Kenobi   
#> # … with 77 more rows