Select variables

Retired: These functions now live in the tidyselect package as tidyselect::vars_select(), tidyselect::vars_rename() and tidyselect::vars_pull(). These dplyr aliases are soft-deprecated and will be deprecated sometimes in the future.

select_vars(vars = chr(), ..., include = chr(), exclude = chr())

rename_vars(vars = chr(), ..., strict = TRUE)

select_var(vars, var = -1)

current_vars(...)

Arguments

vars
A character vector of existing column names.

...
Expressions to compute.

include, exclude
Character vector of column names to always include/exclude.

strict
If TRUE, will throw an error if you attempt to rename a variable that doesn't exist.

var
A variable specified as in the same argument of tidyselect::vars_pull().

Details