src_local {dplyr}	R Documentation

A local source

Description

This function was deprecated since it existed to support a style of testing dplyr backends that turned out not to be useful.

Usage

src_local(tbl, pkg = NULL, env = NULL)

src_df(pkg = NULL, env = NULL)

Arguments

tbl
name of the function used to generate tbl objects

pkg, env
Either the name of a package or an environment object in which to look for objects.

[Package dplyr version 1.0.7 Index]