A local source.

This is mainly useful for testing, since makes it possible to refer to local and remote tables using exactly the same syntax.

src_local(tbl, pkg = NULL, env = NULL)

src_df(pkg = NULL, env = NULL)

Arguments

tbl
name of the function used to generate tbl objects

pkg, env
Either the name of a package or an environment object in which to look for objects.

Details

Generally, src_local() should not be called directly, but instead one of the constructors should be used.

Examples

if (require("Lahman")) {
batting_df <- tbl(src_df("Lahman"), "Batting")
}