tidyeval-compat {dplyr}	R Documentation

Other tidy eval tools

Description

These tidy eval functions are no longer for normal usage, but are still exported from dplyr for backward compatibility. See ?rlang::args_data_masking and vignette("programming") for the latest recommendations.

expr()

enquo()

enquos()

sym()

syms()

as_label()

quo()

quos()

quo_name()

ensym()

ensyms()

enexpr()

enexprs()

[Package dplyr version 1.1.3 Index]