Select top (or bottom) n rows (by value)

This is a convenient wrapper that uses filter() and min_rank() to select the top or bottom entries in each group, ordered by wt.

top_n(x, n, wt)

Arguments

x
a tbl() to filter

n
number of rows to return. If x is grouped, this is the number of rows per group. Will include more than n rows if there are ties.

If n is positive, selects the top n rows. If negative, selects the bottom n rows.

wt
(Optional). The variable to use for ordering. If not specified, defaults to the last variable in the tbl.

This argument is automatically quoted and later evaluated in the context of the data frame. It supports unquoting. See vignette("programming") for an introduction to these concepts.

Examples

df <- data.frame(x = c(10, 4, 1, 6, 3, 1, 1))
df %>% top_n(2)

#> Selecting by x

#>    x
#> 1 10
#> 2  6

# Negative values select bottom from group. Note that we get more
# than 2 values here because there's a tie: top_n() either takes
# all rows with a value, or none.
df %>% top_n(-2)

#> Selecting by x

#>   x
#> 1 1
#> 2 1
#> 3 1

if (require("Lahman")) {
# Find 10 players with most games
# A little nicer with %>%
tbl_df(Batting) %>%
  group_by(playerID) %>%
  tally(G) %>%
  top_n(10)

# Find year with most games for each player
tbl_df(Batting) %>% group_by(playerID) %>% top_n(1, G)
}

#> Selecting by n

#> # A tibble: 19,791 x 22
#> # Groups:   playerID [18,915]
#>    playerID yearID stint teamID lgID      G    AB     R     H   X2B   X3B    HR
#>    <chr>     <int> <int> <fct>  <fct> <int> <int> <int> <int> <int> <int> <int>
#>  1 aardsda…   2009     1 SEA    AL       73     0     0     0     0     0     0
#>  2 aaronha…   1963     1 ML1    NL      161   631   121   201    29     4    44
#>  3 aaronto…   1962     1 ML1    NL      141   334    54    77    20     2     8
#>  4 aasedo01   1986     1 BAL    AL       66     0     0     0     0     0     0
#>  5 abadan01   2003     1 BOS    AL        9    17     1     2     0     0     0
#>  6 abadfe01   2014     1 OAK    AL       69     0     0     0     0     0     0
#>  7 abadijo…   1875     1 PH3    NA       11    45     3    10     0     0     0
#>  8 abbated…   1904     1 BSN    NL      154   579    76   148    18    10     3
#>  9 abbeybe…   1892     1 WAS    NL       27    75     5     9     1     0     0
#> 10 abbeych…   1895     1 WAS    NL      132   511   102   141    14    10     8
#> # ... with 19,781 more rows, and 10 more variables: RBI <int>, SB <int>,
#> #   CS <int>, BB <int>, SO <int>, IBB <int>, HBP <int>, SH <int>, SF <int>,
#> #   GIDP <int>