vars {dplyr}	R Documentation

Select variables

Description

vars() was only needed for the scoped verbs, which have been superseded by the use of across() in an existing verb. See vignette("colwise") for details.

This helper is intended to provide equivalent semantics to select(). It is used for instance in scoped summarising and mutating verbs (mutate_at() and summarise_at()).

Note that verbs accepting a vars() specification also accept a numeric vector of positions or a character vector of column names.

Usage

vars(...)

Arguments

...
<tidy-select> Variables to include/exclude in mutate/summarise. You can use same specifications as in select(). If missing, defaults to all non-grouping variables.

See Also

all_vars() and any_vars() for other quoting functions that you can use with scoped verbs.

[Package dplyr version 1.0.7 Index]