Select variables.

This helper is intended to provide equivalent semantics to select(). It is used for instance in scoped summarising and mutating verbs (mutate_at() and summarise_at()).

vars(...)

Arguments

...
Variables to include/exclude in mutate/summarise. You can use same specifications as in select(). If missing, defaults to all non-grouping variables.

These arguments are automatically quoted and later evaluated in the context of the data frame. They support unquoting. See vignette("programming") for an introduction to these concepts.

Details

Note that verbs accepting a vars() specification also accept an integerish vector of positions or a character vector of column names.

See also

funs(), all_vars() and any_vars() for other quoting functions that you can use with scoped verbs.