vars {dplyr}	R Documentation

Select variables

Description

vars() is superseded because it is only needed for the scoped verbs (i.e. mutate_at(), summarise_at(), and friends), which have been been superseded in favour of across(). See vignette("colwise") for details.

This helper is intended to provide tidy-select semantics for scoped verbs like mutate_at() and summarise_at(). Note that anywhere you can supply vars() specification, you can also supply a numeric vector of column positions or a character vector of column names.

Usage

vars(...)

Arguments

...
<tidy-select> Variables to operate on.

See Also

all_vars() and any_vars() for other quoting functions that you can use with scoped verbs.

[Package dplyr version 1.1.1 Index]